### 2024-07-30 #7054 @editorialbot
[PRE REVIEW]: EMGFlow: A Python package for pre-processing and feature extraction of electromyographic signals
**Submitting author:** @WiIIson (<a href="http://orcid.org/0009-0001-7454-1286">D. William Lawrence Conley</a>)
**Repository:** https://github.com/WiIIson/EMGFlow-Python-Package
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.16
