### 2024-07-30 #7054 @faroit
@Kevin-Mattheus-Moerman This is not in my domain of applications but the signal processing techniques are similar so I can edit this.

I'm on vacation this week so I will start next week
