### 2024-08-08 #7054 @WiIIson
Hi,

Thanks for considering our submission for publication. Here is a list of editors we believe would be relevant to the work:

- [Dr Lara McManus](https://www.laramcmanus.com/LaraCV.html), Trinity College Dublin, [lara.mc-manus@ucdconnect.ie](mailto:lara.mc-manus@ucdconnect.ie)
- [Dr. Angkoon Phinyomark](https://www.angkoonphinyomark.com/about_me), University of New Brunswick, [aphinyom@unb.ca](mailto:aphinyom@unb.ca)
- [Dr Sato Wataru](https://watarusato.shin-gen.jp/ProfileE.html), RIKEN research institute, [wataru.sato.ya@riken.jp](mailto:wataru.sato.ya@riken.jp)
- [Dr Mohammad Soleymani](https://people.ict.usc.edu/~soleymani/), University of Southern California, [soleymani@ict.usc.edu](mailto:soleymani@ict.usc.edu)
- [Dr Madeleine Lowery](https://people.ucd.ie/madeleine.lowery), University College Dublin, [madeleine.lowery@ucd.ie](mailto:madeleine.lowery@ucd.ie)
- [Dr Dominique Makowski](https://profiles.sussex.ac.uk/p592424-dominique-makowski), University of Sussex, [D.Makowski@sussex.ac.uk](mailto:D.Makowski@sussex.ac.uk)
