### 2024-11-01 #7054 @srlivingstone
Hi @faroit, we've provided a longer list of potential reviewers for you to consider. All of these either published in JOSS, or have previously reviewed for JOSS, and so should be familiar with git.
