### 2024-11-22 #7054 @marcoghislieri
Dear All,

Thank you for considering me to review this submission. Unfortunately, I am currently unavailable to take on this task.

I appreciate your understanding.

Sincerely,

Marco Ghislieri, PhD

Assistant Professor with time contract

PolitoBIOMed Lab | Department of Electronics and Telecommunications

Politecnico di Torino

●  Scopus<https://www.scopus.com/authid/detail.uri?authorId=57202379042> / ResearchGate<https://www.researchgate.net/profile/Marco-Ghislieri> / LinkedIn<https://www.linkedin.com/in/marco-ghislieri/>

●  Website<https://www.polito.it/personale?p=marco.ghislieri>

[cid:923a0138-0613-44c4-98da-50a00395108b]
