### 2024-12-06 #7054 @faroit
@WiIIson thanks a lot for your extended list. This is extremely helpful in this case as your submission is slightly out of my research domain. I guess you checked the list already for potential COI's according to the JOSS policy?
