### 2025-01-07 #7054 @WiIIson
The list of potential COI's are all taken from other related paper assignees, so there shouldn't be any problems.

Here are some new potential COI's:
- Monica Bobra (https://github.com/mbobra)
- Mahfuzur Rahman (https://github.com/mahfuz05062)
- Seyed Amir Alavi (https://github.com/samiralavi)
- Kelly L. Rowland (https://github.com/kellyrowland)
- Luca Ferranti (https://github.com/lucaferranti)
- Sam Zorowitz (https://github.com/szorowi1)
- Sam Forbes (https://github.com/samhforbes)
- Jan Freyberg (https://github.com/janfreyberg)
- Axel Loewe (https://github.com/axel-loewe)
- Ariel Rokem (https://github.com/arokem)
