### 2025-01-10 #7054 @faroit
@mbobra @mahfuz05062 @samiralavi @kellyrowland @lucaferranti @szorowi1 @samhforbes @janfreyberg @axelloewe @arokem would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is https://github.com/WiIIson/EMGFlow-Python-Package
