### 2025-01-12 #7054 @janfreyberg
@faroit likewise happy to review it, looks interesting!
