### 2025-01-30 e0c6ff0
SmoothFilterSignals Bug Fix
Fixed a bug where 'gauss' was spelled 'guass' in the check for which method is being used.
