+++ .github/workflows/draft-pdf.yml
@@ -14,7 +14,7 @@ jobs:
           journal: joss
           paper-path: papers/JOSS/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: papers/JOSS/paper.pdf
