+++ .github/workflows/tests.yml
@@ -65,7 +65,7 @@ jobs:
           python -m pip install tox tox-gh-actions
 
       - name: Download coverage data
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           name: covdata
 
