+++ .github/workflows/tests.yml
@@ -68,6 +68,7 @@ jobs:
         uses: actions/download-artifact@v4
         with:
           name: covdata
+          path: .
 
       - name: Combine
         run: |
