### 2025-01-30 43b0bb8
Github actions fix test 2
Testing actions fix by changing download path to './covdata'.
