+++ tox.ini
@@ -7,7 +7,10 @@ setenv =
 deps =
     -r {toxinidir}/EMGFlow-Package/requirements_dev.txt
 commands =
-    python -m coverage run -p -m unittest discover -s {toxinidir}/EMGFlow-Package/
+    python -m coverage run -parallel-mode -m unittest discover -s {toxinidir}/EMGFlow-Package/
+    python -m coverage comine --append
+    python -m coverage report -m --skip-covered -i
+    python -m coverage json -i
 
 [testenv:coverage]
 basepython = python3.10
