### 2025-01-30 bc7c5ef
Github actions test 4
Fixed typo and changed flags
