### 2025-02-03 01782cc
Github actions test 5
Updated the path to coverage.json instead of .coverage.*.

I thought that .coverage was a file path from . (the root folder) to the coverage file, but it might be making GitHub look for files with a coverage extension instead.
