+++ .github/workflows/tests.yml
@@ -44,7 +44,7 @@ jobs:
       uses: actions/upload-artifact@v4
       with:
         name: covdata
-        path: .coverage.*
+        path: coverage.json
 
   coverage:
     name: Coverage
