+++ .github/workflows/tests.yml
@@ -44,7 +44,7 @@ jobs:
       uses: actions/upload-artifact@v4
       with:
         name: covdata
-        path: coverage.json
+        path: .coverage.*
 
   coverage:
     name: Coverage
@@ -68,7 +68,6 @@ jobs:
         uses: actions/download-artifact@v4
         with:
           name: covdata
-          path: coverage
 
       - name: Combine
         run: |
