### 2025-02-03 903a9f6
Github actions test 7
Removed combine and json report generation from testenv commands.
