+++ tox.ini
@@ -8,9 +8,6 @@ deps =
     -r {toxinidir}/EMGFlow-Package/requirements_dev.txt
 commands =
     python -m coverage run -p -m unittest discover -s {toxinidir}/EMGFlow-Package/
-    python -m coverage combine
-    python -m coverage report -m --skip-covered -i
-    python -m coverage json -i
 
 [testenv:coverage]
 basepython = python3.10
