### 2025-02-03 6b5c8fd
Github actions test 8
Added debugging step printing out files that have a coverage extension, set upload step to include hidden files.
