+++ .github/workflows/tests.yml
@@ -45,6 +45,7 @@ jobs:
       with:
         name: covdata
         path: .coverage.*
+        include-hidden-files: true
 
   coverage:
     name: Coverage
