+++ tox.ini
@@ -8,7 +8,7 @@ deps =
     -r {toxinidir}/EMGFlow-Package/requirements_dev.txt
 commands =
     python -m coverage run -p -m unittest discover -s {toxinidir}/EMGFlow-Package/
-    ls -al .coverage*
+    python -m ls -al .coverage*
 
 [testenv:coverage]
 basepython = python3.10
