### 2025-02-04 e0f9c7b
Github actions test 10
Moved code printing out .coverage files to the yml file
