+++ tox.ini
@@ -12,7 +12,7 @@ commands =
 [testenv:coverage]
 basepython = python3.10
 commands =
-    python -m coverage combine
+    python -m coverage combine .coverage.*
     python -m coverage report -m --skip-covered -i
     python -m coverage json -i
     
