### 2025-02-04 2dfa21e
Github actions test 15
Updated tox file to use combine on each coverahe file
