+++ tox.ini
@@ -12,7 +12,7 @@ commands =
 [testenv:coverage]
 basepython = python3.10
 commands =
-    python -m coverage combine .coverage.*
+    /bin/sh -c 'for f in .coverage.*; do coverage combine $f; done'
     python -m coverage report -m --skip-covered -i
     python -m coverage json -i
     
