+++ tox.ini
@@ -11,6 +11,7 @@ commands =
 
 [testenv:coverage]
 basepython = python3.10
+allowlist_externals = /bin/sh
 commands =
     /bin/sh -c 'for f in .coverage.*; do coverage combine $f; done'
     python -m coverage report -m --skip-covered -i
