### 2025-02-21 #7696 @wbaccinelli
> ### General checks
>  **Contribution and authorship:** Has the submitting author ([@WiIIson](https://github.com/WiIIson)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

I see that two author are listed, but [@WiIIson](https://github.com/WiIIson) is the only contributor to the software. Can the contribution of the second author be explicitly explained?
