### 2025-02-21 #7696 @srlivingstone
> I see that two author are listed, but [@WiIIson](https://github.com/WiIIson) is the only contributor to the software. Can the contribution of the second author be explicitly explained?

Hi Dr Baccinelli. [@WiIIson](https://github.com/WiIIson)) is the lead author of the EMGFlow package code. As his thesis [supervisor](https://affectivedatascience.com/people.html), I am the senior author on the paper, which I co-wrote with Mr Conley. I supervised the development of EMGFlow. However, Mr Conley was responsible for writing the code and making all commits. If you like, we could add an "author contributions" section to our manuscript to make this explicit.
