### 2025-02-21 fdb87ee
Updated JOSS Paper
Updated JOSS paper with an author contributions section.
