### 2025-02-21 #7696 @WiIIson
Hi there,

Just confirming that I've updated the documentation to include the author contributions section:
https://github.com/WiIIson/EMGFlow-Python-Package/blob/main/papers/JOSS/paper.pdf
