### 2025-03-02 WiIIson/EMGFlow-Python-Package#1 @samiralavi
JOSS Review: Please add a working example file with sample raw data
I couldn't find a working example due to missing sample data. All of the example codes in the documentation require `/data` with specific raw data, which basically stops readers from running the examples as those data are not provided even in samples. Could you please provide a simple dataset that can be used in the examples to make the example self-sufficient to run?
