### 2025-03-02 WiIIson/EMGFlow-Python-Package#2 @samiralavi
JOSS Review: Please remove temporary artifacts and no-source binaries from the repository
The repository structure is good, however, there are some files that shouldn't have been added to the git source control. Generally, only source files should be version controlled. Here are some examples to remove:
- `EMGFlow-Package/dist`
- `__pycache__`
