### 2025-03-02 #7696 @samiralavi
@faroit I have finished my review. Only two items in the checklist are remaining, which I will update once the authors address the issues that I have created on the software repository. Here are the links to the issues:
https://github.com/WiIIson/EMGFlow-Python-Package/issues/1
https://github.com/WiIIson/EMGFlow-Python-Package/issues/2

@WiIIson please let me know if you think those issues are not valid. Thanks.
