### 2025-03-03 65e4736
Added article DOIs
Added DOIs for the following papers:
- Changes in electromyogram power spectra (1983)
- An electromyographic index (1977)
- A novel feature (2009)
