+++ papers/JOSS/paper.bib
@@ -146,6 +146,7 @@
 
 @article{lindstrom1977electromyographic,
   title={An electromyographic index for localized muscle fatigue},
+  doi = {https://doi.org/10.1152/jappl.1977.43.4.750},
   author={Lindstrom, L\_ and Kadefors, ROLAND and Petersen, INGEMAR},
   journal={Journal of applied physiology},
   volume={43},
@@ -207,6 +208,7 @@
 
 @article{phinyomark2009novel,
   title={A novel feature extraction for robust EMG pattern recognition},
+  doi = {https://doi.org/10.48550/arXiv.0912.3973},
   author={Phinyomark, Angkoon and Limsakul, Chusak and Phukpattaranont, Pornchai},
   journal={arXiv preprint arXiv:0912.3973},
   year={2009}
@@ -343,6 +345,7 @@
 
 @article{van1983changes,
   title={Changes in electromyogram power spectra of facial and jaw-elevator muscles during fatigue},
+  doi = {https://doi.org/10.1152/jappl.1983.54.1.51},
   author={Van Boxtel, A and Goudswaard, P and Van der Molen, GM and Van Den Bosch, WE},
   journal={Journal of Applied Physiology},
   volume={54},
