### 2025-03-03 a013e02
Updated paper.bib DOIs
Some DOIs were formatted incorrectly
