### 2025-03-03 27cd3e0
Update pyproject.toml
Split author names
