+++ EMGFlow-Package/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "EMGFlow"
-version = "1.0.16"
+version = "1.0.17"
 dependencies = [
     "pandas",
     "numpy",
@@ -16,7 +16,11 @@ dependencies = [
 ]
 requires-python = ">=3.9"
 authors = [
-    {name = "William Conley and Steven Livingstone", email = "william@cconley.ca"}
+    {name = "D. William Conley", email = "william@cconley.ca"},
+    {"Steven R. Livingstone"},
+]
+maintainers = [
+    {name = "D. William Conley", email = "william@cconley.ca"}
 ]
 description = "A general EMG processing and feature extraction package."
 readme = "README.md"
