### 2025-03-03 #7696 @WiIIson
> > Hi Dr Baccinelli. [@WiIIson](https://github.com/WiIIson)) is the lead author of the EMGFlow package code. As his thesis [supervisor](https://affectivedatascience.com/people.html), I am the senior author on the paper, which I co-wrote with Mr Conley. I supervised the development of EMGFlow. However, Mr Conley was responsible for writing the code and making all commits. If you like, we could add an "author contributions" section to our manuscript to make this explicit.
>
> Thank you for the explanation [@srlivingstone](https://github.com/srlivingstone). I think that the "author contributions" section would be a fair addition, even though is not needed. I suggest, on the other hand, splitting the names of the authors in the [pyproject.toml](https://github.com/WiIIson/EMGFlow-Python-Package/blob/cf25ca9b6191db24a145164f8595f9007ecd9657/EMGFlow-Package/pyproject.toml#L19).

Thank you for the comment. I have updated the pyproject.toml per your suggestion.
