### 2025-03-10 WiIIson/EMGFlow-Python-Package#3 @wbaccinelli
JOSS Review suggestion: integrate documentation in the GitHub page
While the package documentation is available in the repository, it would be good to integrate it in the `Documentation` section of the existing [GitHub page](https://wiiison.github.io/EMGFlow-Python-Package/#documentation), to make it easier to consult for the users.
I also suggest to automatize the documentation generation process using tools such as [MkDocs](https://www.mkdocs.org) or [Sphinx](https://www.sphinx-doc.org) to increase the maintainability.
