### 2025-03-10 WiIIson/EMGFlow-Python-Package#4 @wbaccinelli
JOSS review: error encountered in `SmoothFilterSignals`
When running the provided [Simple Example](https://github.com/WiIIson/EMGFlow-Python-Package/blob/main/docs/01%20Examples.md#simple-example) with the [dataset](https://zenodo.org/records/6403363) indicated in the paper, I get the following error when running
```python
EMGFlow.SmoothFilterSignals(band_path, smooth_path, sampling_rate, smooth_window)
```
```
Cell In[10], line 4
      2 EMGFlow.NotchFilterSignals(raw_path, notch_path, sampling_rate, notch_vals)
      3 EMGFlow.BandpassFilterSignals(notch_path, band_path, sampling_rate, band_low, band_high)
----> 4 EMGFlow.SmoothFilterSignals(band_path, smooth_path, sampling_rate, smooth_window)

File ~\anaconda3\envs\emgflow\lib\site-packages\EMGFlow\PreprocessSignals.py:836, in SmoothFilterSignals(in_path, out_path, window_size, cols, expression, exp_copy, file_ext, method, sigma)
    833         cols.remove('Time')
    835 # Apply filter to columns
--> 836 for col in cols:
    837     if method == 'rms':
    838         data = ApplyRMSSmooth(data, col, window_size)

TypeError: 'int' object is not iterable

```

Could you, please, address this?
