### 2025-03-10 WiIIson/EMGFlow-Python-Package#5 @wbaccinelli
JOSS review: example in the documentation using `AnalyzeSignals` does not work
When running following line reported in the example usage from the documentation, an error occurs:

```python
df = EMGFlow.AnalyzeSignals(band_path, smooth_path, feature_path, sampling_rate, cols)
```

```
Cell In[11], line 1
----> 1 df = EMGFlow.AnalyzeSignals(band_path, smooth_path, feature_path, sampling_rate, cols)

AttributeError: module 'EMGFlow' has no attribute 'AnalyzeSignals'
```

The documentation should be updated, or the error should be fixed.
