### 2025-03-10 WiIIson/EMGFlow-Python-Package#6 @wbaccinelli
JOSS review: add separate `contributing.md` file
Guidelines for the community to contribute are, at the moment, briefly described in the `README.md` file. I suggest moving the to a devoted `CONTRIBUTING.md` file in the docs, and referring to this file in the `README.md`, as suggested [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#about-contributing-guidelines). In this way it will be easier also to expand the information in the future (e.g., project  style, review process, branches naming), without making the landing point of the package to dense.
