### 2025-03-10 WiIIson/EMGFlow-Python-Package#7 @wbaccinelli
JOSS review: add documentation about the data format
I assume that the data in the input files need to have a specific format, such as the number and naming of the columns, the data types contained in the columns and so on. These specifications should be clearly stated in the documentation to enable the users to prepare their data before using the package.
