### 2025-03-10 WiIIson/EMGFlow-Python-Package#8 @wbaccinelli
JOSS review: add section to explain how to run the tests.
The testing scripts are available and they are run in the CI/CD, but it would be helpful for possible future contributors, to have a paragraph in the documentation explaining how to run them locally. In this way it will be easier for them to test the modifications as intended.
