### 2025-03-10 #7696 @wbaccinelli
@faroit @WiIIson @srlivingstone, I completed my review.
The package is well organized and documented, thank you for your work. I have some points that I raised. I think that some of them would need to be addressed before ticking all the boxes, while other are suggestions that might help you.

The issues that should be addressed for this publication are:
- Fix the errors found when running the examples ( see [this](https://github.com/WiIIson/EMGFlow-Python-Package/issues/4) and [this](https://github.com/WiIIson/EMGFlow-Python-Package/issues/5) issue)
- Add specifications about the data format (see [this issue](https://github.com/WiIIson/EMGFlow-Python-Package/issues/7))
- Add documentation about running tests (see [this issue](https://github.com/WiIIson/EMGFlow-Python-Package/issues/8))
- I also agree with the issues opened by @samiralavi. In particular [this one](https://github.com/WiIIson/EMGFlow-Python-Package/issues/1). While a dataset is indicated in the text of the paper, it should be also indicated in the package documentation to let the users work with the code.

Optional issues:
- Move the documentation to GitHub pages and automatize its release (see [this issue](https://github.com/WiIIson/EMGFlow-Python-Package/issues/3))
- Add a separate file for the community guidelines, and expand them with more information (see [this issue](https://github.com/WiIIson/EMGFlow-Python-Package/issues/6))

I hope that the inputs will help in improving the overall package usability.
