### 2025-03-10 WiIIson/EMGFlow-Python-Package#5 @faroit
cross-referencing the review issue to keep track of the status https://github.com/openjournals/joss-reviews/issues/7696
