### 2025-03-11 89ec0a4
Allowed .tox files to log automation
