+++ EMGFlow-Package/src/EMGFlow/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.16"
+__version__ = "1.0.17"
 from .FileAccess import *
 from .PreprocessSignals import *
 from .ExtractFeatures import *
