+++ pages-docs/06 ExtractFeatures API.md
@@ -1,0 +1,1371 @@
+# `ExtractFeatures` Module Documentation
+
+This module takes preprocessed data, and extracts features from the sEMG signal that capture information in both time and frequency domains. The main function to do this is `ExtractFeatures`. Within this call, individual features are calculated by their own functions, allowing them to be incorporated into your own workflow.
+
+---
+
+## Basic Time-Series Statistics
+
+The `AnalyzeSignal` function calculates some basic statistics that don't involve their own functions. This includes:
+- Minimum voltage
+- Maximum voltage
+- Mean voltage
+- Standard deviation of voltage
+- Skew of voltage
+- Kurtosis of voltage
+- Maximum frequency
+
+The values returned will be a voltage equivalent to the units used in the data provided.
+
+Most of these are commonly understood concepts:
+- **Minimum** is calculated with `np.min`
+- **Maximum** is calculated with `np.max`
+- **Mean** is calculated with `np.mean`
+- **Standard Deviation** is calculated with `np.std`
+
+Skew and kurtosis are lesser known statistical measurements.
+
+### Skew
+
+Skewness describes the symmetry of a dataset, considered more skewed the less symmetrical the left and right distributions of the median are.
+
+Skew is calculated as follows:
+```math
+s=\frac{\frac{\mu-M\\_o}{\sigma}}{\frac{3(\mu-M\\_d)}{\sigma}}
+```
+- $\mu$ <-- Mean
+- $\sigma$ <-- Standard deviation
+- $M_o$ <-- Mode
+- $M_d$ <-- Median
+
+**Skew** is calculated with `scipy.stats.skew`
+
+### Kurtosis
+
+Kurtosis describes the amount of data in the tails of a bell curve of a distribution.
+
+Kurtosis is calculated as follows:
+```math
+k=\frac{1}{N}\sum\\_{i=1}^N\left(\frac{x\\_i-\mu}{\sigma}\right)^4
+```
+- $\mu$ <-- Mean
+- $\sigma$ <-- Standard deviation
+- $N$ <-- Number of data points
+
+**Kurtosis** is calculated with `scipy.stats.kurtosis`
+
+---
+
+## `CalcIEMG`
+
+**Description**
+
+Calculates the Integrated EMG (IEMG) of the signal. The IEMG measures the area under the curve of the signal, which can provide useful information about muscle activity. In an EMG signal, the IEMG describes when the muscle begins contracting, and is related to the signal sequence firing point (Phinyomark et al., 2009).
+
+```python
+CalcIEMG(Signal, col, sr)
+```
+
+**Theory**
+
+In the reference, the IEMG does not account for the sampling rate. Two signal recordings with the same shape but different sampling rate would have different results since we are integrating with respect to time. As such, the calculation made here will include multiplying by sampling rate.
+
+The IEMG is calculated as follows:
+```math
+\text{IEMG}=s\\_r\sum\\_{i=1}^N|x\\_i|
+```
+- $s_r$ <-- Sampling rate
+- $N$ <-- Number of data points
+
+(Spiewak et al., 2018)
+
+**Parameters**
+
+`Signal`: pd.DataFrame 
+- Should have one column called "`Time`" for the time indexes, and other named columns for the values at those times.
+
+`col`: str
+- String name of a column in `Signal` the filters are being applied to.
+
+`sr`: int/float
[…]
