### 2025-03-16 4e1cac2
Added sample data
Added:
- Sample data
- Method for loading sample data (into sample_data variable)
- Changed toml file to include sample data
