+++ EMGFlow-Package/pyproject.toml
@@ -30,5 +30,10 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 
+[tool.hatch.build]
+include = [
+    "src/EMGFlow/data/*"
+]
+
 [project.urls]
 Home = "https://github.com/WiIIson/EMGFlow-Python-Package"
