### 2025-03-16 dfbb64d
Updated example docs
Updated example docs with an example using the sample data
