### 2025-03-16 908e01b
Updated examples
Removed unnecessary lines
