### 2025-03-23 7f0864e
Updated examples
Updated examples: fixed the first example (simple example using the built-in data)
