### 2025-03-24 WiIIson/EMGFlow-Python-Package#1 @WiIIson
Hi @samiralavi, thanks for the suggestion. I have added sample data to the toolkit, and updated the [examples pages](https://wiiison.github.io/EMGFlow-Python-Package/render-doc.html?doc=01) which shows how to access and use these inbuilt data.
