### 2025-03-24 WiIIson/EMGFlow-Python-Package#2 @WiIIson
Thanks for catching that issue. We've updated the .gitignore and these files are no longer included in recent comments.
