### 2025-03-24 #7696 @WiIIson
> [@WiIIson](https://github.com/WiIIson) please let me know if you think those issues are not valid. Thanks.

Hi @samiralavi,

Thank you for those helpful suggestions. As we detail in the individual Issue threads [[1](https://github.com/WiIIson/EMGFlow-Python-Package/issues/1), [2](https://github.com/WiIIson/EMGFlow-Python-Package/issues/2)], we've added sample data to the package, updated the example docs, and updated the .gitignore to remove unwanted files.
