### 2025-03-24 16b6d6a
Updated README badges and links
Added a versions badge

Version and PyPi badges now link to the PyPi page
