### 2025-03-29 820cac0
Updated examples
Updated AnalyzeSignals calls to ExtractFeatures
