### 2025-03-29 d8a139c
Fixed SmoothFilterSignals bug in examples
Removed sampling_rate parameter from examples using SmoothFilterSignals
