### 2025-03-29 eeee5fe
Update 01 Examples.md
Updated example file paths and added description of how to run the tests
