### 2025-03-29 a95fead
Update 03 Processing Pipeline.md
Added comment about file format
