### 2025-03-30 8baea88
Update 03 Processing Pipeline.md
Removed duplication from API descriptions
