+++ pages-docs/03 Processing Pipeline.md
@@ -36,27 +36,8 @@ This module provides preprocessing functions for cleaning sEMG signals prior to 
 
 ### `NotchFilterSignals()`
 
-**Description**
 
 Notch filtering involves filtering specific frequencies. This is typically due to some sort of interference, such as the power source of the device taking the reading.
-
-**Parameters**
-
-Notch filtering is controlled by the function `NotchFilterSignals()`, with the following main parameters:
-- `in_path` <-- File path to input folder
-- `out_path` <-- File path to output folder
-- `sampling_rate` <-- Sampling rate of the data
-- `notch` <-- Notch filters to be applied
-
-`in_path` refers to the folder containing the signal data. The function will be able to detect signals in the folder, and in any subfolders.
-
-`out_path` refers to the folder where the filtered signals will be output to. The function will automatically recreate the file structure of the input folder with additional subfolders.
-
-`sampling_rate` refers to the sampling rate of the data. The function will assume that each signal, and each column, is using the same sampling rate.
-
-`notch` refers to the notch filters being applied to the signal. This takes the form of a list of `(Hz,Q)` tuples. `Hz` is the frequency to which the filter is being applied, and `Q` is the Q-factor (strength of the filter, where a lower number is a stronger filter). The Q-factor needs to be explored to find a suitable value.  When the Q-factor is too high, there will be barely any effect.  When the Q_factor is too, important frequencies around the target being filtered out. All filters in the list will be applied to each column and signal.
-
-**Value-Added**
 
 `NotchFilterSignals()` provides flexibility for use in different regions of the world. Some filtering packages only provide notch filtering for 60Hz, the frequency where power can interfere with signal readings.  However, other regions use 50Hz frequencies.
 
@@ -64,28 +45,7 @@ For more information about further customizations and detail about `NotchFilterS
 
 ### `BandpassFilterSignals()`
 
-**Description**
-
 Bandpass filtering involves specifying a range of frequencies to keep, and removing all other frequencies outside this range. This is useful to remove interference outside the most meaningful range of readings. 
-
-**Parameters**
-
-Bandpass filtering is controlled by the function `BandpassFilterSignals()`, with the following parameters:
-- `in_path` <-- File path to input folder
-- `out_path` <-- File path to output folder
-- `sampling_rate` <-- Sampling rate of the data
-- `low` <-- Lower frequency range
-- `high` <-- Upper frequency range
-
-`in_path` refers to the folder containing the signal data. The function will be able to detect signals in the folder, and in any subfolders.
-
-`out_path` refers to the folder where the filtered signals will be output to. The function will automatically recreate the file structure of the input folder with additional subfolders.
-
-`sampling_rate` refers to the sampling rate of the data. The function will assume that each signal, and each column, is using the same sampling rate.
-
-`low` and `high` are the thresholds for the bandpass filter. All frequencies outside the range provided are filtered out of the signal. The defaults respectively are 20 and 450.
-
-**Value-Added**
 
 `BandpassFilterSignals()` uses bandpass thresholds of 20Hz and 450Hz, as this is default for EMG signals (De Luca et al., 2010). However, there is some disagreement within literature for different muscels, so `BandpassFilterSignals()` provides the option to change the thresholds.
 
@@ -94,25 +54,6 @@ For more information about further customizations and detail about `BandpassFilt
 ### `SmoothFilterSignals()`
 
 Smoothing involves limiting the impacts of noise and outliers in the signal. By default, this function uses the RMS smoothing method, as it is the best choice for smoothing EMG signals (RENSHAW et al., 2010).
-
-Smoothing is controlled by the function `SmoothFilterSignals()`, with the following main parameters:
[…]
