+++ docs/.vitepress/config.ts
@@ -3,12 +3,12 @@ import { defineConfig } from 'vitepress';
 // refer https://vitepress.dev/reference/site-config for details
 export default defineConfig({
   lang: 'en-US',
-  title: 'VitePress',
-  description: 'Vite & Vue powered static site generator.',
+  title: 'EMGFlow',
+  description: 'EMG Workflow package.',
 
   themeConfig: {
-    nav: [
-      { text: 'Example', link: '/example' },
+    //nav: [
+    //  { text: 'Example', link: '/example' },
 
       // {
       //   text: 'Dropdown Menu',
@@ -20,13 +20,20 @@ export default defineConfig({
       // },
 
       // ...
-    ],
+    //],
 
     sidebar: [
       {
         // text: 'Guide',
         items: [
-          { text: 'Example', link: '/example' },
+          { text: 'Examples', link: '/01 Examples' },
+          { text: 'Background', link: '/02 Background' },
+          { text: 'Processing Pipeline', link: '/03 Processing Pipeline' },
+          { text: 'FileAccess API', link: '/04 FileAccess API' },
+          { text: 'PreprocessSignals API', link: '/05 PreprocessSignals API' },
+          { text: 'ExtractFeatures API', link: '/06 ExtractFeatures' },
+          { text: 'OutlierFinder API', link: '/07 OutlierFinder API' },
+          { text: 'PlotSignals API', link: '/08 PlotSignals API' },
           // ...
         ],
       },
