+++ docs/index.md
@@ -1,15 +1,14 @@
----
-layout: home
+# EMGFlow
 
-hero:
-  name: VitePress
-  text: Vite & Vue powered static site generator.
-  tagline: Simple, powerful, and performant. Meet the modern SSG framework you've always wanted.
-  actions:
-    - theme: brand
-      text: Get Started
-      link: /example
-    - theme: alt
-      text: View on GitHub
-      link: https://github.com/vuejs/vitepress
 ---
+
+This is the main page of the EMGFlow package for Python. The source code for this project is also hosted on GitHub where documentation is provided. See the menu bar for quick links to where the package can be downloaded, and the documentation pane on the side for guides on how to use functions as well as examples.
+
+EMGFlow is a Python package for researchers and clinicians to engage in signal processing. EMGFlow provides a broad range of functions to meet your EMG signal processing needs, without prescribing a specific workflow.
+
+EMGFlow follows open standards of data processing, such as CSV files and Pandas data frames to allow easy integration. With functions to extract over 30 different features according to your needs, EMGFlow provides a uniquely deep feature extraction.
+
+While several packages exist for processing physiological and neurological signals, support for surface electromyography (sEMG) has remained limited. Many packages lack a comprehensive set of features that can be extracted from sEMG data, leaving researchers to use a patchwork of tools.
+
+Other packages are orientated around event detection in individual recordings and use GUI-based workflow that requires greater manual intervention. While this design works well for single person recordings, it complicates the extraction of features from large datasets, such as those common to machine learning.
+
