### 2025-03-30 7b2befd
Update 03 Processing Pipeline.md
Removed redundant information
