+++ docs/.vitepress/config.ts
@@ -1,43 +1,72 @@
-import { defineConfig } from 'vitepress';
+import { defineConfig } from 'vitepress'
 
-// refer https://vitepress.dev/reference/site-config for details
+// https://vitepress.dev/reference/site-config
 export default defineConfig({
-  lang: 'en-US',
-  title: 'EMGFlow',
-  description: 'EMG Workflow package.',
-  base: '/EMGFlow-Python-Package/',
+  title: "EMGFlow",
+  titleTemplate: 'EMG Workflow',
+  description: "The open workflow for EMG signal processing and feature extraction",
+
+  // Update time of last content
+  // https://vitepress.dev/reference/default-theme-last-updated
+  lastUpdated: true,
 
   themeConfig: {
-    //nav: [
-    //  { text: 'Example', link: '/example' },
-
-      // {
-      //   text: 'Dropdown Menu',
-      //   items: [
-      //     { text: 'Item A', link: '/item-1' },
-      //     { text: 'Item B', link: '/item-2' },
-      //     { text: 'Item C', link: '/item-3' },
-      //   ],
-      // },
-
-      // ...
-    //],
+    // https://vitepress.dev/reference/default-theme-config
+    nav: [
+      { text: 'Guide', link: '/guide/what-is-emgflow' }, 
+      { text: 'Reference', link: '/reference/api-overview' },
+      { text: 'About Us', link: '/about/authors'}
+    ],
 
     sidebar: [
       {
-        // text: 'Guide',
+        text: 'Introduction',
         items: [
-          { text: 'Examples', link: '/01 Examples' },
-          { text: 'Background', link: '/02 Background' },
-          { text: 'Processing Pipeline', link: '/03 Processing Pipeline' },
-          { text: 'FileAccess API', link: '/04 FileAccess API' },
-          { text: 'PreprocessSignals API', link: '/05 PreprocessSignals API' },
-          { text: 'ExtractFeatures API', link: '/06 ExtractFeatures API' },
-          { text: 'OutlierFinder API', link: '/07 OutlierFinder API' },
-          { text: 'PlotSignals API', link: '/08 PlotSignals API' },
-          // ...
-        ],
+          { text: 'What is EMGFlow?', link: '/guide/what-is-emgflow' },
+          { text: 'Getting started', link: '/guide/getting-started' },
+          { text: 'Examples', link: '/guide/examples' },
+          { text: 'About electromyography', link: '/guide/about-emg' },
+        ]
       },
+
+      {
+        text: 'Reference',
+        items: [
+          { text: 'Overview', link: '/reference/api-overview' },
+          { text: 'Pre-processing', link: '/reference/preprocess-signals' },
+          { text: 'Feature extraction', link: '/reference/feature-extraction' },
+          { text: 'Plotting signals', link: '/reference/plot-signals' },
+          { text: 'Outlier detection', link: '/reference/outlier-detection' },
+          { text: 'File access', link: '/reference/file-access' },
+        ]
+      },
+
+      {
+        text: 'About Us',
+        items: [
+          { text: 'Authors', link: '/about/authors' },
+          { text: 'Citation', link: '/about/citation' },
+          { text: 'Contact us', link: '/about/contact' },
+        ]
+      }, 
     ],
+
+    socialLinks: [
+      { icon: 'github', link: 'https://github.com/WiIIson/EMGFlow-Python-Package' }
+    ],
+
+    // Site footer
+    footer: {
+      message: 'Site built with <a href="https://vitepress.dev/">VitePress</a>',
+      copyright: 'Created by <a href="https://wconley.ca/">William Conley</a> and <a href="https://affectivedatascience.com">Steven R. Livingstone</a>'
+    }, 
[…]
