+++ docs/reference/api-overview.md
@@ -1,0 +1,243 @@
+---
+outline: deep
+---
+
+# API Overview
+
+EMGFlow is broken into 5 models - FileIO, preprocessing signals, feature extraction, detecting outliers, and generating visualizations.
+
+## Modules
+
+### `FileAccess`
+
+These functions provide helper methods for accessing files, and are mostly used internally by the package.
+
+One notable function, `MapFiles`, takes a path to a folder, and generates a dictionary of paths to files contained within. This makes it easier to create a loop over subfiles, reading them in and performing analysis.
+
+`MapFiles` forms the basis for the two modes of analysis offered by EMGFlow - automated, or manual.
+
+The "automated" mode makes the processing pipeline much simpler. In these functions, a file location is provided, default parameters are set, and a file output location is set. The functions then apply the filters to each file found in the folder, and output the filtered files to the output folder. Notable functions include:
+- `NotchFilterSignals`
+- `BandpassFilterSignals`
+- `SmoothFilterSignals`
+
+The "manual" mode allows for additional customization of processing. In these functions, a dataframe is provided, filter parameters are set, and the dataframe is returned. These functions are useful if the processing pipeline requires further additional processing before being outputted, of if the project is not large-scale enough to warrent batch processing. Notable functions include:
+- `ApplyNotchFilters`
+- `ApplyBandpassFilter`
+- `ApplyRMSSmooth`
+
+For more information about file accessing functions, see [FileAccess API](./file-access.md).
+
+### `PreprocessSignals`
+
+This module provides preprocessing functions for cleaning sEMG signals prior to their use in feature extraction. Signal processing is broken into 3 parts: notch filtering, bandpass filtering and smoothing. Each part has additional functions that support more specific needs, explained in more detail in the module descriptions.
+
+#### `NotchFilterSignals()`
+
+##### Description
+
+Notch filtering involves filtering specific frequencies. This is typically due to some sort of interference, such as the power source of the device taking the reading.
+
+##### Parameters
+
+Notch filtering is controlled by the function `NotchFilterSignals()`, with the following main parameters:
+- `in_path` <-- File path to input folder
+- `out_path` <-- File path to output folder
+- `sampling_rate` <-- Sampling rate of the data
+- `notch` <-- Notch filters to be applied
+
+`in_path` refers to the folder containing the signal data. The function will be able to detect signals in the folder, and in any subfolders.
+
+`out_path` refers to the folder where the filtered signals will be output to. The function will automatically recreate the file structure of the input folder with additional subfolders.
+
+`sampling_rate` refers to the sampling rate of the data. The function will assume that each signal, and each column, is using the same sampling rate.
+
+`notch` refers to the notch filters being applied to the signal. This takes the form of a list of `(Hz,Q)` tuples. `Hz` is the frequency to which the filter is being applied, and `Q` is the Q-factor (strength of the filter, where a lower number is a stronger filter). The Q-factor needs to be explored to find a suitable value.  When the Q-factor is too high, there will be barely any effect.  When the Q_factor is too, important frequencies around the target being filtered out. All filters in the list will be applied to each column and signal.
+
+##### Value-Added
+
+`NotchFilterSignals()` provides flexibility for use in different regions of the world. Some filtering packages only provide notch filtering for 60Hz, the frequency where power can interfere with signal readings.  However, other regions use 50Hz frequencies.
+
+For more information about further customizations and detail about `NotchFilterSignals()`, see [PreprocessSignals API](./preprocess-signals.md).
+
+#### `BandpassFilterSignals()`
+
+**Description**
+
[…]
