+++ docs/about/contact.md
@@ -1,0 +1,10 @@
+---
+
+---
+
+# Contact Us :email:
+
+## Authors :scientist:
+
+- <span class="email">douglas<b>.conley</b>@ontariotechu.net</span>
+- <span class="email">steven<b>.livingstone</b>@ontariotechu.ca</span>

[…]
