### 2025-04-03 74ce98c
Fixed LaTeX rendering
