+++ docs/reference/feature-extraction.md
@@ -30,9 +30,7 @@ Skew and kurtosis are lesser known statistical measurements.
 Skewness describes the symmetry of a dataset, considered more skewed the less symmetrical the left and right distributions of the median are.
 
 Skew is calculated as follows:
-```math
-s=\frac{\frac{\mu-M\\_o}{\sigma}}{\frac{3(\mu-M\\_d)}{\sigma}}
-```
+$$s=\frac{\frac{\mu-M_o}{\sigma}}{\frac{3(\mu-M_d)}{\sigma}}$$
 - $\mu$ <-- Mean
 - $\sigma$ <-- Standard deviation
 - $M_o$ <-- Mode
@@ -45,9 +43,9 @@ s=\frac{\frac{\mu-M\\_o}{\sigma}}{\frac{3(\mu-M\\_d)}{\sigma}}
 Kurtosis describes the amount of data in the tails of a bell curve of a distribution.
 
 Kurtosis is calculated as follows:
-```math
-k=\frac{1}{N}\sum\\_{i=1}^N\left(\frac{x\\_i-\mu}{\sigma}\right)^4
-```
+$$
+k=\frac{1}{N}\sum_{i=1}^N\left(\frac{x_i-\mu}{\sigma}\right)^4
+$$
 - $\mu$ <-- Mean
 - $\sigma$ <-- Standard deviation
 - $N$ <-- Number of data points
@@ -62,18 +60,18 @@ k=\frac{1}{N}\sum\\_{i=1}^N\left(\frac{x\\_i-\mu}{\sigma}\right)^4
 
 Calculates the Integrated EMG (IEMG) of the signal. The IEMG measures the area under the curve of the signal, which can provide useful information about muscle activity. In an EMG signal, the IEMG describes when the muscle begins contracting, and is related to the signal sequence firing point (Phinyomark et al., 2009).
 
-```python
+$$python
 CalcIEMG(Signal, col, sr)
-```
+$$
 
 **Theory**
 
 In the reference, the IEMG does not account for the sampling rate. Two signal recordings with the same shape but different sampling rate would have different results since we are integrating with respect to time. As such, the calculation made here will include multiplying by sampling rate.
 
 The IEMG is calculated as follows:
-```math
-\text{IEMG}=s\\_r\sum\\_{i=1}^N|x\\_i|
-```
+$$
+\text{IEMG}=s_r\sum_{i=1}^N|x_i|
+$$
 - $s_r$ <-- Sampling rate
 - $N$ <-- Number of data points
 
@@ -103,7 +101,7 @@ Raises an error if `sr` is less or equal to 0.
 
 **Example**
 
-```python
+python
 # Calculate the IEMG of SignalDF, for column 'column1'
 IEMG = EMGFlow.CalcIEMG(SignalDF, 'column1', 2000)
 ```
@@ -116,16 +114,16 @@ IEMG = EMGFlow.CalcIEMG(SignalDF, 'column1', 2000)
 
 Calculates the Mean Absolute Value (MAV) of the signal. In an EMG signal, the MAV describes the muscle contraction level (Phinyomark et al., 2009).
 
-```python
+$$python
 CalcMAV(Signal, col)
-```
+$$
 
 **Theory**
 
 The MAV is calculated as follows:
-```math
-\text{MAV}=\frac{1}{N}\sum\\_{i=1}^N|x\\_i|
-```
+$$
[…]
