+++ README.md
@@ -1,0 +1,3 @@
+# gh-pages
+
+This repository branch contains the source files for the [documentation pages](https://wiiison.github.io/EMGFlow-Python-Package/).
