+++ README.md
@@ -15,15 +15,11 @@ EMGFlow also includes an easy method for producing detailed graphs of EMG signal
 The simple navigation menu allows you to switch between which files you see as well as which layers are overlayed:
 ![Example 1](EMGFlow_GUI.gif)
 
----
-
 ## Statement Of Need
 
 While several packages exist for processing physiological and neurological signals, support for surface electromyography (sEMG) has remained limited. Many packages lack a comprehensive set of features that can be extracted from sEMG data, leaving researchers to use a patchwork of tools. Other packages are orientated around event detection in individual recordings and use GUI-based workflow that requires greater manual intervention. While this design works well for single person recordings, it complicates the extraction of features from large datasets, such as those common to machine learning.
 
 The EMGFlow package – a portmanteau of EMG and Workflow - fills this gap. Our package provides a flexible pipeline for the extraction of a wide range of sEMG features, with a scalable design that is suited to the processing of large datasets.
-
----
 
 ## Example
 
@@ -56,8 +52,6 @@ EMGFlow.SmoothFilterSignals(band_path, smooth_path, sampling_rate, smooth_window
 df = EMGFlow.ExtractFeatures(band_oath, smooth_path, feature_path, sampling_rate)
 ```
 
----
-
 ## Documentation
 
 [![pypi](https://img.shields.io/badge/documentation-online-brightgreen.svg)](https://wiiison.github.io/EMGFlow-Python-Package/docs-landing-page.html)
@@ -75,8 +69,6 @@ General:
 Examples:
 - [Processing pipeline examples](https://wiiison.github.io/EMGFlow-Python-Package/render-doc.html?doc=01)
 
----
-
 ## Installation
 
 EMGFlow can be installed from PyPI:
@@ -91,13 +83,9 @@ import EMGFlow
 
 Project dependencies can be seen [here](EMGFlow-Package/pyproject.toml).
 
----
-
 ## Contributions
 
 Contributions and community guidelines can be seen [here](https://github.com/WiIIson/EMGFlow-Python-Package/blob/main/.github/CONTRIBUTING.md).
-
----
 
 ## Citations
 
