### 2025-04-04 bf52762
Update CONTRIBUTING.md
Removed line break (auto generated by github)
