+++ .github/CONTRIBUTING.md
@@ -1,7 +1,5 @@
 # Welcome to EMGFlow contributing guide
 
----
-
 We welcome contributions to the project. These can be initiated through the project's issue tracker or through a pull request. Suggestions for feature enhancements, tips, as well as general questions and concerns can also be expressed through direct interaction with contributors and developers.
 
 If you experience any challenges with this module such as bugs, test support, or feature requests, please feel free to use this [issue tracker](https://github.com/WiIIson/EMGFlow-Python-Package/issues).
