### 2025-04-04 5d553de
Update feature-extraction.md
Fixed LaTeX errors
