+++ docs/reference/feature-extraction.md
@@ -60,9 +60,9 @@ $$
 
 Calculates the Integrated EMG (IEMG) of the signal. The IEMG measures the area under the curve of the signal, which can provide useful information about muscle activity. In an EMG signal, the IEMG describes when the muscle begins contracting, and is related to the signal sequence firing point (Phinyomark et al., 2009).
 
-$$python
+```python
 CalcIEMG(Signal, col, sr)
-$$
+```
 
 **Theory**
 
@@ -114,9 +114,9 @@ IEMG = EMGFlow.CalcIEMG(SignalDF, 'column1', 2000)
 
 Calculates the Mean Absolute Value (MAV) of the signal. In an EMG signal, the MAV describes the muscle contraction level (Phinyomark et al., 2009).
 
-$$python
+```python
 CalcMAV(Signal, col)
-$$
+```
 
 **Theory**
 
@@ -160,9 +160,9 @@ MAV = EMGFlow.CalcMAV(SignalDF, 'column1', 2000)
 **Description**
 
 Calculates the Modified Mean Absolute Value (MMAV) of the signal. The MMAV is an alteration of MAV that gives more weight to values in the middle of the signal to reduce error from the beginning and end of the signal.
-$$python
+```python
 CalcMMAV(Signal, col)
-$$
+```
 
 **Theory**
 
@@ -173,7 +173,7 @@ $$
 \text{MMAV}=\frac{1}{N}\sum_{i=1}^N|x_iw_i|
 $$
 $$
-w_i=\left\{ \begin{matrix} 1 & \text{if }0.25N\le n\le 0.75N \\ 0.5 & \text{otherwise} \end{matrix} \right.
+w_i=\left{ \begin{matrix} 1 & \text{if }0.25N\le n\le 0.75N \\ 0.5 & \text{otherwise} \end{matrix} \right}.
 $$
 - $N$ <-- Number of data points
 
@@ -211,9 +211,9 @@ MMAV = EMGFlow.CalcMMAV(SignalDF, 'column1', 2000)
 
 Calculates the Simple Square Integral (SSI) of the signal. In an EMG signal, the SSI describes the energy of the signal (Phinyomark et al., 2009).
 
-$$python
+```python
 CalcSSI(Signal, col, sr)
-$$
+```
 
 **Theory**
 
@@ -265,9 +265,9 @@ SSI = EMGFlow.CalcSSI(SignalDF, 'column1', 2000)
 
 Calculates the Variance (VAR) of the signal. In an EMG signal, the VAR describes the power of the signal (Phinyomark et al., 2009).
 
-$$python
+```python
 CalcVAR(Signal, col)
-$$
+```
 
 **Theory**
 
@@ -310,9 +310,9 @@ VAR = EMGFlow.CalcVAR(SignalDF, 'column1')
 
 Calculates the V-Order of a signal. The V-Order is an alteration of VAR that takes the square root of the result.
 
-$$python
+```python
 CalcVOrder(Signal, col)
-$$
+```
 
 **Theory**
 
@@ -357,9 +357,9 @@ VOrder = EMGFlow.CalcVOrder(SignalDF, 'column1')
 
 Calculates the Root Mean Square (RMS) of a signal. In an EMG signal, the RMS provides information about the constant force, and non-fatiguing contractions of the muscles (Phinyomark et al., 2009).
 
-$$python
+```python
 CalcRMS(Signal, col)
-$$
+```
 
 **Theory**
 
@@ -403,9 +403,9 @@ RMS = EMGFlow.CalcRMS(SignalDF, 'column1')
 
[…]
