### 2025-04-04 e88665d
Update feature-extraction.md
Updated more errors in LaTeX
