### 2025-04-04 7ea4a8b
Update feature-extraction.md
Removed random "." characters from some LaTeX renders
