### 2025-04-04 9c2f920
Removed double line breaks
Removed double line breaks from documentation pages (line breaks are auto-generated by titles)
