+++ docs/reference/outlier-detection.md
@@ -2,7 +2,7 @@
 
 This module provides methods to help detect signal files that contain outliers. This helps for workflows involving batch processing of files, where it might be harder to determine if there are any patterns, or specific files that need additional filters applied.
 
----
+
 
 ## `DetectOutliers`
 
