### 2025-04-05 #7696 @srlivingstone
> > > [@WiIIson](https://github.com/WiIIson) please let me know if you think those issues are not valid. Thanks.
> >
> >
> > Hi [@samiralavi](https://github.com/samiralavi),
> > Thank you for those helpful suggestions. As we detail in the individual Issue threads [[1](https://github.com/WiIIson/EMGFlow-Python-Package/issues/1), [2](https://github.com/WiIIson/EMGFlow-Python-Package/issues/2)], we've added sample data to the package, updated the example docs, and updated the .gitignore to remove unwanted files.
>
> [@WiIIson](https://github.com/WiIIson) [@wbaccinelli](https://github.com/wbaccinelli) please take a look at the open issues and update/close them if the recent changes have addressed them already

Hi @faroit - thanks for checking in. Following @wbaccinelli and @samiralavi's suggestions, @WiIIson and I have recently made extensive revisions to [site documentation](https://wiiison.github.io/EMGFlow-Python-Package/), transitioned to the docs generator VitePress, and further revised our examples with an eye towards simplicity and consistency. Once these are complete, we'll update the respective open issues.

@faroit For our knowledge, should we close issues, or should this be left to the reviewers?
