+++ docs/reference/api-overview.md
@@ -6,6 +6,10 @@ outline: deep
 
 EMGFlow is broken into 5 models - FileIO, preprocessing signals, feature extraction, detecting outliers, and generating visualizations.
 
+## File Format
+
+The EMGFlow Python package works with CSV files, but is planned to expand to other file formats in the future. To prepare your data to be compatible with EMGFlow, it needs to be a CSV file with ideally a "Time" column, and additional columns for the signals you have recorded. "Time" should contain the time from 0 the signal has been recorded for, and the additional columns will have the recording of the signals at that time. Additionally, the file should have a constant sampling rate (time difference between each sequential row).
+
 ## Modules
 
 ### `FileAccess`
@@ -34,27 +38,7 @@ This module provides preprocessing functions for cleaning sEMG signals prior to 
 
 #### `NotchFilterSignals()`
 
-##### Description
-
 Notch filtering involves filtering specific frequencies. This is typically due to some sort of interference, such as the power source of the device taking the reading.
-
-##### Parameters
-
-Notch filtering is controlled by the function `NotchFilterSignals()`, with the following main parameters:
-- `in_path` <-- File path to input folder
-- `out_path` <-- File path to output folder
-- `sampling_rate` <-- Sampling rate of the data
-- `notch` <-- Notch filters to be applied
-
-`in_path` refers to the folder containing the signal data. The function will be able to detect signals in the folder, and in any subfolders.
-
-`out_path` refers to the folder where the filtered signals will be output to. The function will automatically recreate the file structure of the input folder with additional subfolders.
-
-`sampling_rate` refers to the sampling rate of the data. The function will assume that each signal, and each column, is using the same sampling rate.
-
-`notch` refers to the notch filters being applied to the signal. This takes the form of a list of `(Hz,Q)` tuples. `Hz` is the frequency to which the filter is being applied, and `Q` is the Q-factor (strength of the filter, where a lower number is a stronger filter). The Q-factor needs to be explored to find a suitable value.  When the Q-factor is too high, there will be barely any effect.  When the Q_factor is too, important frequencies around the target being filtered out. All filters in the list will be applied to each column and signal.
-
-##### Value-Added
 
 `NotchFilterSignals()` provides flexibility for use in different regions of the world. Some filtering packages only provide notch filtering for 60Hz, the frequency where power can interfere with signal readings.  However, other regions use 50Hz frequencies.
 
@@ -62,172 +46,45 @@ For more information about further customizations and detail about `NotchFilterS
 
 #### `BandpassFilterSignals()`
 
-**Description**
-
 Bandpass filtering involves specifying a range of frequencies to keep, and removing all other frequencies outside this range. This is useful to remove interference outside the most meaningful range of readings. 
-
-**Parameters**
-
-Bandpass filtering is controlled by the function `BandpassFilterSignals()`, with the following parameters:
-- `in_path` <-- File path to input folder
-- `out_path` <-- File path to output folder
-- `sampling_rate` <-- Sampling rate of the data
-- `low` <-- Lower frequency range
-- `high` <-- Upper frequency range
-
-`in_path` refers to the folder containing the signal data. The function will be able to detect signals in the folder, and in any subfolders.
-
-`out_path` refers to the folder where the filtered signals will be output to. The function will automatically recreate the file structure of the input folder with additional subfolders.
-
-`sampling_rate` refers to the sampling rate of the data. The function will assume that each signal, and each column, is using the same sampling rate.
-
-`low` and `high` are the thresholds for the bandpass filter. All frequencies outside the range provided are filtered out of the signal. The defaults respectively are 20 and 450.
-
-**Value-Added**
 
[…]
