+++ docs/guide/examples.md
@@ -6,17 +6,19 @@ outline: deep
 
 This page demonstrates how to use EMGFlow with simple and more advanced examples.
 
-## Simple example
+## Simple examples
+
+The simple examples can run by copying and pasting the code into a Python environment. The only requirement is for the EMGFlow package to be installed
 
 ### Cleaning, visualization, and feature extraction
 
-A simple example outlining the four main steps of the EMG processing pipeline.
+A simple example outlining EMG preprocessing.
 
 ```python
 import EMGFlow
 
-# Load in-built data
-path_names = EMGFlow.make_sample_data()
+# Load built-in data
+sample_data = EMGFlow.sample_data
 
 # Sampling rate
 sampling_rate = 2000
@@ -30,21 +32,24 @@ smooth_window = 50
 # Columns containing data for preprocessing
 cols = ['EMG_zyg', 'EMG_cor']
 
-# Visualization settings
-labels = ['Smooth', 'Bandpass', 'Notch']
-units = 'mV'
+# Preprocess first column of signals ('EMG_zyg')
+sample_data = EMGFlow.ApplyNotchFilters(sample_data, cols[0], sampling_rate, notch_vals)
+sample_data = EMGFlow.ApplyBandpassFilter(sample_data, cols[0], sampling_rate, band_low, band_high)
+sample_data = EMGFlow.ApplyRMSSmooth(sample_data, cols[0], smooth_window)
 
-## 1. Preprocess data
-EMGFlow.ApplyNotchFilters(path_names, sampling_rate, notch_vals, cols)
-EMGFlow.BandpassFilterSignals(path_names, sampling_rate, band_low, band_high, cols)
-EMGFlow.SmoothFilterSignals(path_names, sampling_rate, smooth_window, cols)
-
-## 2. Visualise signals
-EMGFlow.GenPlotDash(path_names, sampling_rate, cols[0], units, labels)
-
-## 3. Extract features and save results in "Features.csv"
-df = EMGFlow.AnalyzeSignals(in_paths, sampling_rate, cols)
+# Preprocess second column of signals ('EMG_cor')
+sample_data = EMGFlow.ApplyNotchFilters(sample_data, cols[1], sampling_rate, notch_vals)
+sample_data = EMGFlow.ApplyBandpassFilter(sample_data, cols[1], sampling_rate, band_low, band_high)
+sample_data = EMGFlow.ApplyRMSSmooth(sample_data, cols[1], smooth_window)
 ```
+
+## Advanced examples
+
+The advanced examples can ran by copying and pasting the code into a Python environment.
+
+In addition, they require having a local "data" folder, with "raw", "notch", "notch_s", "bandpass", "smoothed" and "feature" subfolders. The "raw" subfolder is required to have one or more emotional CSV files inside.
+
+The `cols` parameter may need to be modified to contain some of the column names of the emotional CSV in the "raw" subfolder.
 
 ### Complete pipeline with project data
 
@@ -70,18 +75,16 @@ band_high = 140
 smooth_window = 50
 
 # Columns containing data for preprocessing
-cols = ['col1', 'col3']
+cols = ['col1', 'col2']
 
 # Preprocess signals
 EMGFlow.NotchFilterSignals(raw_path, notch_path, sampling_rate, notch_vals, cols)
 EMGFlow.BandpassFilterSignals(notch_path, band_path, sampling_rate, band_low, band_high, cols)
-EMGFlow.SmoothFilterSignals(band_path, smooth_path, sampling_rate, smooth_window, cols)
+EMGFlow.SmoothFilterSignals(band_path, smooth_path, smooth_window, cols)
 
[…]
