### 2025-04-05 4973bb5
Added make_sample_data function
Added a function to set up test files for using sample data in a workflow

Updated toml to include the sample data

Added an __init__.py file to declare the data folder as part of the package
