### 2025-04-05 9332dfd
Update examples.md
Updated examples to work with new make_sample_data() function
