+++ EMGFlow-Package/src/EMGFlow/PreprocessSignals.py
@@ -7,7 +7,7 @@ import re
 from tqdm import tqdm
 import warnings
 
-#from .FileAccess import *
+from .FileAccess import *
 
 #
 # =============================================================================
