### 2025-04-06 b0440f5
Added CleanSignals() function
Added a CleanSignals function to go over the entire preprocessing pipeline

Added a make_path_dict() function to help in creating a path dictionary

Fixed a bug in ExtractFeatures where the feature file would not generate in the Features folder
