### 2025-04-06 92ddd3b
Update examples.md
Added an additional example using CleanSignals

Changed language to say "signal data" instead of "emotional data"
