### 2025-04-10 41865d8
Updated functions to use path_names
Updated the ExtractFeatures, CleanSignals and GenPlotDash functions to use the path_names dict

make_sample_data() now creates the step folders inside of a Data folder
