### 2025-04-11 a701e61
Update examples.md
Updated examples to use path_names dict

Added an additional example for plotting
