### 2025-04-11 39a0d99
Update PlotSignals.py
"Feature" paths are now automatically removed from the path_names dictionary
