+++ EMGFlow-Package/src/EMGFlow/PlotSignals.py
@@ -318,6 +318,8 @@ def GenPlotDash(path_names, col, units, expression=None, file_ext='csv', autorun
 
     """
     
+    # Remove feature path, and convert dictionary to lists
+    path_names.pop("Feature", None)
     in_paths = list(path_names.values())
     names = list(path_names.keys())
     
