### 2025-04-11 75d4f1b
Updated documentation for ExtractFeatures
Updated documentation for ExtractFeatures to reflect changes using path_names dict
