### 2025-04-11 73213a1
Update ExtractFeatures.py
Added an exception for if the required paths for ExtractFeatures to function are not provided
