### 2025-04-11 2fe1423
Update plot-signals.md
Updated documentation for GenPlotDash to use the path_names dict
