### 2025-04-12 cbc0134
Update paper.md
Updated the number of features reported

Updated the example provided

Deleted unneeded line breaks
