+++ papers/JOSS/paper.md
@@ -27,7 +27,6 @@ bibliography: paper.bib
 
 # Statement of Need
 
-
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Although several packages exist for processing physiological and neurological signals, support for sEMG has remained limited. Many packages lack a comprehensive set of features that can be extracted from sEMG data, leaving researchers to use a patchwork of tools. Other packages are orientated around event detection in individual recordings and use a GUI-based workflow that requires more manual intervention. While this design works well for processing unedited continuous recordings of a single participant, it complicates the extraction of features from large datasets common to machine learning [@abadi_decaf_2015; @chen_emotion_2022; @koelstra_deap_2012; @schmidt_introducing_2018; @sharma_dataset_2019; @zhang_biovid_2016].
 
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;_EMGFlow_, a portmanteau of EMG and Workflow, fills this gap by providing a flexible pipeline for extracting a wide range of sEMG features, with a scalable design suited for large datasets.
@@ -61,8 +60,6 @@ notch_vals = [(50,5)]
 # Apply notch filter to raw sEMG files
 EMGFlow.NotchFilterSignals(raw_path, notch_path, sr, notch_vals, cols)
 ```
-
-
 
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Additional arguments allow users to customize which files are selected and how they are processed. Filtering functions accept an optional regex argument, allowing users to apply filters to specific files. Most functions use common sense defaults, which can be modified task-wide or for select cases. For example, in North America, mains electricity is nominally supplied at 120 VAC 60 Hz, while other countries may supply power at 200-240 VAC 50Hz. This variation in frequency requires different notch filter settings depending on where the data were recorded. _EMGFlow_ accommodates this need by allowing the user to specify the frequency and quality factor of the applied filter. Extending our first example, we now apply an additional notch filter to a subset of files exhibiting noise at 150 Hz, the 3rd harmonic of the mains source.
 
@@ -120,34 +117,36 @@ EMGFlow.GenPlotDash(in_paths, sampling_rate, show_col, units, labels)
 
 ## Feature Extraction Routines
 
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Following data preprocessing, the signal files are ready for feature extraction. _EMGFlow_ extracts 32 features that capture information in both time and frequency domains. The set of 18 time-domain features capture standard statistical moments, including mean, variance, skew, and kurtosis, along with sEMG-specific measures. These include features such as Willison amplitude, an indicator of motor unit firing calculated as the number of times the sEMG amplitude exceeds a threshold, and log-detector, an estimate of the exerted muscle force [@tkach_study_2010]. 
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Following data preprocessing, the signal files are ready for feature extraction. _EMGFlow_ extracts 32 features that capture information in both time and frequency domains. The set of 17 time-domain features capture standard statistical moments, including mean, variance, skew, and kurtosis, along with sEMG-specific measures. These include features such as Willison amplitude, an indicator of motor unit firing calculated as the number of times the sEMG amplitude exceeds a threshold, and log-detector, an estimate of the exerted muscle force [@tkach_study_2010]. 
 
[…]
