### 2025-04-15 5842ec2
Updated examples
Updated titles of examples

Removed first example (too similar to getting started page)
