### 2025-04-16 09cd590
Update PreprocessSignals.py
Changed make_path_dict to make_paths, only takes a root parameter

Changed make_sample_data to also return a path dictionary
