### 2025-04-16 99955c5
Updated examples
Updated all examples to use the new format for make_sample_data

Updated the advanced examples to work without any outside file creation

Changed the "Getting Started" example to use "ExtractFeatures" rather than "AnalyzeSignals"
