+++ docs/guide/examples.md
@@ -4,11 +4,9 @@ outline: deep
 
 # Examples
 
-This page demonstrates how to use EMGFlow with simple and more advanced examples.
+This page demonstrates how to use EMGFlow with simple and more advanced examples. The examples can be ran by copying and pasting the code into a Python environment, ensuring the appropriate packages have been installed.
 
 ## Simple Examples
-
-The simple examples can run by copying and pasting the code into a Python environment. The only requirement is for the EMGFlow package to be installed
 
 ### Using Custom Filter Parameters
 
@@ -17,11 +15,8 @@ A simple example outlining EMG preprocessing and feature extraction using manual
 ```python
 import EMGFlow
 
-# Load sample data
-EMGFlow.make_sample_data()
-
-# Paths for data files
-path_names = EMGFlow.make_path_dict()
+# Load sample data and get path dictionary
+path_names = EMGFlow.make_sample_data()
 
 # Sampling rate
 sampling_rate = 2000
@@ -48,15 +43,20 @@ EMGFlow.SmoothFilterSignals(path_names['Bandpass'], path_names['Smooth'], smooth
 df = EMGFlow.ExtractFeatures(path_names, sampling_rate, cols)
 ```
 
-### Working with individual files
+### Working With Individual Files
 
 A simple example outlining EMG preprocessing using a loaded dataframe.
 
 ```python
 import EMGFlow
+import os
+import pandas as pd
 
-# Load built-in dataframe
-sample_data = EMGFlow.sample_data
+# Load sample data and get path dictionary
+path_names = EMGFlow.make_sample_data()
+
+# Load dataframe from generated data
+sample_data = pd.read_csv(os.path.join(path_names['Raw'], 'sample_data_01.csv'))
 
 # Sampling rate
 sampling_rate = 2000
@@ -81,61 +81,17 @@ sample_data = EMGFlow.ApplyBandpassFilter(sample_data, cols[1], sampling_rate, b
 sample_data = EMGFlow.ApplyRMSSmooth(sample_data, cols[1], smooth_window)
 ```
 
-### Preprocessing and Plotting Pipeline
+## Advanced Examples
 
-A simple example outlining EMG preprocessing and plotting.
-
-```python
-import EMGFlow
-
-# Load sample data
-EMGFlow.make_sample_data()
-
-# Set sampling rate
-sampling_rate = 2000
-
-# Load path dictionary
-path_names = EMGFlow.make_path_dict()
-
-# Clean data
-EMGFlow.CleanSignals(path_names)
-
-# Plot data on the "EMG_zyg" column
-EMGFlow.GenPlotDash(path_names, 'EMG_zyg', 'mV')
-```
-
-## Advanced examples
-
-The advanced examples can ran by copying and pasting the code into a Python environment.
-
-In addition, they require having a local "data" folder, with "raw", "notch", "notch_s", "bandpass", "smoothed" and "feature" subfolders. The "raw" subfolder is required to have one or more signal CSV files inside.
-
-The `cols` parameter may need to be modified to contain some of the column names of the signal CSV in the "raw" subfolder.
-
-They serve more as a template of how analysis can be done, to be replicated with your own data.
-
-### Custom filters for individiual files
+### Custom Filters for Individiual Files
 
 A more advanced example applying conditional filters based on file names.
 
 ```python
 import EMGFlow
 
-# Paths for data files
-raw_path = 'data/raw'
-notch_path = 'data/notch'
[…]
