### 2025-04-17 #7696 @faroit
> @faroit For our knowledge, should we close issues, or should this be left to the reviewers?

@wbaccinelli For most issues it is clear to see when they were are addressed (ideally when links to PR are provided) so closing from either side is ok. For some others, feedback might be required (eg. https://github.com/WiIIson/EMGFlow-Python-Package/issues/8), so its best if the reviewer can confirm that this was addressed and then close. Hope that helps
