### 2025-04-22 ab74008
Added Mermaid
Added mermaid for creating mindmaps and flowcharts

Added a flowchart of EMGFlow module layout
