### 2025-04-22 9caff78
Updated mermaid config
