### 2025-04-22 e846973
Updated mermaid charts
Updated mindmap subsection titles

Added flowchart for processing workflow
