### 2025-04-23 2daaf96
Added MMAV2
Added documentation for MMAV2

Changed "MMAV" to "MMAV1"

Added citation for MMAV2
