+++ docs/reference/feature-extraction.md
@@ -155,25 +155,25 @@ MAV = EMGFlow.CalcMAV(SignalDF, 'column1', 2000)
 
 
 
-## `CalcMMAV`
+## `CalcMMAV1`
 
 **Description**
 
-Calculates the Modified Mean Absolute Value (MMAV) of the signal. The MMAV is an alteration of MAV that gives more weight to values in the middle of the signal to reduce error from the beginning and end of the signal.
+Calculates the Modified Mean Absolute Value 1 (MMAV1) of the signal. The MMAV1 is an alteration of MAV that gives more weight to values in the middle of the signal to reduce error from the beginning and end of the signal.
 ```python
-CalcMMAV(Signal, col)
+CalcMMAV1(Signal, col)
 ```
 
 **Theory**
 
-The MMAV is identical to MAV, except it introduces a weight to the calculation. Values are given a weight of 1 when they are between the 25th and 74th percentile, and 0.5 outside. The article describing the MMAV also listed another modification, MMAV2, but the description appeared to be flawed, and as such the measure was not included.
+The MMAV1 is identical to MAV, except it introduces a weight to the calculation. Values are given a weight of 1 when they are between the 25th and 75th percentile, and 0.5 outside.
 
-The MMAV is calculated as follows:
+The MMAV1 is calculated as follows:
 $$
-\text{MMAV}=\frac{1}{N}\sum_{i=1}^N|x_iw_i|
+\text{MMAV1}=\frac{1}{N}\sum_{i=1}^N|x_iw_i|
 $$
 $$
-w_i=\left\{ \begin{matrix} 1 & \text{if }0.25N\le n\le 0.75N \\ 0.5 & \text{otherwise} \end{matrix} \right\}
+w_i=\left\{ \begin{matrix} 1 & \text{if }0.25N\le i\le 0.75N \\ 0.5 & \text{otherwise} \end{matrix} \right\}
 $$
 - $N$ <-- Number of data points
 
@@ -189,8 +189,8 @@ $$
 
 **Returns**
 
-`CalcMMAV`: float
-- Returns the value of the MMAV.
+`CalcMMAV1`: float
+- Returns the value of the MMAV1.
 
 **Error**
 
@@ -199,9 +199,61 @@ Raises an error if `col` is not found in `Signal`.
 **Example**
 
 ```python
-# Calculate the MMAV of SignalDF, for column 'column1'
-MMAV = EMGFlow.CalcMMAV(SignalDF, 'column1', 2000)
+# Calculate the MMAV1 of SignalDF, for column 'column1'
+MMAV1 = EMGFlow.CalcMMAV1(SignalDF, 'column1', 2000)
 ```
+
+
+
+## `CalcMMAV2`
+
+**Description**
+
+Calculates the Modified Mean Absolute Value 2 (MMAV2) of the signal. The MMAV2 is an alteration of MAV that gives more weight to values in the middle of the signal to reduce error from the beginning and end of the signal.
+```python
+CalcMMAV2(Signal, col)
+```
+
+**Theory**
+
+The MMAV2 is identical to MAV, except it introduces a weight to the calculation. Values are given a weight of 1 when they are between the 25th and 75th percentile, a weight of $\frac{4i}{N}$ when below the 25th percentile, and a weight of $\frac{4(i-N)}{N}$ when above the 75th percentile. This makes the weight smoothly increase and decrease approaching the center of the signal, as opposed to MMAV1 which instantly jumps from a weight of 0.5 to 1.
+
[…]
