### 2025-04-23 e36bee8
Update examples.md
Updated examples to work with new version of make_sample_data and make_paths
