### 2025-04-23 884e20a
Added module structure mindmaps
Added "module structure" mindmaps to each page.

The overview page has a high-level description of each page, and the module pages and an expanded view of all functions in that module.
