+++ docs/reference/outlier-detection.md
@@ -2,7 +2,18 @@
 
 This module provides methods to help detect signal files that contain outliers. This helps for workflows involving batch processing of files, where it might be harder to determine if there are any patterns, or specific files that need additional filters applied.
 
+## Module Structure
 
+```mermaid
+mindmap
+    root((EMGFlow))
+        EF(Extract Features)
+        FA(File Access)
+        DO(Detect Outliers)
+            DetectOutliers
+        PlS(Plot Signals)
+        PrS(Preprocess Signals)
+```
 
 ## `DetectOutliers`
 
