+++ docs/guide/getting-started.md
@@ -55,59 +55,3 @@ C --> D[Smooth Filter]
 D --> E[Extract Features]
 D --> F[Plot Signal]
 ```
-
-## EMGFlow Module Structure
-
-```mermaid
-mindmap
-    root((EMGFlow))
-        EF(Extract Features)
-            CalcIEMG
-            CalcMAV
-            CalcMMAV
-            CalcSSI
-            CalcVAR
-            CalcVOrder
-            CalcRMS
-            CalcWL
-            CalcWAMP
-            CalcLOG
-            CalcMFL
-            CalcAP
-            CalcSpecFlux
-            CalcMDF
-            CalcMNF
-            CalcTwitchRatio
-            CalcTwitchIndex
-            CalcTwitchSlope
-            CalcSC
-            CalcSF
-            CalcSS
-            CalcSDec
-            CalcSEntropy
-            CalcSRoll
-            CalcSBW
-            ExtractFeatures
-        FA(File Access)
-            ReadFileType
-            MapFiles
-            ConvertMapFiles
-        DO(Detect Outliers)
-            DetectOutliers
-        PS(Plot Signals)
-            GenPlotDash
-            PlotSpectrum
-            PlotCompareSignals
-        PrS(Preprocess Signals)
-            EMG2PSD
-            ApplyNotchFilters
-            NotchFilterSignals
-            ApplyBandpassFilter
-            BandpassFilterSignals
-            ApplyFWR
-            ApplyBoxcarSmooth
-            ApplyRMSSmooth
-            ApplyGaussianSmooth
-            ApplyLoessSmooth
-            SmoothFilterSignals
-```
