### 2025-04-23 87f870e
Added mindmap links
Added workaround code for mindmap links
