### 2025-04-23 2c5ea58
Updated mindmap links
Updated method of using mindmap links
