### 2025-04-24 a302da3
Update paper.md
Updated examples with new format
