### 2025-04-24 f60c6c7
Updated examples
Updated "getting started" example with new format

Also removed index.js code for dynamically adding links to mindmaps
