+++ docs/guide/getting-started.md
@@ -15,8 +15,11 @@ _EMGFlow_ extracts a comprehensive set of 32 statistical features from sEMG sign
 ```python
 import EMGFlow
 
-# Load in-built data
-path_names = EMGFlow.make_sample_data()
+# Get path dictionary
+path_names = EMGFlow.make_paths()
+
+# Load sample data
+EMGFlow.make_sample_data(path_names)
 
 # Preprocess signals
 EMGFlow.CleanSignals(path_names, sampling_rate = 2000)
