### 2025-04-24 1463e43
Updated JOSS paper
Added an additional figure for an overview of the processing pipeline

Rewrote the processing pipeline section to work with the new data used.
