### 2025-04-28 dac3cb4
Updated make_paths and make_sample_data
Added root parameter in make_paths to default to None, does not create a "data" folder if you specify the path

Added exceptions to make_sample_data if correct keys are not present

Moved both functions to the FileAccess module
