+++ EMGFlow-Package/src/EMGFlow/PreprocessSignals.py
@@ -1,4 +1,3 @@
-import importlib.resources as resources
 import scipy
 import pandas as pd
 import numpy as np
@@ -867,42 +866,6 @@ def SmoothFilterSignals(in_path, out_path, window_size, cols=None, expression=No
 # =============================================================================
 #
 
-def make_paths(root=os.getcwd()):
-    """
-    Generates a file structure for signal files, and returns a dictionary of
-    the locations for these files.
-    
-    A "Data" folder is created, with "Raw", "Notch", "Bandpass", "Smooth" and
-    "Feature" subfolders
-
-    Parameters
-    ----------
-    root : str, optional
-        Root of the data to be generated. The default is os.cwd().
-
-    Returns
-    -------
-    path_dict : [str] dict
-        A dictionary of file locations with keys for stage in the processing
-        pipeline.
-
-    """
-    # Create dictionary
-    path_dict = {
-        'Raw':os.path.join(root, 'Data', 'Raw'),
-        'Notch':os.path.join(root, 'Data', 'Notch'),
-        'Bandpass':os.path.join(root, 'Data', 'Bandpass'),
-        'Smooth':os.path.join(root, 'Data', 'Smooth'),
-        'Feature':os.path.join(root, 'Data', 'Feature')
-    }
-    
-    # Create folders
-    for value in path_dict.values():
-        os.makedirs(value, exist_ok=True)
-    
-    # Return dictionary
-    return path_dict
-
 def CleanSignals(path_names, sampling_rate=2000):
     """
     Automates the EMG preprocessing workflow, performing notch filtering,
@@ -916,11 +879,27 @@ def CleanSignals(path_names, sampling_rate=2000):
     sampling_rate : float
         Sampling rate of the Signal.
 
+    Raises
+    ------
+    Exception
+        An exception is raised if the provided 'path_names' dictionary doesn't
+        contain a 'Raw', 'Notch', 'Bandpass' or 'Smooth' path key.
+
     Returns
     -------
     None.
 
     """
+    
+    # Raise exceptions if paths not found
+    if 'Raw' not in path_names:
+        raise Exception('Raw path not detected in provided dictionary (path_names)')
+    if 'Notch' not in path_names:
+        raise Exception('Notch path not detected in provided dictionary (path_names)')
+    if 'Bandpass' not in path_names:
+        raise Exception('Bandpass path not detected in provided dictionary (path_names)')
+    if 'Smooth' not in path_names:
+        raise Exception('Smooth path not detected in provided dictionary (path_names)')
     
     # Default values for notch filtering and window size
     notch = [(50,5)]
@@ -931,45 +910,3 @@ def CleanSignals(path_names, sampling_rate=2000):
     BandpassFilterSignals(path_names['Notch'], path_names['Bandpass'], sampling_rate)
     SmoothFilterSignals(path_names['Bandpass'], path_names['Smooth'], window_size)
     return
-
-#
-# =============================================================================
-#
-
-def make_sample_data(path_names):
-    """
-    Generates sample data in the 'Raw' folder of a provided dictionary of file
-    locations.
-
-    Parameters
-    ----------
-    path_names : [str] dict
-        Dictionary of file locations.
-
-    Raises
-    ------
-    Exception
-        An exception is raised if the provided 'path_names' dictionary doesn't
-        contain a 'Raw' path key.
-
-    Returns
-    -------
-    None.
-
-    """
-    
-    # Check that a 'raw' folder exists
-    if 'Raw' not in path_names:
-        raise Exception('Raw path not detected in provided dictionary (path_names)')
-    
-    # Load the sample data
[…]
