### 2025-04-28 7f75157
Updated documentation and language
Updated outdated documentation for ExtractFeatures

Added documentation for make_paths and make_sample_data in FileAccess

Changed all instances of "pre-processing" to "preprocessing" for consistency
