### 2025-04-29 WiIIson/EMGFlow-Python-Package#1 @faroit
@samiralavi @WiIIson can this be closed?

ref: https://github.com/openjournals/joss-reviews/issues/7696
