### 2025-04-29 WiIIson/EMGFlow-Python-Package#2 @faroit
ref https://github.com/openjournals/joss-reviews/issues/7696
