### 2025-04-29 1af88f2
Update FileAccess.py
Modified make_sample_data and make_paths to fit with naming format of other functions
