### 2025-04-29 f177c47
Updated make_paths and make_sample_data names
Changed to makePaths and makeSampleData
