### 2025-04-29 841b93a
Update getting-started.md
Updated naming format for makeSampleData and makePaths
