### 2025-04-29 cabb017
Updated FileAccess description and preprocess-signals documentation
Added additional text to FileAccess describing the makePaths and makeSampleData functions.

Changed examples in PreprocessSignals to use new makePaths function.
