### 2025-04-30 ea58f23
Update FileAccess.py
Capitalized MakeSampleData and MakePaths to be consistent with other function names
