### 2025-04-30 dfd2398
Updated documentation formatting
Updated makePaths and makeSampleData to MakePaths and MakeSampleData (more consistent with other function names)

Updated some examples to use the MakePaths dict.
