+++ docs/guide/getting-started.md
@@ -16,10 +16,10 @@ _EMGFlow_ extracts a comprehensive set of 32 statistical features from sEMG sign
 import EMGFlow
 
 # Get path dictionary
-path_names = EMGFlow.makePaths()
+path_names = EMGFlow.MakePaths()
 
 # Load sample data
-EMGFlow.makeSampleData(path_names)
+EMGFlow.MakeSampleData(path_names)
 
 # Preprocess signals
 EMGFlow.CleanSignals(path_names, sampling_rate = 2000)
