### 2025-04-30 a1c018c
Writing update
Tweaked a few sentences to flow better

Changed some instances of improper function formatting

Changed author name to "D. William Conley"
