+++ docs/guide/getting-started.md
@@ -24,11 +24,11 @@ EMGFlow.MakeSampleData(path_names)
 # Preprocess signals
 EMGFlow.CleanSignals(path_names, sampling_rate = 2000)
 
-# Plot data on the "EMG_zyg" column
-EMGFlow.GenPlotDash(path_names, 'EMG_zyg', 'mV')
-
 # Extract features and save results in "Features.csv" in feature_path
 df = EMGFlow.ExtractFeatures(path_names, sampling_rate = 2000)
+
+# Plot data on the "EMG_zyg" column
+EMGFlow.GenPlotDash(path_names, 'EMG_zyg', 'mV')
 ```
 
 ## Input data format
