+++ docs/guide/what-is-emgflow.md
@@ -26,6 +26,6 @@ _EMGFlow_, a portmanteau of EMG and Workflow, addresses this limitation by offer
 
 ## Contributing
 
-We welcome contributions to the project. These can be initiated through [GitHub](https://github.com/WiIIson/EMGFlow-Python-Package) on the project's issue tracker or through a pull request. Suggestions for feature enhancements, tips, as well as general questions and concerns can also be expressed through direct interaction with contributors and developers. You can also contact us over email.
+We welcome contributions to the project. These can be initiated through [GitHub](https://github.com/WiIIson/EMGFlow-Python-Package) on the project's [issue tracker](https://github.com/WiIIson/EMGFlow-Python-Package/issues) or through a pull request. Suggestions for feature enhancements, tips, as well as general questions and concerns can also be expressed through direct interaction with contributors and developers. You can also contact us over email.
 
 If you experience any challenges with this module such as bugs, test support, or feature requests, please feel free to use this issue tracker.
