+++ docs/reference/api-overview.md
@@ -8,7 +8,7 @@ EMGFlow is broken into 5 modules - FileIO, preprocessing signals, feature extrac
 
 ## File Format
 
-The EMGFlow Python package works with CSV files, but is planned to expand to other file formats in the future. To prepare your data to be compatible with EMGFlow, it needs to be a CSV file with ideally a "Time" column, and additional columns for the signals you have recorded. "Time" should contain the time from 0 the signal has been recorded for, and the additional columns will have the recording of the signals at that time. Additionally, the file should have a constant sampling rate (time difference between each sequential row).
+The EMGFlow Python package works with CSV files, but is planned to expand to other file formats in the future. To prepare your data to be compatible with EMGFlow, it needs to be a CSV file with, ideally, a "Time" column, and additional columns for the signals you have recorded. "Time" should contain the time from 0 the signal has been recorded for, and the additional columns will have the recording of the signals at that time. Additionally, the file should have a constant sampling rate (time difference between each sequential row).
 
 ## Modules
 
@@ -30,7 +30,7 @@ These functions provide helper methods for accessing files.
 
 `MakePaths` is the beginning of the EMG processing pipeline, it generates folders for storing EMG data, and returns a dictionary with the locations of these folders. These locations are accessed under the keys "Raw", "Notch", "Bandpass", "Smooth" and "Feature". This dictionary is passed to many of the processing functions.
 
-`MakeSampleData` is a supplementary function that writes sample data for testing to the "Raw" folder of your `makePaths` dictionary
+`MakeSampleData` is a supplementary function that writes sample data for testing to the "Raw" folder of your `MakePaths` dictionary
 
 `MapFiles` is widely used internally in the EMGFlow, it takes a path to a folder, and generates a dictionary of paths to files contained within. This makes it easier to create a loop over subfiles, reading them in and performing analysis.
 
@@ -46,7 +46,7 @@ The "manual" mode allows for additional customization of processing. In these fu
 - `ApplyBandpassFilter`
 - `ApplyRMSSmooth`
 
-For more information about file accessing functions, see [FileAccess API](./file-access.md).
+For more information about file accessing functions, see [File access](./file-access.md).
 
 ### `PreprocessSignals`
 
@@ -58,7 +58,7 @@ Notch filtering involves filtering specific frequencies. This is typically due t
 
 `NotchFilterSignals()` provides flexibility for use in different regions of the world. Some filtering packages only provide notch filtering for 60Hz, the frequency where power can interfere with signal readings.  However, other regions use 50Hz frequencies.
 
-For more information about further customizations and detail about `NotchFilterSignals()`, see [PreprocessSignals API](./preprocess-signals.md).
+For more information about further customizations and detail about `NotchFilterSignals()`, see [Preprocessing](./preprocess-signals.md).
 
 #### `BandpassFilterSignals()`
 
@@ -66,7 +66,7 @@ Bandpass filtering involves specifying a range of frequencies to keep, and remov
 
 `BandpassFilterSignals()` uses bandpass thresholds of 20Hz and 450Hz, as this is default for EMG signals (De Luca et al., 2010). However, there is some disagreement within literature for different muscels, so `BandpassFilterSignals()` provides the option to change the thresholds.
 
-For more information about further customizations and detail about `BandpassFilterSignals()`, see [PreprocessSignals API](./preprocess-signals.md).
+For more information about further customizations and detail about `BandpassFilterSignals()`, see [Preprocessing](./preprocess-signals.md).
 
 #### `SmoothFilterSignals()`
 
@@ -74,7 +74,7 @@ Smoothing involves limiting the impacts of noise and outliers in the signal. By 
 
[…]
