+++ EMGFlow-Package/src/EMGFlow/extract_features.py
@@ -6,8 +6,8 @@ import re
 from tqdm import tqdm
 import warnings
 
-from .FileAccess import *
-from .PreprocessSignals import EMG2PSD
+from .access_files import *
+from .preprocess_signals import EMG2PSD
 
 #
 # =============================================================================
@@ -22,7 +22,7 @@ A collection of functions for extracting features.
 #
 
 
-def CalcIEMG(Signal, col, sr):
+def calc_iemg(Signal, col, sr):
     """
     Calculate the Integreated EMG (IEMG) of a Signal.
 
@@ -64,7 +64,7 @@ def CalcIEMG(Signal, col, sr):
 #
 
 # Calculate the Mean Absolute Value (MAV) of a signal
-def CalcMAV(Signal, col):
+def calc_mav(Signal, col):
     """
     Calculate the Mean Absolute Value (MAV) of a Signal.
 
@@ -99,7 +99,7 @@ def CalcMAV(Signal, col):
 # =============================================================================
 #
 
-def CalcMMAV1(Signal, col):
+def calc_mmav1(Signal, col):
     """
     Calculate the Modified Mean Absolute Value 1 (MMAV1) of a Signal.
 
@@ -141,7 +141,7 @@ def CalcMMAV1(Signal, col):
 # =============================================================================
 #
 
-def CalcMMAV2(Signal, col):
+def calc_mmav2(Signal, col):
     """
     Calculate the Modified Mean Absolute Value 2 (MMAV2) of a Signal.
 
@@ -185,7 +185,7 @@ def CalcMMAV2(Signal, col):
 # =============================================================================
 #
 
-def CalcSSI(Signal, col, sr):
+def calc_ssi(Signal, col, sr):
     """
     Calculate the Simple Square Integreal (SSI) of a Signal.
 
@@ -226,7 +226,7 @@ def CalcSSI(Signal, col, sr):
 # =============================================================================
 #
 
-def CalcVAR(Signal, col):
+def calc_var(Signal, col):
     """
     Calculate the Variance (VAR) of a Signal.
 
@@ -261,7 +261,7 @@ def CalcVAR(Signal, col):
 # =============================================================================
 #
 
-def CalcVOrder(Signal, col):
+def calc_vorder(Signal, col):
     """
     Calculate the V-Order of a Signal.
 
@@ -288,14 +288,14 @@ def CalcVOrder(Signal, col):
     if col not in list(Signal.columns.values):
         raise Exception("Column " + col + " not in Signal")
     
-    vOrder = np.sqrt(CalcVAR(Signal, col))
+    vOrder = np.sqrt(calc_var(Signal, col))
     return vOrder
 
 #
 # =============================================================================
 #
 
-def CalcRMS(Signal, col):
+def calc_rms(Signal, col):
     """
     Calculate the Root Mean Square (RMS) of a Signal.
 
@@ -330,7 +330,7 @@ def CalcRMS(Signal, col):
 # =============================================================================
 #
 
-def CalcWL(Signal, col):
+def calc_wl(Signal, col):
     """
     Calculate the Waveform Length (WL) of a Signal.
 
@@ -367,7 +367,7 @@ def CalcWL(Signal, col):
 # =============================================================================
 #
 
-def CalcWAMP(Signal, col, threshold):
+def calc_wamp(Signal, col, threshold):
     """
     Calculate the Willison Amplitude (WAMP) of a Signal.
 
@@ -406,7 +406,7 @@ def CalcWAMP(Signal, col, threshold):
 # =============================================================================
 #
 
-def CalcLOG(Signal, col):
+def calc_log(Signal, col):
     """
     Calculate the Log Detector (LOG) of a Signal.
 
@@ -442,7 +442,7 @@ def CalcLOG(Signal, col):
[…]

[…]
