### 2025-05-01 03fc509
Updated examples with new documentation
Updated examples with new names for functions and conventions
