+++ docs/guide/getting-started.md
@@ -16,19 +16,19 @@ _EMGFlow_ extracts a comprehensive set of 32 statistical features from sEMG sign
 import EMGFlow
 
 # Get path dictionary
-path_names = EMGFlow.MakePaths()
+pathNames = EMGFlow.make_paths()
 
 # Load sample data
-EMGFlow.MakeSampleData(path_names)
+EMGFlow.make_sample_data(pathNames)
 
 # Preprocess signals
-EMGFlow.CleanSignals(path_names, sampling_rate = 2000)
+EMGFlow.clean_signals(pathNames, samplingRate=2000)
 
 # Extract features and save results in "Features.csv" in feature_path
-df = EMGFlow.ExtractFeatures(path_names, sampling_rate = 2000)
+df = EMGFlow.extract_features(pathNames, samplingRate=2000)
 
 # Plot data on the "EMG_zyg" column
-EMGFlow.GenPlotDash(path_names, 'EMG_zyg', 'mV')
+EMGFlow.plot_dashboard(pathNames, 'EMG_zyg', 'mV')
 ```
 
 ## Input data format
