+++ docs/reference/feature-extraction.md
@@ -7,6 +7,8 @@ This module takes preprocessed data, and extracts features from the sEMG signal 
 ```mermaid
 mindmap
     root((EMGFlow))
+        FA(File Access)
+        DO(Detect Outliers)
         EF(Extract Features)
             CalcIEMG
             CalcMAV
@@ -35,8 +37,6 @@ mindmap
             CalcSRoll
             CalcSBW
             ExtractFeatures
-        FA(File Access)
-        DO(Detect Outliers)
         PlS(Plot Signals)
         PrS(Preprocess Signals)
 ```
