+++ docs/guide/examples.md
@@ -16,34 +16,34 @@ A simple example outlining EMG preprocessing and feature extraction using manual
 import EMGFlow
 
 # Get path dictionary
-path_names = EMGFlow.MakePaths()
+pathNames = EMGFlow.make_paths()
 
 # Load sample data
-EMGFlow.MakeSampleData(path_names)
+EMGFlow.make_sample_data(pathNames)
 
 # Sampling rate
-sampling_rate = 2000
+samplingRate = 2000
 
 # Filter parameters
-notch_vals = [(50, 5)]
-band_low = 20
-band_high = 140
-smooth_window = 50
+notchVals = [(50, 5)]
+bandLow = 20
+bandHigh = 140
+smoothWindow = 50
 
 # Columns containing data for preprocessing
 cols = ['EMG_zyg', 'EMG_cor']
 
 # 1. Apply notch filters
-EMGFlow.NotchFilterSignals(path_names['Raw'], path_names['Notch'], sampling_rate, notch_vals, cols)
+EMGFlow.notch_filter_signals(pathNames['Raw'], pathNames['Notch'], samplingRate, notchVals, cols)
 
 # 2. Apply bandpass filter
-EMGFlow.BandpassFilterSignals(path_names['Notch'], path_names['Bandpass'], sampling_rate, band_low, band_high, cols)
+EMGFlow.bandpass_filter_signals(pathNames['Notch'], pathNames['Bandpass'], samplingRate, bandLow, bandHigh, cols)
 
 # 3. Apply smoothing filter
-EMGFlow.SmoothFilterSignals(path_names['Bandpass'], path_names['Smooth'], smooth_window, cols)
+EMGFlow.smooth_filter_signals(pathNames['Bandpass'], pathNames['Smooth'], smoothWindow, cols)
 
 # 4. Extract features
-df = EMGFlow.ExtractFeatures(path_names, sampling_rate, cols)
+df = EMGFlow.extract_features(pathNames, samplingRate, cols)
 ```
 
 ### Working With Individual Files
@@ -56,35 +56,35 @@ import os
 import pandas as pd
 
 # Get path dictionary
-path_names = EMGFlow.MakePaths()
+pathNames = EMGFlow.make_paths()
 
 # Load sample data
-EMGFlow.MakeSampleData(path_names)
+EMGFlow.make_sample_data(pathNames)
 
 # Load dataframe from generated data
-sample_data = pd.read_csv(os.path.join(path_names['Raw'], 'sample_data_01.csv'))
+sampleData = pd.read_csv(os.path.join(pathNames['Raw'], 'sample_data_01.csv'))
 
 # Sampling rate
-sampling_rate = 2000
+samplingRate = 2000
 
 # Custom filter parameters
-notch_vals = [(50, 5)]
-band_low = 20
-band_high = 140
-smooth_window = 50
+notchVals = [(50, 5)]
+bandLow = 20
+bandHigh = 140
+smoothWindow = 50
 
 # Columns containing data for preprocessing
 cols = ['EMG_zyg', 'EMG_cor']
 
 # Preprocess first column of signals ('EMG_zyg')
-sample_data = EMGFlow.ApplyNotchFilters(sample_data, cols[0], sampling_rate, notch_vals)
-sample_data = EMGFlow.ApplyBandpassFilter(sample_data, cols[0], sampling_rate, band_low, band_high)
-sample_data = EMGFlow.ApplyRMSSmooth(sample_data, cols[0], smooth_window)
+sampleData = EMGFlow.apply_notch_filters(sampleData, cols[0], samplingRate, notchVals)
[…]
