### 2025-05-01 43fc7bf
Updated page names and flowchart
Changed names of pages and modules to be more consistent with documentation modules
