+++ docs/reference/plot-signals.md
@@ -7,7 +7,7 @@ The plotting module `PlotSignals` provides functions to help visualize individua
 ```mermaid
 mindmap
     root((EMGFlow))
-        FA(File Access)
+        AF(File Access)
         DO(Detect Outliers)
         EF(Extract Features)
         PlS(Plot Signals)
