### 2025-05-01 82b253d
Reordered mindmaps
Reordered and renamed mindmaps according to stage of processing
