+++ docs/reference/preprocess-signals.md
@@ -7,10 +7,6 @@ The `PreProcessSignals` module provides preprocessing functions for cleaning sEM
 ```mermaid
 mindmap
     root((EMGFlow))
-        AF(File Access)
-        DO(Detect Outliers)
-        EF(Extract Features)
-        PlS(Plot Signals)
         PrS(Preprocess Signals)
             EMG2PSD
             ApplyNotchFilters
@@ -23,6 +19,10 @@ mindmap
             ApplyGaussianSmooth
             ApplyLoessSmooth
             SmoothFilterSignals
+        EF(Extract Features)
+        PlS(Plot Signals)
+        DO(Detect Outliers)
+        AF(File Access)
 ```
 
 ## `EMG2PSD`
