+++ docs/.vitepress/config.ts
@@ -38,11 +38,11 @@ export default withMermaid({
         text: 'Reference',
         items: [
           { text: 'Overview', link: '/reference/api-overview' },
+          { text: 'Access Files', link: '/reference/access-files' },
+          { text: 'Detect Outliers', link: '/reference/detect-outliers' },
+          { text: 'Plot Signals', link: '/reference/plot-signals' },
           { text: 'Preprocess Signals', link: '/reference/preprocess-signals' },
           { text: 'Extract Features', link: '/reference/extract-features' },
-          { text: 'Plot Signals', link: '/reference/plot-signals' },
-          { text: 'Detect Outliers', link: '/reference/detect-outliers' },
-          { text: 'Access Files', link: '/reference/access-files' },
         ]
       },
 
