+++ docs/reference/api-overview.md
@@ -4,11 +4,11 @@ outline: deep
 
 # API Overview
 
-EMGFlow is broken into 5 modules - FileIO, preprocessing signals, feature extraction, detecting outliers, and generating visualizations.
+EMGFlow is broken into 5 modules: `access_files` for file IO, `detect_outliers` for detecting outliers, `preprocess_signals` for preprocessing signals, `plot_signals` generating visualizations, and `extract_features` feature extraction.
 
 ## File Format
 
-The EMGFlow Python package works with CSV files, but is planned to expand to other file formats in the future. To prepare your data to be compatible with EMGFlow, it needs to be a CSV file with, ideally, a "Time" column, and additional columns for the signals you have recorded. "Time" should contain the time from 0 the signal has been recorded for, and the additional columns will have the recording of the signals at that time. Additionally, the file should have a constant sampling rate (time difference between each sequential row).
+The EMGFlow Python package works with CSV files, but is planned to expand to other file formats in the future. To prepare your data to be compatible with EMGFlow, it needs to be a CSV file with, ideally, a "Time" column, and additional columns for the signals you have recorded. "Time" should contain the time since the beginning of the recording, and the additional columns will have the recording of the signals at that time. Additionally, the file should have a constant sampling rate (time difference between each sequential row).
 
 ## Modules
 
@@ -17,90 +17,90 @@ The EMGFlow Python package works with CSV files, but is planned to expand to oth
 ```mermaid
 mindmap
     root((EMGFlow))
-        PrS(Preprocess Signals)
-        EF(Extract Features)
-        PlS(Plot Signals)
-        DO(Detect Outliers)
         AF(Access Files)
+        DO(Detect Outliers)
+        PrS(Preprocess Signals)
+        PlS(Plot Signals)
+        EF(Extract Features)
 ```
 
-### `FileAccess`
+### `access_files` Module
 
-These functions provide helper methods for accessing files.
+These functions provide helper methods for accessing files, as well as functions for creating an EMG workflow filestructure.
 
-`MakePaths` is the beginning of the EMG processing pipeline, it generates folders for storing EMG data, and returns a dictionary with the locations of these folders. These locations are accessed under the keys "Raw", "Notch", "Bandpass", "Smooth" and "Feature". This dictionary is passed to many of the processing functions.
+`make_paths` is the beginning of the EMG processing pipeline. It generates folders for storing EMG data, and returns a dictionary with the locations of these folders. These locations are accessed under the keys "Raw", "Notch", "Bandpass", "Smooth" and "Feature" - each for a different stage of processing. This dictionary is passed to many of the processing functions. If you have your own file structure you want to use instead, you can create your own dictionary to use instead, manually setting the locations of the paths under these keys.
 
-`MakeSampleData` is a supplementary function that writes sample data for testing to the "Raw" folder of your `MakePaths` dictionary
+`make_sample_data` provides built-in data to test out a data processing pipeline. It writes EMG data files to the "Raw" directory of a provided dictionary.
 
-`MapFiles` is widely used internally in the EMGFlow, it takes a path to a folder, and generates a dictionary of paths to files contained within. This makes it easier to create a loop over subfiles, reading them in and performing analysis.
+`map_files` is widely used internally to the EMGFlow package. It takes a path to a folder, and generates a dictionary of paths to the files contained within based on their names. This makes it easier to loop over subfiles, reading and analyzing them.
 
-This forms the basis for the two modes of analysis offered by EMGFlow - automated, or manual.
+This function forms the basis for the two modes of analysis offered by EMGFlow: automated or manual.
 
[…]
