### 2025-05-02 c968a59
Update access-files.md
Updated function documentation to new format
