+++ docs/reference/access-files.md
@@ -1,4 +1,4 @@
-# `FileAccess` Module Documentation
+# `access_files` Module Documentation
 
 These functions provide helper methods for accessing files, and are mostly used internally by the package.
 
@@ -8,27 +8,27 @@ These functions provide helper methods for accessing files, and are mostly used 
 mindmap
     root((EMGFlow))
         AF(Access Files)
-            MakePaths
-            MakeSampleData
-            ReadFileType
-            MapFiles
-            ConvertMapFiles
+            make_paths
+            make_sample_data
+            read_file_type
+            map_files
+            convert_map_files
         DO(Detect Outliers)
         PrS(Preprocess Signals)
         PlS(Plot Signals)
         EF(Extract Features)
 ```
 
-## `MakePaths`
+## `make_paths`
 
 **Description**
 
-`MakePaths` generates a file structure for signal file processing, and returns a dictionary of the locations of these files.
+`make_paths` generates a file structure for signal files, and returns a dictionary of the locations of these files.
 
-A "Data" folder is created, with "Raw", "Notch", "Bandpass", "Smooth" and "Feature" subfolders. Alternatively, a path can be provided for these subfolders to be created in instead.
+"Raw", "Notch", "Bandpass", "Smooth" and "Feature" folders are created at a given location. By default, if no path is provided, these folders will be created inside a "Data" folder in the current working directory.
 
 ```python
-MakePaths(root=None)
+make_paths(root=None)
 ```
 
 **Parameters**
@@ -38,55 +38,55 @@ MakePaths(root=None)
 
 **Returns**
 
-`path_dict`: dict (str)
+`make_paths`: dict (str)
 - A dictionary of file locations with keys for the stage in the processing pipeline.
 
 **Example**
 
 ```python
 # Create folders and get locations
-path_names = EMGFlow.MakePaths()
+pathNames = EMGFlow.make_paths()
 ```
 
-## `MakeSampleData`
+## `make_sample_data`
 
 **Description**
 
-`MakeSampleData` generates sample data in the "Raw" folder of a provided dictionary of file locations.
+`make_sample_data` generates sample data in the "Raw" folder of a provided dictionary of file locations.
 
 ```python
-MakeSampleData(path_names)
+make_sample_data(pathNames)
 ```
 
 **Parameters**
 
-`path_names`: dict (str)
+`pathNames`: dict (str)
 - A dictionary of file locations
 
 **Returns**
 
-`None`
+`make_sample_data`: None
 
 **Error**
 
-Raises an error if `path_names` does not contain a `Raw` path key.
+Raises an error if `pathNames` does not contain a `Raw` path key.
 
 **Example**
 
 ```python
 # Create file paths, then create sample data
-path_names = EMGFlow.MakePaths()
-EMGFlow.MakeSampleData(path_names)
+pathNames = EMGFlow.make_paths()
+EMGFlow.make_sample_data(pathNames)
 ```
 
-## `ReadFileType`
+## `read_file_type`
 
 **Description**
 
-`ReadFileType` is a safe wrapper for reading files of a given extension.
+`read_file_type` is a safe wrapper for reading files of a given extension.
 
 ```python
-ReadFileType(path, file_ext)
+read_file_type(path, fileExt)
 ```
 
 **Parameters**
@@ -94,19 +94,19 @@ ReadFileType(path, file_ext)
 `path`: str
 - String filepath of file to read.
 
-`file_ext`: str
+`fileExt`: str
[…]
