### 2025-05-02 3be0d42
Updated documentation for make_sample_data
Updated documentation with new update for make_sample_data, where the function creates subfolders to put the data in
