### 2025-05-02 3a87c41
Updated map_files
Updated map_files to work with files with the same name stored in different folders
