### 2025-05-02 28253fd
Added an exception for plot_dashboard
Added an exception for if the user provides a list for "col" (an easy mistake)
