### 2025-05-02 a1e69bf
Update plot-signals.md
Added new exception for if col is a list for plot_dashboard
