### 2025-05-05 eb0e830
Updated Exception messages
Updated exception messages to convert variables to string - helps catch if an incorrect type is provided

Updated plot_dashboard to work with newest version of shiny
