### 2025-05-05 d37d039
Updated detect-outliers page
Updated detect-outliers page

Updated access-files for consistency
