+++ papers/JOSS/paper.bib
@@ -95,6 +95,26 @@
 	file = {Full Text PDF:C\:\\Users\\willi\\Zotero\\storage\\H64S2N83\\Fridlund and Cacioppo - 1986 - Guidelines for Human Electromyographic Research.pdf:application/pdf;Snapshot:C\:\\Users\\willi\\Zotero\\storage\\NLWTMFVK\\j.1469-8986.1986.tb00676.html:text/html},
 }
 
+@Inbook{gabrieli_pysiology_2020,
+	author="Gabrieli, Giulio
+	and Azhari, Atiqah
+	and Esposito, Gianluca",
+	editor="Esposito, Anna
+	and Faundez-Zanuy, Marcos
+	and Morabito, Francesco Carlo
+	and Pasero, Eros",
+	title="PySiology: A Python Package for Physiological Feature Extraction",
+	bookTitle="Neural Approaches to Dynamics of Signal Exchanges",
+	year="2020",
+	publisher="Springer Singapore",
+	address="Singapore",
+	pages="395--402",
+	abstract="Physiological signals have been widely used to measure continuous data from the autonomic nervousGabrieli, Giulio system in the fields of computer science, psychology, and human--computer interaction. Signal processing andAzhari, Atiqah feature estimation of physiological measurements can beEsposito, Gianluca performed with several commercial tools. Unfortunately, those tools possess a steep learning curve and do not usually allow for complete customization of estimation parameters. For these reasons, we designed PySiology, an open-source package for the estimation of features from physiological signals, suitable for both novice and expert users. This package provides clear documentation of utilized methodology, guided functionalities for semi-automatic feature estimation, and options for extensive customization. In this article, a brief introduction to the features of the package, and to its design workflow, are presented. To demonstrate the usage of the package in a real-world context, an advanced example of image valence estimation from physiological measurements (ECG, EMG, and EDA) is described. Preliminary tests have shown high reliability of feature estimated using PySiology.",
+	isbn="978-981-13-8950-4",
+	doi="10.1007/978-981-13-8950-4_35",
+	url="https://doi.org/10.1007/978-981-13-8950-4_35"
+}
+
 @misc{greene_peakaffectds_2022,
 	title = {{PeakAffectDS}},
 	url = {https://zenodo.org/records/6403363},
@@ -204,6 +224,13 @@
 	year = {2020},
 	note = {Publisher: Frontiers},
 	keywords = {clinical application, Kinesiology, Physiotherapy, Rehabilitation, Surface Electromography, tutorial},
+}
+
+@misc{mermaidjs,
+  title        = {{Mermaid.js}: Generation of diagrams and flowcharts from text},
+  author       = {Sveidqvist, Knut and the Mermaid Contributors},
+  howpublished = {\url{https://mermaid-js.github.io}},
+  note         = {Version 11.6.x, accessed April 2025}
 }
 
 @article{phinyomark2009novel,
@@ -352,6 +379,24 @@
   number={1},
   pages={51--58},
   year={1983}
+}
+
+@misc{vitepress,
+  title        = {{VitePress}: Vite \& Vue powered static-site generator},
+  author       = {You, Evan and the VitePress Contributors},
+  howpublished = {\url{https://vitepress.dev}},
+  note         = {1.6.3, accessed April 2025}
+}
+
+@article{virtanen_scipy_2020,
+  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
[…]
