### 2025-05-06 adb7858
Updated documentation consistency
Updated documentation pages (access_files, detect_outliers, extract_features) for sylistic improvements and consistency
