+++ EMGFlow-Package/src/EMGFlow/detect_outliers.py
@@ -15,7 +15,7 @@ from .access_files import *
 #
 
 """
-A collection of functions for finding outliers while testing
+A collection of functions for finding outliers while testing.
 """
 
 #
@@ -26,6 +26,14 @@ def detect_outliers(inPath, samplingRate, threshold, cols=None, low=None, high=N
     """
     Looks at all Signals contained in a filepath, returns a dictionary of file
     names and locations that have outliers.
+    
+    Works by interpolating an inverse function from the peaks of the signal's
+    spectrum. The function then calculates the 'metric' aggregate of the
+    differences between the predicted spectrum intensity of the inverse
+    function, and the actual spectrum intensity of the peaks. Finally, if the
+    largest difference between the predicted and actual values is greater than
+    the metric average multiplied by the threshold value, the file is flagged
+    for having an outlier and is added to the dictionary.
 
     Parameters
     ----------
@@ -33,7 +41,7 @@ def detect_outliers(inPath, samplingRate, threshold, cols=None, low=None, high=N
         Filepath to a directory to read Signal files.
     samplingRate : float
         Sampling rate of the Signal.
-    cols : TYPE
+    cols : list-str
         List of columns of the Signal to search for outliers in. The default is
         None, in which case outliers are searched for in every column except
         for 'time'.
@@ -52,8 +60,8 @@ def detect_outliers(inPath, samplingRate, threshold, cols=None, low=None, high=N
         in which case no upper threshold is used.
     metric : function, optional
         Some summary function that defines the metric used for finding
-        outliers. The default is np.median, but others such as np.mean can be
-        used instead.
+        outliers. The default is 'np.median', but others such as 'np.mean' can
+        be used instead.
     expression : str, optional
         A regular expression. If provided, will only search for outliers in
         files whose names match the regular expression. The default is None.
@@ -67,47 +75,55 @@ def detect_outliers(inPath, samplingRate, threshold, cols=None, low=None, high=N
     Raises
     ------
     Exception
-        An exception is raised if samplingRate is less or equal to 0.
+        'windowSize' must be an integer greater than 0.
     Exception
-        An exception is raised if threshold is less or equal to 0.
+        An exception is raised if 'samplingRate' is less or equal to 0.
     Exception
-        An exception is raised if low is greater than high.
+        An exception is raised if 'threshold' is less or equal to 0.
     Exception
-        An exception is raised if low or high are negative.
+        An exception is raised if 'low' is greater than 'high'.
     Exception
-        An exception is raised if metric is not a valid summary function.
+        An exception is raised if 'low' or 'high' are negative.
     Exception
-        An exception is raised if a column in cols is not in the data file.
+        An exception is raised if 'metric' is not a valid summary function.
     Exception
-        Raises an exception if a file cannot not be read in inPath.
+        An exception is raised if a column in 'cols' is not in a data file.
+    Exception
+        Raises an exception if a file cannot not be read in 'inPath'.
     Exception
         Raises an exception if an unsupported file format was provided for
-        fileExt.
+        'fileExt'.
     Exception
-        Raises an exception if expression is not None or a valid regular
+        Raises an exception if 'expression' is not None or a valid regular
         expression.
 
     Returns
     -------
-    dict
+    outliers : dict-str
         Dictionary of file names/locations as keys/values for each file
         detected that contains an outlier.
 
     """
     
+    if not type(windowSize) == int:
+        raise Exception("'windowSize' must be a valid integer.")
+    
+    if windowSize <= 0:
+        raise Exception("'windowSize' cannot be 0 or less.")
+    
     if expression is not None:
         try:
[…]
