### 2025-05-06 47533a3
Updated paper table
