+++ papers/JOSS/paper.md
@@ -80,7 +80,8 @@ notch_vals_subset = [(150, 25)]
 participant_pattern = '02\/\w+.csv'
 
 # Apply custom notch filter
-EMGFlow.notch_filter_signals(path_names['Notch'], path_names['Notch'], sampling_rate, notch_vals_subset, cols, expression=participant_pattern)
+EMGFlow.notch_filter_signals(path_names['Notch'], path_names['Notch'], 
+sampling_rate, notch_vals_subset, cols, expression=participant_pattern)
 ```
 
 ## Visualization of Preprocessing Stages
@@ -96,10 +97,12 @@ band_high = 140
 smooth_window = 50
 
 # Apply bandpass filter
-EMGFlow.bandpass_filter_signals(path_names['Notch'], path_names['Bandpass'], sampling_rate, band_low, band_high, cols)
+EMGFlow.bandpass_filter_signals(path_names['Notch'], path_names['Bandpass'], 
+sampling_rate, band_low, band_high, cols)
 
 # Apply smoothing filter
-EMGFlow.smooth_filter_signals(path_names['Bandpass'], path_names['Smooth'], smooth_window, cols)
+EMGFlow.smooth_filter_signals(path_names['Bandpass'], path_names['Smooth'], 
+smooth_window, cols)
 
 # Data column and units for plotting
 col = 'EMG_zyg'
