### 2025-05-07 77fa188
Documentation update
Updated code files with more consistent documentation for the in-function documentation
