+++ EMGFlow-Package/src/EMGFlow/preprocess_signals.py
@@ -27,7 +27,7 @@ def emg_to_psd(sigVals, samplingRate=1000, normalize=True):
 
     Parameters
     ----------
-    sigVals : float list
+    sigVals : list-float
         A list of float values. A column of a Signal.
     samplingRate : float
         Sampling rate of the Signal.
@@ -38,12 +38,12 @@ def emg_to_psd(sigVals, samplingRate=1000, normalize=True):
     Raises
     ------
     Exception
-        An exception is raised if the sampling rate is less or equal to 0
+        An exception is raised if 'samplingRate' is less or equal to 0
 
     Returns
     -------
-    psd : DataFrame
-        A DataFrame containing a 'Frequency' and 'Power' column. The Power
+    psd : pd.DataFrame
+        A dataframe containing a 'Frequency' and 'Power' column. The 'Power'
         column indicates the intensity of each frequency in the Signal
         provided. Results will be normalized if 'normalize' is set to True.
     
@@ -98,14 +98,14 @@ def apply_notch_filters(Signal, col, samplingRate, notchVals):
 
     Parameters
     ----------
-    Signal : DataFrame
-        A Pandas DataFrame containing a 'Time' column, and additional columns
+    Signal : pd.DataFrame
+        A Pandas dataframe containing a 'Time' column, and additional columns
         for signal data.
     col : str
         Column of the Signal to apply the filter to.
     samplingRate : float
         Sampling rate of the Signal.
-    notchVals : list
+    notchVals : list-tuple
         A list of (Hz, Q) tuples corresponding to the notch filters being
         applied. Hz is the frequency to apply the filter to, and Q is the
         Q-score (an intensity score where a higher number means a less extreme
@@ -114,17 +114,17 @@ def apply_notch_filters(Signal, col, samplingRate, notchVals):
     Raises
     ------
     Exception
-        An exception is raised if the column is not found in the Signal.
+        An exception is raised if the column 'col' is not found in the Signal.
     Exception
-        An exception is raised if the sampling rate is less or equal to 0.
+        An exception is raised if 'samplingRate' is less or equal to 0.
     Exception
-        An exception is raised if a Hz value in notchVals is greater than
+        An exception is raised if a Hz value in 'notchVals' is greater than
         samplingRate/2 or less than 0
 
     Returns
     -------
-    DataFrame
-        A copy of Signal after the notch filters are applied.
+    Signal : pd.DataFrame
+        A copy of the 'Signal' dataframe with the notch filters are applied.
 
     """
 
@@ -136,18 +136,18 @@ def apply_notch_filters(Signal, col, samplingRate, notchVals):
 
     def apply_notch_filter(Signal, col, samplingRate, notch):
         """
-        Apply a notch filter to a signal
+        Apply a notch filter to a signal.
 
         Parameters
         ----------
-        Signal : DataFrame
-            A Pandas DataFrame containing a 'Time' column, and additional
+        Signal : pd.DataFrame
+            A Pandas dataframe containing a 'Time' column, and additional
             columns for signal data.
         col : str
             Column of the Signal to apply the filter to.
         samplingRate : float
             Sampling rate of the Signal.
-        notch : (int, int) tuple
+        notch : tuple
             Notch filter data. Should be a (Hz, Q) tuple where Hz is the
             frequency to apply the filter to, and Q. is the Q-score (an
             intensity score where a higher number means a less extreme filter).
@@ -160,8 +160,9 @@ def apply_notch_filters(Signal, col, samplingRate, notchVals):
 
         Returns
         -------
-        Signal_col : Series
-            A Pandas Series of the provided column with the notch filter applied
+        SignalCol : Series
[…]
