### 2025-05-08 87704a4
Updated documentation for consistency
Updated documentation across all code files for consistency and correctness.
