+++ EMGFlow-Package/src/EMGFlow/extract_features.py
@@ -22,7 +22,7 @@ A collection of functions for extracting features.
 #
 
 
-def calc_iemg(Signal, col, sr):
+def calc_iemg(Signal, col, samplingRate):
     """
     Calculate the Integreated EMG (IEMG) of a Signal.
 
@@ -33,7 +33,7 @@ def calc_iemg(Signal, col, sr):
         for signal data.
     col : str
         Column of the Signal to apply the summary to.
-    sr : float
+    samplingRate : float
         Sampling rate of the Signal.
 
     Raises
@@ -41,7 +41,7 @@ def calc_iemg(Signal, col, sr):
     Exception
         An exception is raised if 'col' is not a column of 'Signal'.
     Exception
-        An exception is raised if 'sr' is less or equal to 0.
+        An exception is raised if 'samplingRate' is less or equal to 0.
 
     Returns
     -------
@@ -53,10 +53,10 @@ def calc_iemg(Signal, col, sr):
     if col not in list(Signal.columns.values):
         raise Exception("Column '" + str(col) + "' not found in the Signal.")
     
-    if sr <= 0:
+    if samplingRate <= 0:
         raise Exception("Sampling rate cannot be less or equal to 0.")
     
-    IEMG = np.sum(np.abs(Signal[col]) * sr)
+    IEMG = np.sum(np.abs(Signal[col]) * samplingRate)
     return IEMG
 
 #
@@ -184,7 +184,7 @@ def calc_mmav2(Signal, col):
 # =============================================================================
 #
 
-def calc_ssi(Signal, col, sr):
+def calc_ssi(Signal, col, samplingRate):
     """
     Calculate the Simple Square Integreal (SSI) of a Signal.
 
@@ -195,7 +195,7 @@ def calc_ssi(Signal, col, sr):
         for signal data.
     col : str
         Column of the Signal to apply the summary to.
-    sr : float
+    samplingRate : float
         Sampling rate of the Signal.
 
     Raises
@@ -203,7 +203,7 @@ def calc_ssi(Signal, col, sr):
     Exception
         An exception is raised if 'col' is not a column of 'Signal'.
     Exception
-        An exception is raised if sr is less or equal to 0.
+        An exception is raised if samplingRate is less or equal to 0.
 
     Returns
     -------
@@ -215,10 +215,10 @@ def calc_ssi(Signal, col, sr):
     if col not in list(Signal.columns.values):
         raise Exception("Column '" + str(col) + "' not found in the Signal.")
     
-    if sr <= 0:
+    if samplingRate <= 0:
         raise Exception("Sampling rate cannot be 0 or negative")
     
-    SSI = np.sum((np.abs(Signal[col]) * sr) ** 2)
+    SSI = np.sum((np.abs(Signal[col]) * samplingRate) ** 2)
     return SSI
 
 #
@@ -512,7 +512,7 @@ def calc_ap(Signal, col):
 # =============================================================================
 #
 
-def calc_spec_flux(Signal1, diff, col, sr, diffSr=None):
+def calc_spec_flux(Signal1, diff, col, samplingRate, diffSr=None):
     """
     Calculate the spectral flux of a Signal.
 
@@ -529,7 +529,7 @@ def calc_spec_flux(Signal1, diff, col, sr, diffSr=None):
         Column of the Signal to apply the summary to. If a second signal is
         provided for 'diff', a column of the same name should be available for
         use.
-    sr : float
+    samplingRate : float
         Sampling rate of the Signal.
     diffSr : float, optional
         Sampling rate for the second Signal if provided. The default is None,
@@ -541,7 +541,7 @@ def calc_spec_flux(Signal1, diff, col, sr, diffSr=None):
     Exception
[…]
