+++ EMGFlow-Package/src/EMGFlow/preprocess_signals.py
@@ -43,8 +43,8 @@ def emg_to_psd(sigVals, samplingRate=1000, normalize=True):
     Returns
     -------
     psd : pd.DataFrame
-        A dataframe containing a 'Frequency' and 'Power' column. The 'Power'
-        column indicates the intensity of each frequency in the Signal
+        A Pandas dataframe containing a 'Frequency' and 'Power' column. The
+        'Power' column indicates the intensity of each frequency in the Signal
         provided. Results will be normalized if 'normalize' is set to True.
     
     """
@@ -114,7 +114,7 @@ def apply_notch_filters(Signal, col, samplingRate, notchVals):
     Raises
     ------
     Exception
-        An exception is raised if the column 'col' is not found in the Signal.
+        An exception is raised if 'col' is not a column of 'Signal'.
     Exception
         An exception is raised if 'samplingRate' is less or equal to 0.
     Exception
@@ -160,7 +160,7 @@ def apply_notch_filters(Signal, col, samplingRate, notchVals):
 
         Returns
         -------
-        SignalCol : Series
+        SignalCol : pd.Series
             A Pandas series of the provided column with the notch filter
             applied.
 
@@ -220,7 +220,7 @@ def notch_filter_signals(inPath, outPath, samplingRate, notch, cols=None, expres
         A regular expression. If provided, will only filter files whose names
         match the regular expression. The default is None.
     expCopy : bool, optional
-        If true, copies files that don't match the regular expression to the
+        If True, copies files that don't match the regular expression to the
         output folder without filtering. The default is False, which ignores
         files that don't match.
     fileExt : str, optional
@@ -318,8 +318,8 @@ def apply_bandpass_filter(Signal, col, samplingRate, low, high):
 
     Parameters
     ----------
-    Signal : DataFrame
-        A Pandas DataFrame containing a 'Time' column, and additional columns
+    Signal : pd.DataFrame
+        A Pandas dataframe containing a 'Time' column, and additional columns
         for signal data.
     col : str
         Column of the Signal to apply the filter to.
@@ -333,16 +333,19 @@ def apply_bandpass_filter(Signal, col, samplingRate, low, high):
     Raises
     ------
     Exception
-        An exception is raised if the column is not found in the Signal.
+        An exception is raised if 'col' is not a column of 'Signal'.
     Exception
-        An exception is raised if the sampling rate is less or equal to 0.
+        An exception is raised if 'samplingRate' is less or equal to 0.
     Exception
-        An exception is raised if high is not higher than low.
+        An exception is raised if 'high' is not higher than 'low'.
+    Exception
+        An exception is raised if 'high' or 'low' are higher than 1/2 of
+        'samplingRate'
 
     Returns
     -------
-    Signal : DataFrame
-        A copy of Signal after the bandpass filter is applied.
+    Signal : pd.DataFrame
+        A copy of 'Signal' after the bandpass filter is applied.
 
     """
     
@@ -378,7 +381,7 @@ def bandpass_filter_signals(inPath, outPath, samplingRate, low=20, high=450, col
     
     Parameters
     ----------
-    inPath : dict
+    inPath : dict-str
         Filepath to a directory to read Signal files.
     outPath : str
         Filepath to an output directory.
@@ -396,7 +399,7 @@ def bandpass_filter_signals(inPath, outPath, samplingRate, low=20, high=450, col
         A regular expression. If provided, will only filter files whose names
         match the regular expression. The default is None.
     expCopy : bool, optional
-        If true, copies files that don't match the regular expression to the
+        If True, copies files that don't match the regular expression to the
         output folder without filtering. The default is False, which ignores
         files that don't match.
     fileExt : str, optional
[…]
