### 2025-05-08 5f288ed
Updated READMEs
Updated the README pages for the GitHub, and PyPi pages.

Updated the pyproject.toml for the new dependency (importlib-resources)
