+++ EMGFlow-Package/src/EMGFlow/access_files.py
@@ -15,6 +15,112 @@ A collection of functions for accessing files.
 # =============================================================================
 #
 
+def make_paths(root=None):
+    """
+    Generates a file structure for signal files, and returns a dictionary of
+    the locations for these files.
+    
+    A 'Data' folder is created, with 'Raw', 'Notch', 'Bandpass', 'Smooth' and
+    'Feature' subfolders.
+
+    Parameters
+    ----------
+    root : str, optional
+        Root of the data to be generated. The default is None.
+
+    Returns
+    -------
+    pathNames : dict-str
+        A dictionary of file locations with keys for stage in the processing
+        pipeline.
+
+    """
+    
+    if root is None:
+        root = os.path.join(os.getcwd(), 'Data')
+    else:
+        root = os.path.normpath(root)
+    
+    # Create dictionary
+    pathNames = {
+        'Raw':os.path.join(root, 'Raw'),
+        'Notch':os.path.join(root, 'Notch'),
+        'Bandpass':os.path.join(root, 'Bandpass'),
+        'Smooth':os.path.join(root, 'Smooth'),
+        'Feature':os.path.join(root, 'Feature')
+    }
+    
+    # Create folders
+    for value in pathNames.values():
+        os.makedirs(value, exist_ok=True)
+    
+    # Return dictionary
+    return pathNames
+
+#
+# =============================================================================
+#
+
+def make_sample_data(pathNames):
+    """
+    Generates sample data in the 'Raw' folder of a provided dictionary of file
+    locations.
+
+    Parameters
+    ----------
+    pathNames : dict-str
+        Dictionary of file locations.
+
+    Raises
+    ------
+    Exception
+        An exception is raised if the provided 'pathNames' dictionary doesn't
+        contain a 'Raw' path key.
+    Exception
+        An exception is raised if the sample data cannot be loaded.
+
+    Returns
+    -------
+    None.
+
+    """
+    
+    # Check that a 'raw' folder exists
+    if 'Raw' not in pathNames:
+        raise Exception('Raw path not detected in pathNames.')
+    
+    # Load the sample data
+    try:
+        sample_data_01 = pd.read_csv(importlib_resources.files("EMGFlow").joinpath(os.path.join("data", "sample_data_01.csv")))
+        sample_data_02 = pd.read_csv(importlib_resources.files("EMGFlow").joinpath(os.path.join("data", "sample_data_02.csv")))
+        sample_data_03 = pd.read_csv(importlib_resources.files("EMGFlow").joinpath(os.path.join("data", "sample_data_03.csv")))
+        sample_data_04 = pd.read_csv(importlib_resources.files("EMGFlow").joinpath(os.path.join("data", "sample_data_04.csv")))
+    except:
+        raise Exception('Failed to load EMGFlow sample data.')
+    
+    # Write the sample data
+    os.makedirs(os.path.join(pathNames['Raw'], '01'), exist_ok=True)
+    os.makedirs(os.path.join(pathNames['Raw'], '02'), exist_ok=True)
+    
+    data_path_01 = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
+    data_path_02 = os.path.join(pathNames['Raw'], '01', 'sample_data_02.csv')
+    data_path_03 = os.path.join(pathNames['Raw'], '02', 'sample_data_03.csv')
+    data_path_04 = os.path.join(pathNames['Raw'], '02', 'sample_data_04.csv')
+    
+    if not os.path.exists(data_path_01):
+        sample_data_01.to_csv(data_path_01, index=False)
+    if not os.path.exists(data_path_02):
+        sample_data_02.to_csv(data_path_02, index=False)
+    if not os.path.exists(data_path_03):
[…]
