+++ EMGFlow-Package/tests/test_extract_features.py
@@ -1,156 +1,278 @@
 import unittest
-import pandas as pd
-import numpy as np
+import importlib
 import os
-import sys
+import shutil
 
-from src.EMGFlow.ExtractFeatures import *
+# Load EMGFlow from local files
+filePath = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src', 'EMGFlow', '__init__.py'))
+spec = importlib.util.spec_from_file_location("EMGFlow", filePath)
+EMGFlow = importlib.util.module_from_spec(spec)
+spec.loader.exec_module(EMGFlow)
 
-test_df = pd.DataFrame({'r1':[1,2,4,2,5,3,1,5,7,3,7,8,4,2,5,3,5,3,2,1,6,3,6,1,2]})
-test_df_2 = pd.DataFrame({'r1':[1,-2,3,-4,5,6,-7]})
-test_sr = 1000
+import numpy as np
 
 class TestSimple(unittest.TestCase):
+
+#
+# =============================================================================
+#
     
     def setUp(self):
-        if os.path.exists('./Testing') == False:
-            os.mkdir('./Testing')
-            time_col = np.array(range(500)) / 100
-            emg_col = np.sin(time_col) + (np.random.rand(500)/10)
-            df = pd.DataFrame({'Time':time_col, 'EMG':emg_col})
-            df.to_csv('./Testing/Data.csv', index=False)
-        if os.path.exists('./Testing_out') == False:
-            os.mkdir('./Testing_out')
-        if os.path.exists('./Testing_plots') == False:
-            os.mkdir('./Testing_plots')
+        pass
 
 #
 # =============================================================================
 #
 
-    # =========================
-    # TEST FEATURE CALCULATIONS
-    # =========================
 
-    def test_CalcIEMG(self):
-        val = CalcIEMG(test_df, 'r1', test_sr)
-        self.assertAlmostEqual(val, 91000, 6)
+    def test_calc_iemg(self):
+        pathNames = EMGFlow.make_paths()
+        EMGFlow.make_sample_data(pathNames)
+        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
+        Signal = EMGFlow.read_file_type(filePath, 'csv')
+        IEMG = EMGFlow.calc_iemg(Signal, 'EMG_zyg', 2000)
+        self.assertIsInstance(IEMG, float)
     
-    def test_CalcMAV(self):
-        val = CalcMAV(test_df, 'r1')
-        self.assertAlmostEqual(val, 3.64, 6)
+    def test_calc_mav(self):
+        pathNames = EMGFlow.make_paths()
+        EMGFlow.make_sample_data(pathNames)
+        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
+        Signal = EMGFlow.read_file_type(filePath, 'csv')
+        MAV = EMGFlow.calc_mav(Signal, 'EMG_zyg')
+        self.assertIsInstance(MAV, float)
     
-    def test_CalcMMAV(self):
-        val = CalcMMAV(test_df, 'r1')
-        self.assertAlmostEqual(val, 2.9, 6)
+    def test_calc_mmav1(self):
+        pathNames = EMGFlow.make_paths()
+        EMGFlow.make_sample_data(pathNames)
+        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
+        Signal = EMGFlow.read_file_type(filePath, 'csv')
+        MMAV1 = EMGFlow.calc_mmav1(Signal, 'EMG_zyg')
[…]
