+++ EMGFlow-Package/tests/test_preprocess_signals.py
@@ -1,168 +1,119 @@
 import unittest
-import pandas as pd
-import numpy as np
+import importlib
 import os
-import sys
+import shutil
 
-from src.EMGFlow.PreprocessSignals import *
+# Load EMGFlow from local files
+filePath = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src', 'EMGFlow', '__init__.py'))
+spec = importlib.util.spec_from_file_location("EMGFlow", filePath)
+EMGFlow = importlib.util.module_from_spec(spec)
+spec.loader.exec_module(EMGFlow)
 
-test_df = pd.DataFrame({'r1':[1,2,4,2,5,3,1,5,7,3,7,8,4,2,5,3,5,3,2,1,6,3,6,1,2]})
-test_df_2 = pd.DataFrame({'r1':[1,-2,3,-4,5,6,-7]})
-test_sr = 1000
+import pandas as pd
 
 class TestSimple(unittest.TestCase):
+
+#
+# =============================================================================
+#
     
     def setUp(self):
-        if os.path.exists('./Testing') == False:
-            os.mkdir('./Testing')
-            time_col = np.array(range(500)) / 100
-            emg_col = np.sin(time_col) + (np.random.rand(500)/10)
-            df = pd.DataFrame({'Time':time_col, 'EMG':emg_col})
-            df.to_csv('./Testing/Data.csv', index=False)
-        if os.path.exists('./Testing_out') == False:
-            os.mkdir('./Testing_out')
-        if os.path.exists('./Testing_plots') == False:
-            os.mkdir('./Testing_plots')
+        pathNames = EMGFlow.make_paths()
+        EMGFlow.make_sample_data(pathNames)
+        pass
 
 #
 # =============================================================================
 #
-
-    # ============================
-    # TEST MAIN PIPELINE FUNCTIONS
-    # ============================
-
-    def test_NotchFilterSignals(self):
-        NotchFilterSignals('./Testing/', './Testing_out/', 100, [(10,4)], cols=['EMG'])
-
-    def test_BandpassFilterSignals(self):
-        BandpassFilterSignals('./Testing/', './Testing_out/', 100, 10, 40, cols=['EMG'])
-        
-    def test_SmoothFilterSignals(self):
-        SmoothFilterSignals('./Testing/', './Testing_out/', 5, ['EMG'])
     
-
-#
-# =============================================================================
-#
-
-    def test_EMG2PSD(self):
-        test = EMG2PSD(test_df['r1']).round(6)
-        ans = pd.DataFrame({'Frequency':[166.666667,208.333333,250.000000,
-                                         291.666667,333.333333,375.000000,
-                                         416.666667,458.333333,500.000000],
-                          'Power':[0.730713,0.797440,0.900509,
-                                   1.000000,0.887714,0.456161,
-                                   0.119441,0.310694,0.275484],
-                          '':[4,5,6,7,8,9,10,11,12]}).set_index('')
-        self.assertTrue(ans.equals(test))
+    def test_emg_to_psd(self):
+        pathNames = EMGFlow.make_paths()
+        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
+        Signal = EMGFlow.read_file_type(filePath, 'csv')
+        PSD = EMGFlow.emg_to_psd(Signal['EMG_zyg'], 2000)
+        self.assertIsInstance(PSD, pd.DataFrame)
     
[…]

[…]
