### 2025-05-13 78db377
Updated variable case
Changed case of all variables to snake_case
