+++ EMGFlow-Package/src/EMGFlow/access_files.py
@@ -30,7 +30,7 @@ def make_paths(root=None):
 
     Returns
     -------
-    pathNames : dict-str
+    path_names : dict-str
         A dictionary of file locations with keys for stage in the processing
         pipeline.
 
@@ -42,7 +42,7 @@ def make_paths(root=None):
         root = os.path.normpath(root)
     
     # Create dictionary
-    pathNames = {
+    path_names = {
         'Raw':os.path.join(root, 'Raw'),
         'Notch':os.path.join(root, 'Notch'),
         'Bandpass':os.path.join(root, 'Bandpass'),
@@ -51,30 +51,30 @@ def make_paths(root=None):
     }
     
     # Create folders
-    for value in pathNames.values():
+    for value in path_names.values():
         os.makedirs(value, exist_ok=True)
     
     # Return dictionary
-    return pathNames
+    return path_names
 
 #
 # =============================================================================
 #
 
-def make_sample_data(pathNames):
+def make_sample_data(path_names):
     """
     Generates sample data in the 'Raw' folder of a provided dictionary of file
     locations.
 
     Parameters
     ----------
-    pathNames : dict-str
+    path_names : dict-str
         Dictionary of file locations.
 
     Raises
     ------
     Exception
-        An exception is raised if the provided 'pathNames' dictionary doesn't
+        An exception is raised if the provided 'path_names' dictionary doesn't
         contain a 'Raw' path key.
     Exception
         An exception is raised if the sample data cannot be loaded.
@@ -86,8 +86,8 @@ def make_sample_data(pathNames):
     """
     
     # Check that a 'raw' folder exists
-    if 'Raw' not in pathNames:
-        raise Exception('Raw path not detected in pathNames.')
+    if 'Raw' not in path_names:
+        raise Exception('Raw path not detected in path_names.')
     
     # Load the sample data
     try:
@@ -99,13 +99,13 @@ def make_sample_data(pathNames):
         raise Exception('Failed to load EMGFlow sample data.')
     
     # Write the sample data
-    os.makedirs(os.path.join(pathNames['Raw'], '01'), exist_ok=True)
-    os.makedirs(os.path.join(pathNames['Raw'], '02'), exist_ok=True)
+    os.makedirs(os.path.join(path_names['Raw'], '01'), exist_ok=True)
+    os.makedirs(os.path.join(path_names['Raw'], '02'), exist_ok=True)
     
-    data_path_01 = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-    data_path_02 = os.path.join(pathNames['Raw'], '01', 'sample_data_02.csv')
-    data_path_03 = os.path.join(pathNames['Raw'], '02', 'sample_data_03.csv')
-    data_path_04 = os.path.join(pathNames['Raw'], '02', 'sample_data_04.csv')
+    data_path_01 = os.path.join(path_names['Raw'], '01', 'sample_data_01.csv')
+    data_path_02 = os.path.join(path_names['Raw'], '01', 'sample_data_02.csv')
+    data_path_03 = os.path.join(path_names['Raw'], '02', 'sample_data_03.csv')
+    data_path_04 = os.path.join(path_names['Raw'], '02', 'sample_data_04.csv')
     
     if not os.path.exists(data_path_01):
         sample_data_01.to_csv(data_path_01, index=False)
@@ -121,7 +121,7 @@ def make_sample_data(pathNames):
 # =============================================================================
 #
 
-def read_file_type(path, fileExt):
+def read_file_type(path, file_ext):
     """
     Safe wrapper for reading files of a given extension.
     
@@ -134,7 +134,7 @@ def read_file_type(path, fileExt):
     ----------
     path : str
         Path of the file to read.
-    fileExt : str
[…]
