### 2025-05-13 1ded9dd
Updated examples with variable names
Updated all example vatiable names to snake_case
