### 2025-05-13 32dfa06
Update access_files.py
Updated documentation for make_paths
