### 2025-05-13 481641a
Updated descriptions
Updated function descriptions
