+++ EMGFlow-Package/src/EMGFlow/preprocess_signals.py
@@ -38,6 +38,9 @@ def emg_to_psd(sig_vals, sampling_rate=1000, normalize=True):
     Raises
     ------
     Exception
+        An exception is raised if 'sig_vals' is a pd.DataFrame, not a column of
+        a dataframe.
+    Exception
         An exception is raised if 'sampling_rate' is less or equal to 0
 
     Returns
@@ -48,6 +51,9 @@ def emg_to_psd(sig_vals, sampling_rate=1000, normalize=True):
         provided. Results will be normalized if 'normalize' is set to True.
     
     """
+    
+    if isinstance(sig_vals, pd.DataFrame):
+        raise Exception("sig_vals must be a column of the dataframe, not the entire dataframe.")
     
     if sampling_rate <= 0:
         raise Exception("Sampling rate must be greater or equal to 0")
