### 2025-05-13 bb61d57
Update access-files.md
Updated documentation with new function conventions
