+++ docs/reference/access-files.md
@@ -19,6 +19,10 @@ mindmap
         EF(Extract Features)
 ```
 
+
+
+
+
 ## `make_paths`
 
 **Description**
@@ -33,7 +37,7 @@ make_paths(root=None)
 
 **Parameters**
 
-`root`: str (None)
+`root`: str, optional (None)
 - Root of the data to be generated. The default is None.
 
 **Returns**
@@ -48,74 +52,92 @@ make_paths(root=None)
 path_names = EMGFlow.make_paths()
 ```
 
+
+
+
+
 ## `make_sample_data`
 
 **Description**
 
 Generates sample data in the 'Raw' folder of a provided dictionary of file locations.
 
+Creates '01' and '02' folders, which each contain two sample data files ('01/sample_data_01.csv', '01/sample_data_02.csv', '02/sample_data_03.csv', '02/sample_data_04.csv')
+
 ```python
-make_sample_data(pathNames)
+make_sample_data(path_names)
 ```
 
 **Parameters**
 
-`pathNames`: dict-str
-- A dictionary of string file locations
+`path_names`: dict-str
+- Dictionary of file locations.
 
-**Error**
+**Raises**
 
-Raises an error if `pathNames` does not contain a `Raw` path key.
+An exception is raised if the provided `path_names` dictionary doesn't contain a 'Raw' path key.
+
+An exception is raised if the sample data cannot be loaded.
 
 **Returns**
 
-`make_sample_data`: None
+None.
 
 **Example**
 
 ```python
 # Create file paths, then create sample data
-pathNames = EMGFlow.make_paths()
-EMGFlow.make_sample_data(pathNames)
+path_names = EMGFlow.make_paths()
+EMGFlow.make_sample_data(path_names)
 ```
+
+
+
+
 
 ## `read_file_type`
 
 **Description**
 
-`read_file_type` is a safe wrapper for reading files of a given extension.
+Safe wrapper for reading files of a given extension.
+
+Switches between different reading methods based on the instructions provided.
+    
+Supported formats that can be read are: 'csv'.
 
 ```python
-read_file_type(path, fileExt)
+read_file_type(path, file_ext)
 ```
 
 **Parameters**
 
 `path`: str
-- String filepath of file to read.
+- Path of the file to read.
 
-`fileExt`: str
-- String extension of the files to read.
+`file_ext`: str
+- File extension to read.
 
-**Error**
+**Raises**
 
-Raises an error if the file could not be read.
+An exception is raised if the file could not be read.
 
-Raises an error if an unsupported file format was provided for `fileExt`.
+An exception is raised if an unsupported file format was provided for `file_ext`.
 
 **Returns**
 
-`read_file_type`: pd.DataFrame
+`file`: pd.DataFrame
 - Returns a Pandas dataframe of the file contents.
 
 **Example**
 
 ```python
 # Read a csv file
-path = 'data/raw/file01.csv'
+path = '/Data/Raw/01/sample_data_01.csv'
 ext = 'csv'
 df = EMGFlow.read_file_type(path, ext)
 ```
+
+
 
 
 
@@ -123,45 +145,47 @@ df = EMGFlow.read_file_type(path, ext)
 
 **Description**
 
-`map_files` generates a dictionary of file name and location keys/values from a folder and its subfolders.
+Generate a dictionary of file names and locations (keys/values) from the subfiles of a folder.
 
 ```python
-map_files(inPath, fileExt='csv', expression=None, base=None)
+map_files(in_path, file_ext='csv', expression=None, base=None)
 ```
 
 **Parameters**
 
-`inPath`: str
-- String filepath to a directory containing Signal files.
+`in_path`: str
+- The filepath to a directory to read Signal files.
 
[…]
