### 2025-05-13 8b88927
Update access_files.py
Updated code for consistency with documentation
