### 2025-05-14 c005957
Updated documentation
Updated detect-outliers documentation for new format

Changed "Signal" to lower case in access-files
