### 2025-05-14 03395d6
Update detect_outliers.py
Updated order of parameter documentation to match function
