+++ docs/reference/preprocess-signals.md
@@ -10,182 +10,192 @@ mindmap
         AF(File Access)
         DO(Detect Outliers)
         PrS(Preprocess Signals)
-            EMG2PSD
-            ApplyNotchFilters
-            NotchFilterSignals
-            ApplyBandpassFilter
-            BandpassFilterSignals
-            ApplyFWR
-            ApplyBoxcarSmooth
-            ApplyRMSSmooth
-            ApplyGaussianSmooth
-            ApplyLoessSmooth
-            SmoothFilterSignals
+            emg_to_psd
+            apply_notch_filters
+            notch_filter_signals
+            apply_bandpass_filter
+            bandpass_filter_signals
+            apply_fwr
+            apply_boxcar_smooth
+            apply_rms_smooth
+            apply_gaussian_smooth
+            apply_loess_smooth
+            smooth_filter_signals
         PlS(Plot Signals)
         EF(Extract Features)
 ```
 
-## `EMG2PSD`
+
+
+
+
+## `emg_to_psd`
 
 ### Description
 
-Creates a PSD (power spectrum density) of a Signal. Uses the Welch method, meaning it can be used as a Long Term Average Spectrum (LTAS).
+Creates a PSD (power spectrum density) dataframe of a Signal. Uses the Welch method, meaning it can be used as a Long Term Average Spectrum (LTAS).
 
 ```python
-EMG2PSD(Sig_vals, sr=1000, normalize=True)
+emg_to_psd(sig_vals, sampling_rate=1000, normalize=True)
 ```
 
 ### Parameters
 
-`Sig_vals`: float list
-- A list of float values. A column of a Signal.
+`sig_vals`: list-float
+- A list of float values. A column of a signal dataframe.
 
-`sr`: int/float (1000)
-- Numerical value of the sampling rate of the `Signal`. This is the number of entries recorded per second, or the inverse of the difference in time between entries.
+`sampling_rate`: int, float, optional (1000)
+- Sampling rate of `sig_vals`. This is the number of entries recorded per second, or the inverse of the difference in time between entries.
 
-`normalize`: bool (True)
-- Normalizes the result of the PSD by its maximum strength.
+`normalize`: bool, optional (True)
+- If True, will normalize the result of the PSD by its maximum strength. If False, will not. The default is True.
+
+### Raises
+
+An exception is raised if `sig_vals` is a `pd.DataFrame`, not a column of a dataframe.
+
+An exception is raised if `sampling_rate` is less or equal to 0.
 
 ### Returns
 
-`EMG2PSD`: pd.DataFrame
-- Returns a dictionary of frequencies and related strengths with the columns "Frequency" and "Power".
-
-### Error
-
-Raises an error if the sampling rate is less or equal to 0.
+`psd`: pd.DataFrame
+- A Pandas dataframe containing a 'Frequency' and 'Power' column. The 'Power' column indicates the intensity of each corresponding frequency in `sig_vals`. Results will be normalized if `normalize` is set to True.
 
 ### Example
 
 ```python
-sr = 2000
-PSD = EMGFlow.EMG2PSD(SignalDF['col1'], sr)
+sampling_rate = 2000
+PSD = EMGFlow.emg_to_psd(Signal['EMG_zyg'], sampling_rate)
 ```
 
 
 
-## `ApplyNotchFilters`
+
+
+## `apply_notch_filters`
 
 **Description**
 
-Applies a list of notch filters to a `Signal`, using the `scipy.signal.iirnotch` method. Returns a new DataFrame object and does not modify the `Signal` provided.
+Applies a list of notch filters to a `Signal` dataframe, using the `scipy.signal.iirnotch` method. Returns a new dataframe object and does not modify the input `Signal`.
 
[…]
