### 2025-05-15 36dddc3
Visual changes
Changed page display order

Updated function mindmaps
