### 2025-05-15 3c625e1
Update preprocess_signals.py
Updated return variable for apply_x_filter functions to be distinct from Signal input
