+++ EMGFlow-Package/src/EMGFlow/preprocess_signals.py
@@ -41,7 +41,7 @@ def emg_to_psd(sig_vals, sampling_rate=1000, normalize=True):
         An exception is raised if 'sig_vals' is a pd.DataFrame, not a column of
         a dataframe.
     Exception
-        An exception is raised if 'sampling_rate' is less or equal to 0
+        An exception is raised if 'sampling_rate' is less or equal to 0.
 
     Returns
     -------
@@ -125,11 +125,11 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
         An exception is raised if 'sampling_rate' is less or equal to 0.
     Exception
         An exception is raised if a Hz value in 'notch_vals' is greater than
-        sampling_rate/2 or less than 0
+        sampling_rate/2 or less than 0.
 
     Returns
     -------
-    Signal : pd.DataFrame
+    notch_Signal : pd.DataFrame
         A copy of the 'Signal' dataframe with the notch filters are applied.
 
     """
@@ -189,12 +189,12 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
         
         return SignalCol
     
-    Signal = Signal.copy()
+    notch_Signal = Signal.copy()
     
     # Applies apply_notch_filter for every notch_val tuple
     for i in range(len(notch_vals)):
-        Signal[col] = apply_notch_filter(Signal, col, sampling_rate, notch_vals[i])
-    return Signal
+        notch_Signal[col] = apply_notch_filter(notch_Signal, col, sampling_rate, notch_vals[i])
+    return notch_Signal
 
 #
 # =============================================================================
@@ -244,7 +244,7 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
         An exception is raised if 'sampling_rate' is less or equal to 0.
     Exception
         An exception is raised if a Hz value in 'notch_vals' is greater than
-        sampling_rate/2 or less than 0
+        sampling_rate/2 or less than 0.
     Exception
         An exception is raised if a file cannot not be read in 'in_path'.
     Exception
@@ -346,11 +346,11 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high):
         An exception is raised if 'high' is not higher than 'low'.
     Exception
         An exception is raised if 'high' or 'low' are higher than 1/2 of
-        'sampling_rate'
+        'sampling_rate'.
 
     Returns
     -------
-    Signal : pd.DataFrame
+    band_Signal : pd.DataFrame
         A copy of 'Signal' after the bandpass filter is applied.
 
     """
@@ -368,12 +368,12 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high):
         raise Exception("'high' must be higher than 'low'.")
     
     
-    Signal = Signal.copy()
+    band_Signal = Signal.copy()
     # Here, the "5" is the order of the butterworth filter
     # (how quickly the signal is cut off)
     b, a = scipy.signal.butter(5, [low, high], fs=sampling_rate, btype='band')
-    Signal[col] = scipy.signal.lfilter(b, a, Signal[col])
-    return Signal
+    band_Signal[col] = scipy.signal.lfilter(b, a, band_Signal[col])
+    return band_Signal
 
 #
 # =============================================================================
@@ -425,7 +425,7 @@ def bandpass_filter_signals(in_path, out_path, sampling_rate, low=20, high=450, 
         An exception is raised if 'high' is not higher than 'low'.
     Exception
         An exception is raised if 'high' or 'low' are higher than 1/2 of
-        'sampling_rate'
+        'sampling_rate'.
     Exception
[…]
