### 2025-05-15 6517bb2
Updated documentation
Updated plot-signals page of documentation

Fixed mistake on preprocess-signals page
