+++ docs/reference/plot-signals.md
@@ -11,203 +11,72 @@ mindmap
         DO(Detect Outliers)
         PrS(Preprocess Signals)
         PlS(Plot Signals)
-            GenPlotDash
-            PlotSpectrum
-            PlotCompareSignals
+            plot_dashboard
         EF(Extract Features)
 ```
 
-## `GenPlotDash`
+## `plot_dashboard`
 
 **Description:**
 
-Creates a HTML dashboard from a series of signal file paths to compare filtering progress at different stages. Has a side bar menu to navigate the signal file and stage being displayed.
+Creates an HTML dashboard from a `path_names` filepath dictionary (see `make_paths`). Has a side bar menu to navigate the signal file and stage being displayed.
 
 The visualization is created in the default browser, and is opened automatically. The function will automatically create plots for any paths provided in the dictionary, using the keys as the legend.
 
+Each filepath in the dictionary is expected to contain the same named files to display them at the different stages. To avoid error, the `plot_dashboard` ignores the 'Feature' filepath.
+
 ```python
-GenPlotDash(path_names, sampling_rate, col, units, expression=None, file_ext='csv', autorun=True)
+plot_dashboard(path_names, col, units, expression=None, file_ext='csv', autorun=True)
 ```
 
 **Parameters:**
 
-`path_names`: dictionary of strings
-- A dictionary of keys (stage of preprocessing) and values (filepath to that stage).
-
-`sampling_rate`: int/float
-- Numerical value of the sampling rate of the `Signal`. This is the number of entries recorded per second, or the inverse of the difference in time between entries.
+- A dictionary of keys (stage of preprocessing) and values (filepath to that stage). The provided dictionary is required to have a `Raw`, `Notch`, `Bandpass`, and `Smooth` path.
 
 `col`: str
-- String column name to display in the visualization.
+- Column of the signal files being plotted.
 
 `units`: str
-- Units to use for the y axis of the plot, same units used for the column values.
+- Units to use for the y axis of the plot. The same units as recorded in the column values.
 
-`expression`: str (None)
-- String regular expression. If provided, will only create visualizations for `Signal` files whose names match the regular expression, and will ignore everything else.
+`expression`: str, optional (None)
+- A regular expression. If provided, will only visualize files whose names match the regular expression. The default is None.
 
-`file_ext`: str ("csv")
-- String extension of the files to read. Any file in `in_path` with this extension will be considered to be a `Signal` file, and treated as such. The default is `'csv'`.
+`file_ext`: str, optional ("csv")
+- File extension for files to read. Only reads files with this extension. The default is 'csv'.
 
-`autorun`: bool (True)
-- Boolean controlling the behavior of the function. If True (default), will automatically run the visualization in the default browser. If False, will return a shiny.App instance.
+`autorun`: bool, optional (True)
+- Boolean controlling the behavior of the function. If True (default), will automatically run the visualization in the default browser. If False, will return a `shiny.App` instance.
 
-**Returns:**
+**Raises**
 
-`GenPlotDash`: None or shiny.app
+An exception is raised if the directories in `path_names` don't contain the same files.
 
-**Error**
+An exception is raised if `col` is not found in any of the signal files read.
 
-Raises an error if `col` is a list.
+An exception is raised if a file cannot not be read in a path in `path_names`.
 
-Raises an error if the directories referenced in `path_names` don't contain the same files.
[…]
