### 2025-05-15 d809c09
Updated extract-features mindmap
Updated the mindmap on the extract-features page
