### 2025-05-16 6a2cae2
Updated documentation pages
Updated extract-features with new documentation

Made minor changes to the other pages to make them more consistent with the source code documentation
