+++ docs/reference/plot-signals.md
@@ -1,6 +1,6 @@
-# `PlotSignals` Module Documentation
+# `plot_signals` Module Documentation
 
-The plotting module `PlotSignals` provides functions to help visualize individual, or large batches of signal data. This helps visually see what is happening in a signal to identify outliers, and determine the kinds of filters that need to be applied.
+These functions help visualize individual, or large batches of signal data. This helps visually see what is happening in a signal to identify outliers, and determine the kinds of filters that need to be applied.
 
 ## Module Structure
 
