+++ docs/reference/extract-features.md
@@ -1,6 +1,6 @@
-# `ExtractFeatures` Module Documentation
+# `extract_features` Module Documentation
 
-This module takes preprocessed data, and extracts features from the sEMG signal that capture information in both time and frequency domains. The main function to do this is `ExtractFeatures`. Within this call, individual features are calculated by their own functions, allowing them to be incorporated into your own workflow.
+These functions extract features from the sEMG signal, capturing information in both time and frequency domains. The main function to do this is `extract_features`. Within this call, individual features are calculated by their own functions, allowing them to be incorporated into your own workflow.
 
 ## Module Structure
 
@@ -45,24 +45,16 @@ mindmap
 
 ## Basic Time-Series Statistics
 
-The `AnalyzeSignal` function calculates some basic statistics that don't involve their own functions. This includes:
-- Minimum voltage
-- Maximum voltage
-- Mean voltage
-- Standard deviation of voltage
-- Skew of voltage
-- Kurtosis of voltage
-- Maximum frequency
+The `extract_features` function calculates some basic statistics that don't involve their own functions. This includes:
+- Minimum voltage (using `np.min`)
+- Maximum voltage (using `np.max`)
+- Mean voltage (using `np.mean`)
+- Standard deviation of voltage (using `np.std`)
+- Skew of voltage (using `scipy.stats.skew`)
+- Kurtosis of voltage (using `scipy.stats.kurtosis`)
+- Maximum frequency (using `np.max`)
 
-The values returned will be a voltage equivalent to the units used in the data provided.
-
-Most of these are commonly understood concepts:
-- **Minimum** is calculated with `np.min`
-- **Maximum** is calculated with `np.max`
-- **Mean** is calculated with `np.mean`
-- **Standard Deviation** is calculated with `np.std`
-
-Skew and kurtosis are lesser known statistical measurements.
+Skew and kurtosis are less common, and have a more detailed explanation below.
 
 ### Skew
 
@@ -93,14 +85,16 @@ $$
 
 
 
-## `CalcIEMG`
+
+
+## `calc_iemg`
 
 **Description**
 
 Calculates the Integrated EMG (IEMG) of the signal. The IEMG measures the area under the curve of the signal, which can provide useful information about muscle activity. In an EMG signal, the IEMG describes when the muscle begins contracting, and is related to the signal sequence firing point (Phinyomark et al., 2009).
 
 ```python
-CalcIEMG(Signal, col, sr)
+calc_iemg(Signal, col, sampling_rate)
 ```
 
 **Theory**
@@ -119,42 +113,44 @@ $$
 **Parameters**
 
 `Signal`: pd.DataFrame 
-- Should have one column called "`Time`" for the time indexes, and other named columns for the values at those times.
+- A Pandas dataframe containing a 'Time' column, and additional columns for signal data.
 
 `col`: str
-- String name of a column in `Signal` the filters are being applied to.
+- Column of `Signal` the feature is calculated from.
 
-`sr`: int/float
-- Numerical value of the sampling rate of the `Signal`. This is the number of entries recorded per second, or the inverse of the difference in time between entries.
[…]
