### 2025-05-21 d2d7000
Updated extract-features flowchart
Updated names of some functions
