### 2025-05-26 f05d13e
Updated documentation for simplicity
Dropped "documentation" from documentation headings

Updated title hierarchy of "extract_features" module

Updated flowchart on "getting-started" page
