+++ docs/reference/extract-features.md
@@ -1,4 +1,4 @@
-# `extract_features` Module Documentation
+# `extract_features` Module
 
 These functions extract features from the sEMG signal, capturing information in both time and frequency domains. The main function to do this is `extract_features`. Within this call, individual features are calculated by their own functions, allowing them to be incorporated into your own workflow.
 
@@ -43,7 +43,94 @@ mindmap
             extract_features
 ```
 
-## Basic Time-Series Statistics
+
+
+## Extract Features
+
+The `extract_features` function is the main function of the `extract_features` module. It applies all the feature extraction functions available in this module to signal dataframe files, and generates a comprehensive 'Features.csv' file.
+
+Each extracted feature is available as its own function for easy use.
+
+### `extract_features`
+
+**Description**
+
+Extracts usable features from two sets of signal files (before and after being smoothed). Writes output to a new folder directory, as specified by the 'Feature' key value of the `path_names` dictionary. Output is both saved to the disk as a plaintext file, 'Features.csv', and is also returned as a dataframe object.
+
+Components of a `Signal` dataframe:
+- Has a column named `Time` containing time indexes
+- `Time` indexes are all equally spaced apart
+- Has one (or more) columns with any other name, holding the value of the electrical signal read at that time
+
+All files contained within the folder and subfolder with the proper extension are assumed to be signal files. All signal files within the folder and subfolders should have the same change in time between entries.
+```python
+extract_features(path_names, sampling_rate, cols=None, expression=None, file_ext='csv', short_name=True)
+```
+
+**Theory**
+
+This function requires a path to smoothed and unsmoothed data. This is because while time-series features are extracted from smoothed data, spectral features are not. High-frequency components of the signal can be lost in the smoothing, and we want to ensure the spectral features are as accurate as possible.
+
+**Parameters**
+
+`path_names`: dict-str
+- A dictionary of file locations with keys for the stage in the processing pipeline.
+
+`sampling_rate`: int, float
+- Sampling rate of the signal files. This is the number of entries recorded per second, or the inverse of the difference in time between entries.
+
+`cols`: str, optional (None)
+- List of columns of the signal files to extract features from. The default is None, in which case features are extracted from every column except for 'Time'.
+
+`expression`: str, optional (None)
+- A regular expression. If provided, will only extract features from files whose names match the regular expression. The default is None.
+
+`file_ext`: str, optional ('csv')
+- File extension for files to read. Only reads files with this extension. The default is 'csv'.
+
+`short_names`: bool, optional (True)
+- If true, makes the key column of the feature files the relative path of the file. If false, uses the full system path. The default is True.
+
+**Raises**
+
[…]
