+++ docs/reference/detect-outliers.md
@@ -1,4 +1,4 @@
-# `detect_outliers` Module Documentation
+# `detect_outliers` Module
 
 These functions help detect signal files that contain outliers. This helps for workflows involving batch processing of files, where it might be harder to determine if there are any patterns, or specific files that need additional filters applied.
 
