+++ EMGFlow-Package/tests/test_extract_features.py
@@ -1,280 +1,0 @@
-import unittest
-import importlib
-import os
-import shutil
-
-# Load EMGFlow from local files
-#filePath = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src', 'EMGFlow', '__init__.py'))
-#spec = importlib.util.spec_from_file_location("EMGFlow", filePath)
-#EMGFlow = importlib.util.module_from_spec(spec)
-#spec.loader.exec_module(EMGFlow)
-
-import src.EMGFlow as EMGFlow
-
-import numpy as np
-
-class TestSimple(unittest.TestCase):
-
-#
-# =============================================================================
-#
-    
-    def setUp(self):
-        pass
-
-#
-# =============================================================================
-#
-
-
-    def test_calc_iemg(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-        Signal = EMGFlow.read_file_type(filePath, 'csv')
-        IEMG = EMGFlow.calc_iemg(Signal, 'EMG_zyg', 2000)
-        self.assertIsInstance(IEMG, float)
-    
-    def test_calc_mav(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-        Signal = EMGFlow.read_file_type(filePath, 'csv')
-        MAV = EMGFlow.calc_mav(Signal, 'EMG_zyg')
-        self.assertIsInstance(MAV, float)
-    
-    def test_calc_mmav1(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-        Signal = EMGFlow.read_file_type(filePath, 'csv')
-        MMAV1 = EMGFlow.calc_mmav1(Signal, 'EMG_zyg')
-        self.assertIsInstance(MMAV1, float)
-    
-    def test_calc_mmav2(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-        Signal = EMGFlow.read_file_type(filePath, 'csv')
-        MMAV2 = EMGFlow.calc_mmav2(Signal, 'EMG_zyg')
-        self.assertIsInstance(MMAV2, float)
-    
-    def test_calc_ssi(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-        Signal = EMGFlow.read_file_type(filePath, 'csv')
-        SSI = EMGFlow.calc_ssi(Signal, 'EMG_zyg', 2000)
-        self.assertIsInstance(SSI, float)
-    
-    def test_calc_var(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-        Signal = EMGFlow.read_file_type(filePath, 'csv')
-        VAR = EMGFlow.calc_var(Signal, 'EMG_zyg')
-        self.assertIsInstance(VAR, float)
-    
-    def test_calc_vorder(self):
[…]
