+++ EMGFlow-Package/tests/test_access_files.py
@@ -1,15 +1,8 @@
 import unittest
-import importlib
 import os
 import shutil
 
-# Load EMGFlow from local files
-#filePath = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src', 'EMGFlow', '__init__.py'))
-#spec = importlib.util.spec_from_file_location("EMGFlow", filePath)
-#EMGFlow = importlib.util.module_from_spec(spec)
-#spec.loader.exec_module(EMGFlow)
-
-import src.EMGFlow as EMGFlow
+from src.EMGFlow.access_files import *
 
 import pandas as pd
 
@@ -27,12 +20,12 @@ class TestSimple(unittest.TestCase):
 #
     
     def test_make_paths(self):
-        pathNames = EMGFlow.make_paths()
+        pathNames = make_paths()
         self.assertIsInstance(pathNames, dict)
         self.assertTrue(os.path.isdir('Data'))
     
     def test_make_sample_data(self):
-        pathNames = EMGFlow.make_paths()
+        pathNames = make_paths()
         EMGFlow.make_sample_data(pathNames)
         self.assertTrue(os.path.exists(os.path.join('Data', 'Raw', '01', 'sample_data_01.csv')))
         self.assertTrue(os.path.exists(os.path.join('Data', 'Raw', '01', 'sample_data_02.csv')))
@@ -40,28 +33,28 @@ class TestSimple(unittest.TestCase):
         self.assertTrue(os.path.exists(os.path.join('Data', 'Raw', '02', 'sample_data_04.csv')))
     
     def test_read_file_type(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        df = EMGFlow.read_file_type(os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv'), 'csv')
+        pathNames = make_paths()
+        make_sample_data(pathNames)
+        df = read_file_type(os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv'), 'csv')
         self.assertIsInstance(df, pd.DataFrame)
     
     def test_map_files(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filedirs = EMGFlow.map_files(pathNames['Raw'])
+        pathNames = make_paths()
+        make_sample_data(pathNames)
+        filedirs = map_files(pathNames['Raw'])
         self.assertIsInstance(filedirs, dict)
     
     def test_convert_map_files(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filedirs = EMGFlow.convert_map_files(pathNames['Raw'])
-        filedirs = EMGFlow.convert_map_files(filedirs)
+        pathNames = make_paths()
+        make_sample_data(pathNames)
+        filedirs = convert_map_files(pathNames['Raw'])
+        filedirs = convert_map_files(filedirs)
         self.assertIsInstance(filedirs, dict)
     
     def test_map_files_fuse(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        pathDF = EMGFlow.map_files_fuse([pathNames, pathNames], ['test1', 'test2'])
+        pathNames = make_paths()
+        make_sample_data(pathNames)
+        pathDF = map_files_fuse([pathNames, pathNames], ['test1', 'test2'])
         self.assertIsInstance(pathDF, pd.DataFrame)
     
 #
