+++ EMGFlow-Package/setup.cfg
@@ -1,0 +1,29 @@
+[metadata]
+name = EMGFlow
+description = The open workflow for EMG signal processing and feature extraction.
+author = William Conley
+license = GPL
+license_file = LICENSE
+platforms = unix, linux, osx, cygwin, win32
+classifiers = 
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+
+[options]
+packages =
+    EMGFlow
+install_requires =
+    pandas
+    numpy
+    scipy
+    matplotlib
+    tqdm
+    shiny
+    nest-asyncio
+python_requires = >=3.9
+package_dir =
+    =src
+zip_safe = no
+
+[flake8]
+max-line-length = 160

[…]
