### 2025-05-28 dd0cd12
Updated file structure
Reconfigured file structure of GitHub to be more aligned with standards
