+++ requirements.txt
@@ -5,3 +5,4 @@ matplotlib==3.9.0
 tqdm==4.66.4
 shiny==0.9.0
 nest-asyncio==1.6.0
+importlib-resources==6.5.2
