+++ .github/workflows/tests.yml
@@ -1,112 +1,0 @@
-name: Tests
-
-on:
-  - push
-  - pull_request
-
-defaults:
-  run:
-    shell: bash
-
-jobs:
-  test:
-    name: Python ${{ matrix.python-version }} on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu-latest, windows-latest]
-        python-version: ['3.9', '3.10']
-    
-    steps:
-    - name: Check out the repo
-      uses: actions/checkout@v4
-
-    - name: Set up Python
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install tox tox-gh-actions
-
-    - name: Run tox for ${{ matrix.python-version }}
-      run: |
-        python -m tox
-
-    - name: List files
-      run: ls -al .coverage*
-
-    - name: Upload coverage data
-      uses: actions/upload-artifact@v4
-      with:
-        name: covdata-${{ matrix.os }}-${{ matrix.python-version }}
-        path: .coverage.*
-        include-hidden-files: true
-
-  coverage:
-    name: Coverage
-    needs: test
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out the repo
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: '3.10'
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install tox tox-gh-actions
-
-      - name: Download coverage data (Windows 3.9)
-        uses: actions/download-artifact@v4
-        with:
-          name: covdata-windows-latest-3.9
-          path: ./.coverage.windows_3.9/*
-      
-      - name: Download coverage data (Windows 3.10)
-        uses: actions/download-artifact@v4
-        with:
-          name: covdata-windows-latest-3.10
-          path: ./.coverage.windows_3.10/*
-      
-      - name: Download coverage data (Ubuntu 3.9)
-        uses: actions/download-artifact@v4
-        with:
-          name: covdata-ubuntu-latest-3.9
-          path: ./.coverage.ubuntu_3.9/*
-      
-      - name: Download coverage data (Ubuntu 3.10)
-        uses: actions/download-artifact@v4
-        with:
-          name: covdata-ubuntu-latest-3.10
-          path: ./.coverage.ubuntu_3.10/*
-
-      - name: List files
-        run: ls -a
-
-      - name: Combine
-        run: |
-          python -m tox -e coverage
-          export TOTAL=$(python -c "import json;print(json.load(open('coverage.json'))['totals']['percent_covered_display'])")
-          echo "total=$TOTAL" >> $GITHUB_ENV
-          echo "### Total coverage: ${TOTAL}%" >> $GITHUB_STEP_SUMMARY
-      
-      - name: Make badge
-        uses: schneegans/dynamic-badges-action@v1.4.0
-        with:
-          auth: ${{ secrets.GIST_SECRET }}
-          gistID: da00e624c5ba768584057a6ea7c53860
-          filename: covbadge.json
-          label: Coverage
-          message: ${{ env.total }}%
-          minColorRange: 50
-          maxColorRange: 90
-          valColorRange: ${{ env.total }}
