+++ tox.ini
@@ -1,23 +1,0 @@
-[tox]
-envlist = py39, py310, coverage
-
-[testenv]
-setenv =
-    PYTHONPATH = {toxinidir}
-deps =
-    -r {toxinidir}/EMGFlow-Package/requirements_dev.txt
-commands =
-    python -m coverage run -p -m unittest discover -s {toxinidir}/EMGFlow-Package/
-
-[testenv:coverage]
-basepython = python3.10
-allowlist_externals = /bin/sh
-commands =
-    /bin/sh -c 'for f in .coverage.*; do coverage combine $f; done'
-    python -m coverage report -m --skip-covered -i
-    python -m coverage json -i
-    
-[gh-actions]
-python =
-    3.9: py39
-    3.10: py310
