+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,12 @@
 name: Draft PDF
-on: [push]
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
 
 jobs:
   paper:
