### 2025-05-28 0a23731
Added automated coverage
Updated YML and README to automatically update coverage badge
