+++ .github/workflows/run-tests.yml
@@ -28,15 +28,51 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .[test]
+          pip install .[test] coverage codecov
       
-      - name: Run tests with unittest
+      - name: Run tests with coverage
         run: |
-          python -m unittest discover -s tests -v
+          coverage run -m unittest discover -s tests -v
+          coverage xml
+          coverage html
+          coverage report > coverage.txt
+      
+      - name: Extract coverage percentage
+        run: |
+          COVERAGE=$(grep 'TOTAL' coverage.txt | awk '{print $NF}' | sed 's/%//')
+          echo "COVERAGE_PERCENT=$COVERAGE" >> $GITHUB_ENV
+      
+      - name: Generate shields.io badge SVG
+        run: |
+          COVERAGE=${{ env.COVERAGE_PERCENT }}
+          COLOR="red"
+          if [ "$COVERAGE" -ge 90 ]; then COLOR="brightgreen"; fi
+          if [ "$COVERAGE" -ge 80 ] && [ "$COVERAGE" -lt 90 ]; then COLOR="green"; fi
+          if [ "$COVERAGE" -ge 70 ] && [ "$COVERAGE" -lt 80 ]; then COLOR="yellowgreen"; fi
+          if [ "$COVERAGE" -ge 60 ] && [ "$COVERAGE" -lt 70 ]; then COLOR="yellow"; fi
+          if [ "$COVERAGE" -ge 50 ] && [ "$COVERAGE" -lt 60 ]; then COLOR="orange"; fi
+          mkdir -p badges
+          curl "https://img.shields.io/badge/coverage-$COVERAGE%25-$COLOR.svg" -o badges/coverage.svg
+      
+      - name: Commit badge SVG
+        if: matrix.python-version == '3.12' && github.event_name == 'push' && github.ref == 'refs/heads/main'
+        run: |
+          git config user.name "GitHub Actions"
+          git config user.email "actions@github.com"
+          git add badges/coverage.svg
+          git commit -m "Update coverage badge SVG [ci skip]" || echo "No changes to commit"
+          git push
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       
       - name: Upload test results
         if: always()
         uses: actions/upload-artifact@v4
         with:
           name: test-results-${{ matrix.python-version }}
-          path: test-results/
+          path: |
+            test-results/
+            htmlcov/
+            coverage.xml
+            coverage.txt
+            badges/coverage.svg
