+++ .github/workflows/run-tests.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          token: ${{ secrets.GITHUB_TOKEN }}
+          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions
 
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
@@ -30,14 +30,16 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .[test] coverage codecov
+          pip install .[test] coverage
+          # Ensure coverage is included in dependencies
       
       - name: Run tests with coverage
         run: |
           coverage run -m unittest discover -s tests -v
           coverage xml -o coverage-${{ matrix.python-version }}.xml
           coverage html -d htmlcov-${{ matrix.python-version }}
-          coverage report > coverage-${{ matrix-python-version }}.txt
+          coverage report > coverage-${{ matrix.python-version }}.txt
+        # Generate coverage reports with version-specific names
       
       - name: Upload coverage data
         uses: actions/upload-artifact@v4
@@ -47,18 +49,19 @@ jobs:
             coverage-${{ matrix.python-version }}.xml
             coverage-${{ matrix.python-version }}.txt
             htmlcov-${{ matrix.python-version }}/
-          
+        # Upload coverage data for aggregation
+
   aggregate-coverage:
     runs-on: ubuntu-latest
-    needs: test
+    needs: test # Run after all test jobs complete
     if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-
+    
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-      
+          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions
+
       - name: Set up Python 3.12
         uses: actions/setup-python@v5
         with:
@@ -74,19 +77,21 @@ jobs:
         with:
           pattern: coverage-data-*
           path: coverage-data
-          merge-multiple: true
-
+          merge-multiple: true # Merge artifacts from all matrix jobs
+      
       - name: Combine coverage data
         run: |
           coverage combine coverage-data/coverage-*.xml
           coverage xml -o coverage-combined.xml
           coverage html -d htmlcov-combined
           coverage report > coverage-combined.txt
-
+        # Combine coverage data and generate reports
+      
       - name: Extract combined coverage percentage
         run: |
           COVERAGE=$(grep 'TOTAL' coverage-combined.txt | awk '{print $NF}' | sed 's/%//')
           echo "COVERAGE_PERCENT=$COVERAGE" >> $GITHUB_ENV
+        # Extract the combined coverage percentage
       
       - name: Generate shields.io badge SVG
         run: |
@@ -99,9 +104,9 @@ jobs:
           if [ "$COVERAGE" -ge 50 ] && [ "$COVERAGE" -lt 60 ]; then COLOR="orange"; fi
           mkdir -p badges
           curl "https://img.shields.io/badge/coverage-$COVERAGE%25-$COLOR.svg" -o badges/coverage.svg
+        # Download SVG badge from shields.io
       
       - name: Commit badge SVG
-        if: matrix.python-version == '3.12' && github.event_name == 'push' && github.ref == 'refs/heads/main'
         run: |
           git config user.name "GitHub Actions"
           git config user.email "actions@github.com"
@@ -110,6 +115,7 @@ jobs:
           git push
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        # Commit with [ci skip] to prevent infinite loop
       
[…]
