### 2025-05-28 c11b916
Updated coverage merging
Now generates raw coverage files, renames them to prevent conflictions
