+++ .github/workflows/run-tests.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions or use BADGE_TOKEN
+          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions
 
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
@@ -30,29 +30,26 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e .[test] coverage
-          # Install package in editable mode to ensure imports work
+          pip install .[test] coverage
+          # Ensure coverage is included in dependencies
       
       - name: Run tests with coverage
         run: |
-          export PYTHONPATH=$PYTHONPATH:$GITHUB_WORKSPACE
-          coverage run --source=emgflow -m unittest discover -s tests -v
-          mv .coverage .coverage.${{ matrix.python-version }}
+          coverage run -m unittest discover -s tests -v
           coverage xml -o coverage-${{ matrix.python-version }}.xml
           coverage html -d htmlcov-${{ matrix.python-version }}
           coverage report > coverage-${{ matrix.python-version }}.txt
-        # Specify source, set PYTHONPATH, and generate reports
+        # Generate coverage reports with version-specific names
       
       - name: Upload coverage data
         uses: actions/upload-artifact@v4
         with:
           name: coverage-data-${{ matrix.python-version }}
           path: |
-            .coverage.${{ matrix.python-version }}
             coverage-${{ matrix.python-version }}.xml
             coverage-${{ matrix.python-version }}.txt
             htmlcov-${{ matrix.python-version }}/
-        # Upload raw .coverage file and reports
+        # Upload coverage data for aggregation
 
   aggregate-coverage:
     runs-on: ubuntu-latest
@@ -63,7 +60,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions or use BADGE_TOKEN
+          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions
 
       - name: Set up Python 3.12
         uses: actions/setup-python@v5
@@ -84,12 +81,11 @@ jobs:
       
       - name: Combine coverage data
         run: |
-          mv coverage-data/.coverage.* .
-          coverage combine
+          coverage combine coverage-data/coverage-*.xml
           coverage xml -o coverage-combined.xml
           coverage html -d htmlcov-combined
           coverage report > coverage-combined.txt
-        # Move .coverage files to root, combine, and generate reports
+        # Combine coverage data and generate reports
       
       - name: Extract combined coverage percentage
         run: |
@@ -118,7 +114,7 @@ jobs:
           git commit -m "Update coverage badge SVG [ci skip]" || echo "No changes to commit"
           git push
         env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # Or BADGE_TOKEN
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         # Commit with [ci skip] to prevent infinite loop
       
       - name: Upload combined results
