+++ .github/workflows/run-tests.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions
+          token: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
@@ -31,7 +31,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install .[test] coverage
-          # Ensure coverage is included in dependencies
       
       - name: Run tests with coverage
         run: |
@@ -39,28 +38,30 @@ jobs:
           coverage xml -o coverage-${{ matrix.python-version }}.xml
           coverage html -d htmlcov-${{ matrix.python-version }}
           coverage report > coverage-${{ matrix.python-version }}.txt
-        # Generate coverage reports with version-specific names
+          mv .coverage .coverage-${{ matrix.python-version }}  # Rename .coverage file
+        # Generate coverage reports and preserve .coverage file
       
       - name: Upload coverage data
         uses: actions/upload-artifact@v4
         with:
           name: coverage-data-${{ matrix.python-version }}
           path: |
+            .coverage-${{ matrix.python-version }}
             coverage-${{ matrix.python-version }}.xml
             coverage-${{ matrix.python-version }}.txt
             htmlcov-${{ matrix.python-version }}/
-        # Upload coverage data for aggregation
+        # Upload .coverage file along with other reports
 
   aggregate-coverage:
     runs-on: ubuntu-latest
-    needs: test # Run after all test jobs complete
+    needs: test
     if: github.event_name == 'push' && github.ref == 'refs/heads/main'
     
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          token: ${{ secrets.GITHUB_TOKEN }} # Ensure write permissions
+          token: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Set up Python 3.12
         uses: actions/setup-python@v5
@@ -77,21 +78,26 @@ jobs:
         with:
           pattern: coverage-data-*
           path: coverage-data
-          merge-multiple: true # Merge artifacts from all matrix jobs
+          merge-multiple: true
+      
+      - name: Move coverage files to correct location
+        run: |
+          mkdir -p .cov-data
+          find coverage-data -type f -name '.coverage-*' -exec mv {} .cov-data/ \;
+        # Move .coverage files to a single directory for combining
       
       - name: Combine coverage data
         run: |
-          coverage combine coverage-data/coverage-*.xml
+          coverage combine .cov-data/.coverage-*
           coverage xml -o coverage-combined.xml
           coverage html -d htmlcov-combined
           coverage report > coverage-combined.txt
-        # Combine coverage data and generate reports
+        # Combine .coverage files and generate reports
       
       - name: Extract combined coverage percentage
         run: |
           COVERAGE=$(grep 'TOTAL' coverage-combined.txt | awk '{print $NF}' | sed 's/%//')
           echo "COVERAGE_PERCENT=$COVERAGE" >> $GITHUB_ENV
-        # Extract the combined coverage percentage
       
       - name: Generate shields.io badge SVG
         run: |
@@ -104,7 +110,6 @@ jobs:
           if [ "$COVERAGE" -ge 50 ] && [ "$COVERAGE" -lt 60 ]; then COLOR="orange"; fi
           mkdir -p badges
           curl "https://img.shields.io/badge/coverage-$COVERAGE%25-$COLOR.svg" -o badges/coverage.svg
-        # Download SVG badge from shields.io
       
       - name: Commit badge SVG
[…]
