### 2025-05-28 31a0f95
Update run-tests.yml
Added debugging steps
