+++ .github/workflows/run-tests.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .[test] coverage
+          pip install .[test] coverage==7.6.1  # Pin coverage version for consistency
       
       - name: Run tests with coverage
         run: |
@@ -38,8 +38,7 @@ jobs:
           coverage xml -o coverage-${{ matrix.python-version }}.xml
           coverage html -d htmlcov-${{ matrix.python-version }}
           coverage report > coverage-${{ matrix.python-version }}.txt
-          mv .coverage .coverage-${{ matrix.python-version }}  # Rename .coverage file
-        # Generate coverage reports and preserve .coverage file
+          mv .coverage .coverage-${{ matrix.python-version }}
       
       - name: Upload coverage data
         uses: actions/upload-artifact@v4
@@ -50,7 +49,6 @@ jobs:
             coverage-${{ matrix.python-version }}.xml
             coverage-${{ matrix.python-version }}.txt
             htmlcov-${{ matrix.python-version }}/
-        # Upload .coverage file along with other reports
 
   aggregate-coverage:
     runs-on: ubuntu-latest
@@ -71,7 +69,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install coverage
+          pip install coverage==7.6.1  # Pin coverage version for consistency
       
       - name: Download coverage data
         uses: actions/download-artifact@v4
@@ -80,19 +78,26 @@ jobs:
           path: coverage-data
           merge-multiple: true
       
+      - name: Debug downloaded files
+        run: |
+          ls -R coverage-data
+          find coverage-data -type f -name '.coverage-*'
+        # Inspect the directory structure to debug file locations
+      
       - name: Move coverage files to correct location
         run: |
           mkdir -p .cov-data
           find coverage-data -type f -name '.coverage-*' -exec mv {} .cov-data/ \;
-        # Move .coverage files to a single directory for combining
+          ls -l .cov-data/  # Debug to confirm files are moved
+        # Move .coverage files to .cov-data/
       
       - name: Combine coverage data
         run: |
-          coverage combine .cov-data/.coverage-*
+          coverage combine .cov-data/ || echo "No coverage files found or combine failed"
           coverage xml -o coverage-combined.xml
           coverage html -d htmlcov-combined
           coverage report > coverage-combined.txt
-        # Combine .coverage files and generate reports
+        # Combine .coverage files from .cov-data/ directory
       
       - name: Extract combined coverage percentage
         run: |
