### 2025-05-28 bbebb3f
Update run-tests.yml
Added additional debugging steps
