+++ .github/workflows/run-tests.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .[test] coverage==7.6.1  # Pin coverage version for consistency
+          pip install .[test] coverage==7.6.1
       
       - name: Run tests with coverage
         run: |
@@ -38,7 +38,12 @@ jobs:
           coverage xml -o coverage-${{ matrix.python-version }}.xml
           coverage html -d htmlcov-${{ matrix.python-version }}
           coverage report > coverage-${{ matrix.python-version }}.txt
-          mv .coverage .coverage-${{ matrix.python-version }}
+          mv .coverage .coverage-${{ matrix.python-version }} || echo "No .coverage file generated"
+      
+      - name: Debug coverage file
+        run: |
+          ls -l .coverage-${{ matrix.python-version }} || echo "Coverage file not found"
+        # Confirm the .coverage file exists before upload
       
       - name: Upload coverage data
         uses: actions/upload-artifact@v4
@@ -49,6 +54,8 @@ jobs:
             coverage-${{ matrix.python-version }}.xml
             coverage-${{ matrix.python-version }}.txt
             htmlcov-${{ matrix.python-version }}/
+          retention-days: 1
+        # Reduced retention for artifacts to save space
 
   aggregate-coverage:
     runs-on: ubuntu-latest
@@ -69,7 +76,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install coverage==7.6.1  # Pin coverage version for consistency
+          pip install coverage==7.6.1
       
       - name: Download coverage data
         uses: actions/download-artifact@v4
@@ -81,23 +88,25 @@ jobs:
       - name: Debug downloaded files
         run: |
           ls -R coverage-data
-          find coverage-data -type f -name '.coverage-*'
-        # Inspect the directory structure to debug file locations
+          find coverage-data -type f -name '.coverage-*' || echo "No .coverage-* files found"
       
       - name: Move coverage files to correct location
         run: |
           mkdir -p .cov-data
-          find coverage-data -type f -name '.coverage-*' -exec mv {} .cov-data/ \;
-          ls -l .cov-data/  # Debug to confirm files are moved
-        # Move .coverage files to .cov-data/
+          find coverage-data -type f -name '.coverage-*' -exec mv {} .cov-data/ \; || echo "No .coverage files to move"
+          ls -l .cov-data/ || echo "No files in .cov-data"
       
       - name: Combine coverage data
         run: |
-          coverage combine .cov-data/ || echo "No coverage files found or combine failed"
-          coverage xml -o coverage-combined.xml
-          coverage html -d htmlcov-combined
-          coverage report > coverage-combined.txt
-        # Combine .coverage files from .cov-data/ directory
+          if [ -n "$(ls -A .cov-data)" ]; then
+            coverage combine .cov-data/
+            coverage xml -o coverage-combined.xml
+            coverage html -d htmlcov-combined
+            coverage report > coverage-combined.txt
+          else
+            echo "No coverage files found to combine"
+            exit 1
+          fi
       
       - name: Extract combined coverage percentage
         run: |
@@ -136,3 +145,4 @@ jobs:
             coverage-combined.txt
             htmlcov-combined/
             badges/coverage.svg
+          retention-days: 1
