### 2025-05-28 8409d0f
Update run-tests.yml
Restructured yml workflow
