+++ .github/workflows/run-tests.yml
@@ -1,92 +1,151 @@
-aggregate-coverage:
-  runs-on: ubuntu-latest
-  needs: test
-  if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-  
-  
-  steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-      with:
-        token: ${{ secrets.GITHUB_TOKEN }}
+name: Run Unit Tests
 
-    - name: Set up Python 3.12
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.12"
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
     
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install coverage==7.6.1
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install .[test] coverage==7.6.1
+      
+      - name: Run tests with coverage
+        run: |
+          coverage run -m unittest discover -s tests -v
+          if [ ! -f .coverage ]; then
+            echo "Error: No .coverage file generated"
+            exit 1
+          fi
+          mv .coverage .coverage-${{ matrix.python-version }}
+          coverage xml -o coverage-${{ matrix.python-version }}.xml
+          coverage html -d htmlcov-${{ matrix.python-version }}
+          coverage report > coverage-${{ matrix.python-version }}.txt
+      
+      - name: Debug coverage file
+        run: |
+          ls -l .coverage-${{ matrix.python-version }} || echo "Coverage file not found"
+      
+      - name: Upload coverage data
+        uses: actions/upload-artifact@v4
+        with:
+          name: coverage-data-${{ matrix.python-version }}
+          path: |
+            .coverage-${{ matrix.python-version }}
+            coverage-${{ matrix.python-version }}.xml
+            coverage-${{ matrix.python-version }}.txt
+            htmlcov-${{ matrix.python-version }}/
+          retention-days: 1
+
+  aggregate-coverage:
+    runs-on: ubuntu-latest
+    needs: test
+    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
     
-    - name: Download coverage data
-      uses: actions/download-artifact@v4
-      with:
-        pattern: coverage-data-*
-        path: coverage-data
-        merge-multiple: true
-    
-    - name: Debug downloaded files
-      run: |
-        echo "Directory structure of coverage-data:"
-        find coverage-data -type f -ls
-        echo "Looking for .coverage-* files:"
-        find coverage-data -type f -name '.coverage-*' || echo "No .coverage-* files found"
-    
-    - name: Move coverage files to correct location
-      run: |
-        mkdir -p .cov-data
-        find coverage-data -type f -name '.coverage-*' -exec mv {} .cov-data/ \; || { echo "No .coverage files found"; exit 1; }
-        ls -l .cov-data/ || echo "No files in .cov-data"
-    
-    - name: Combine coverage data
-      run: |
[…]
