### 2025-05-29 d274127
Update run-tests.yml
Added additional debugging
