### 2025-05-29 c540803
Update run-tests.yml
Modified debugging steps
