### 2025-05-29 7055a6d
Update run-tests.yml
Updated debugging
