+++ .github/workflows/run-tests.yml
@@ -15,7 +15,7 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
-    
+
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
@@ -26,30 +26,32 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-      
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install .[test] coverage==7.6.1
-      
+
       - name: Run tests with coverage
         run: |
           mkdir -p coverage-data/${{ matrix.python-version }}
-          coverage run -m unittest discover -s tests -v || { echo "Unit tests failed or no tests found"; exit 1; }
+          # Use --source to ensure consistent source paths
+          coverage run --source=. -m unittest discover -s tests -v || { echo "Unit tests failed or no tests found"; exit 1; }
           coverage xml -o coverage-data/${{ matrix.python-version }}/coverage.xml
           coverage html -d coverage-data/${{ matrix.python-version }}/htmlcov
           coverage report > coverage-data/${{ matrix.python-version }}/coverage.txt
-          mv .coverage coverage-data/${{ matrix.python-version }}/.coverage || { echo "No .coverage file generated for Python ${{ matrix.python-version }}"; exit 1; }
-      
+          mv .coverage coverage-data/${{ matrix.python-version }}/.coverage.${{ matrix.python-version }} || { echo "No .coverage file generated for Python ${{ matrix.python-version }}"; exit 1; }
+
       - name: Debug coverage file
         run: |
-          ls -l coverage-data/${{ matrix.python-version }}/.coverage || echo "Coverage file not found for Python ${{ matrix.python-version }}"
-          if [ -s coverage-data/${{ matrix.python-version }}/.coverage ]; then
+          ls -l coverage-data/${{ matrix.python-version }}/.coverage.${{ matrix.python-version }} || echo "Coverage file not found for Python ${{ matrix.python-version }}"
+          if [ -s coverage-data/${{ matrix.python-version }}/.coverage.${{ matrix.python-version }} ]; then
             echo "Coverage file exists and is non-empty for Python ${{ matrix.python-version }}"
           else
             echo "Coverage file is missing or empty for Python ${{ matrix.python-version }}"
+            exit 1
           fi
-      
+
       - name: Upload coverage data
         uses: actions/upload-artifact@v4
         with:
@@ -61,7 +63,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: test
     if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-    
+
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
@@ -72,64 +74,61 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.12"
-      
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install coverage==7.6.1
-      
+
       - name: Download coverage data
         uses: actions/download-artifact@v4
         with:
           pattern: coverage-data-*
           path: coverage-data
           merge-multiple: true
-      
+
       - name: Debug downloaded files
         run: |
           ls -R coverage-data
-          find coverage-data -type f -name '.coverage' || echo "No .coverage files found"
-      
+          find coverage-data -type f -name '.coverage.*' || echo "No .coverage files found"
+
       - name: Move coverage files to correct location
         run: |
           mkdir -p .cov-data
[…]
