### 2025-05-29 22d2383
Update run-tests.yml
Updated how coverage files are generated
