### 2025-05-29 e0fff37
Added coverage consistency files
Added a .coveragerc file for coverage consistency

Added coverage dependencies

Updated workflow
