+++ pyproject.toml
@@ -31,12 +31,6 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 
-[project.optional-dependencies]
-test = [
-  "coverage==7.6.1",
-  "unittest"
-]
-
 [project.urls]
 Homepage = "https://github.com/WiIIson/EMGFlow-Python-Package"
 Documentation = "https://wiiison.github.io/EMGFlow-Python-Package/"
