### 2025-05-29 dfdb535
Update run-tests.yml
Simplified aggregate-coverage step
