+++ .github/workflows/run-tests.yml
@@ -58,123 +58,96 @@ jobs:
           path: coverage-data/${{ matrix.python-version }}/
           retention-days: 1
 
-  aggregate-coverage:
-    runs-on: ubuntu-latest
-    needs: test
-    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+aggregate-coverage:
+  runs-on: ubuntu-latest
+  needs: test
+  if: github.event_name == 'push' && github.ref == 'refs/heads/main'
 
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
+  steps:
+    - name: Checkout code
+      uses: actions/checkout@v4
+      with:
+        token: ${{ secrets.GITHUB_TOKEN }}
 
-      - name: Set up Python 3.12
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.12"
+    - name: Set up Python 3.12
+      uses: actions/setup-python@v5
+      with:
+        python-version: "3.12"
 
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install coverage==7.6.1
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install coverage==7.6.1
 
-      - name: Download coverage data for Python 3.9
-        uses: actions/download-artifact@v4
-        with:
-          name: coverage-data-3.9
-          path: coverage-data/3.9
+    - name: Download all coverage data
+      uses: actions/download-artifact@v4
+      with:
+        pattern: coverage-data-*
+        path: coverage-data
+        merge-multiple: true
 
-      - name: Download coverage data for Python 3.10
-        uses: actions/download-artifact@v4
-        with:
-          name: coverage-data-3.10
-          path: coverage-data/3.10
+    - name: Debug downloaded files
+      run: |
+        ls -R coverage-data || echo "No coverage data found"
+        find coverage-data -type f -name '.coverage' || echo "No .coverage files found"
 
-      - name: Download coverage data for Python 3.11
-        uses: actions/download-artifact@v4
-        with:
-          name: coverage-data-3.11
-          path: coverage-data/3.11
-
-      - name: Download coverage data for Python 3.12
-        uses: actions/download-artifact@v4
-        with:
-          name: coverage-data-3.12
-          path: coverage-data/3.12
-
-      - name: Debug downloaded files
-        run: |
-          ls -R coverage-data
-          find coverage-data -type f -name '.coverage' || echo "No .coverage files found"
-
-      - name: Move coverage files to correct location
-        run: |
-          mkdir -p .cov-data
-          find coverage-data -type f -name '.coverage' -exec mv {} .cov-data/ \; || { echo "No .coverage files to move"; exit 1; }
-          ls -l .cov-data/ || echo "No files in .cov-data"
-
-      - name: Validate coverage files
-        run: |
-          if ls .cov-data/.coverage >/dev/null 2>&1; then
-            for file in .cov-data/.coverage; do
-              if [ -s "$file" ]; then
-                echo "Valid coverage file: $file"
-              else
-                echo "Empty or invalid coverage file: $file"
-                exit 1
-              fi
-            done
-          else
-            echo "No coverage files found in .cov-data/"
-            exit 1
-          fi
-
-      - name: Combine coverage data
-        run: |
[…]
