+++ .github/workflows/run-tests.yml
@@ -58,96 +58,96 @@ jobs:
           path: coverage-data/${{ matrix.python-version }}/
           retention-days: 1
 
-aggregate-coverage:
-  runs-on: ubuntu-latest
-  needs: test
-  if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+  aggregate-coverage:
+    runs-on: ubuntu-latest
+    needs: test
+    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
 
-  steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-      with:
-        token: ${{ secrets.GITHUB_TOKEN }}
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
 
-    - name: Set up Python 3.12
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.12"
+      - name: Set up Python 3.12
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
 
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install coverage==7.6.1
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install coverage==7.6.1
 
-    - name: Download all coverage data
-      uses: actions/download-artifact@v4
-      with:
-        pattern: coverage-data-*
-        path: coverage-data
-        merge-multiple: true
+      - name: Download all coverage data
+        uses: actions/download-artifact@v4
+        with:
+          pattern: coverage-data-*
+          path: coverage-data
+          merge-multiple: true
 
-    - name: Debug downloaded files
-      run: |
-        ls -R coverage-data || echo "No coverage data found"
-        find coverage-data -type f -name '.coverage' || echo "No .coverage files found"
+      - name: Debug downloaded files
+        run: |
+          ls -R coverage-data || echo "No coverage data found"
+          find coverage-data -type f -name '.coverage' || echo "No .coverage files found"
 
-    - name: Combine coverage data
-      run: |
-        mkdir -p .cov-data
-        find coverage-data -type f -name '.coverage' -exec cp {} .cov-data/ \;
-        if ls .cov-data/.coverage* >/dev/null 2>&1; then
-          cd .cov-data
-          coverage combine --debug=combine || { echo "Coverage combine failed"; exit 1; }
-          coverage xml -o ../coverage-combined.xml
-          coverage html -d ../htmlcov-combined
-          coverage report > ../coverage-combined.txt
-        else
-          echo "No valid .coverage files found in .cov-data/"
-          exit 1
-        fi
-      env:
-        COVERAGE_PATH: ${{ github.workspace }}
+      - name: Combine coverage data
+        run: |
+          mkdir -p .cov-data
+          find coverage-data -type f -name '.coverage' -exec cp {} .cov-data/ \;
+          if ls .cov-data/.coverage* >/dev/null 2>&1; then
+            cd .cov-data
+            coverage combine --debug=combine || { echo "Coverage combine failed"; exit 1; }
+            coverage xml -o ../coverage-combined.xml
+            coverage html -d ../htmlcov-combined
+            coverage report > ../coverage-combined.txt
+          else
+            echo "No valid .coverage files found in .cov-data/"
+            exit 1
+          fi
+        env:
[…]
