### 2025-05-29 ebfd6ab
Update extract-features.md
Updated sidebar display
