### 2025-05-29 d1d8392
Update run-tests.yml
Reverted to version that worked
