### 2025-05-29 5a207c3
Update run-tests.yml
Added average calculator
