### 2025-05-29 402b746
Update run-tests.yml
Fixed average calculation
