### 2025-05-29 1284ea1
Update run-tests.yml
Added debugging step
