+++ .github/workflows/run-tests.yml
@@ -10,10 +10,11 @@ on:
 
 jobs:
   test:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
 
     strategy:
       matrix:
+        os: [ubuntu-latest, windows-latest]
         python-version: ["3.9", "3.10", "3.11", "3.12"]
     
     steps:
@@ -35,18 +36,20 @@ jobs:
           coverage run -m unittest discover -s tests -v
           coverage xml
           coverage html
-          coverage report > coverage-${{ matrix.python-version }}.txt
+          coverage report > coverage-${{ matrix.os }}-${{ matrix.python-version }}.txt
+        shell: bash # Ensure consistent shell for Windows and Ubuntu
       
       - name: Store coverage percentage
         run: |
-          COVERAGE=$(grep 'TOTAL' coverage-${{ matrix.python-version }}.txt | awk '{print $NF}' | sed 's/%//')
-          echo "COVERAGE_${{ matrix.python-version }}=$COVERAGE" >> $GITHUB_ENV
+          COVERAGE=$(grep 'TOTAL' coverage-${{ matrix.os }}-${{ matrix.python-version }}.txt | awk '{print $NF}' | sed 's/%//')
+          echo "COVERAGE_${{ matrix.os }}_${{ matrix.python-version }}=$COVERAGE" >> $GITHUB_ENV
+        shell: bash
       
       - name: Upload coverage data
         uses: actions/upload-artifact@v4
         with:
-          name: coverage-data-${{ matrix.python-version }}
-          path: coverage-${{ matrix.python-version }}.txt
+          name: coverage-data-${{ matrix.os }}-${{ matrix.python-version }}
+          path: coverage-${{ matrix.os }}-${{ matrix.python-version }}.txt
 
   generate-badge:
     needs: test
@@ -83,6 +86,7 @@ jobs:
             echo "No coverage files found"
             exit 1
           fi
+        shell: bash
 
       - name: Generate shields.io badge SVG
         run: |
@@ -104,10 +108,4 @@ jobs:
           git commit -m "Update coverage badge SVG [ci skip]" || echo "No changes to commit"
           git push
         env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Upload test results
-        uses: actions/upload-artifact@v4
-        with:
-          name: coverage-badge
-          path: badges/coverage.svg
+          GITHUB_TOKEN: ${{ secrets
