### 2025-05-29 9322408
Update run-tests.yml
Added missing code
