### 2025-05-30 07ecadc
Update run-tests.yml
Fixed calculation on windows
