### 2025-05-30 7daba8d
Re-Added additional tests
Added additional tests for modules
