### 2025-06-03 d8aad19
Updated handle_outliers module
Changed name from "detect_outliers" to "handle_outliers"

Renamed "detect_outliers" function to "detect_spectral_outliers"
