+++ tests/test_handle_outliers.py
@@ -20,7 +20,7 @@ class TestSimple(unittest.TestCase):
     def test_detect_outliers(self):
         pathNames = EMGFlow.make_paths()
         EMGFlow.make_sample_data(pathNames)
-        outliers = EMGFlow.detect_outliers(pathNames['Raw'], 2000, 2, window_size=15)
+        outliers = EMGFlow.detect_spectral_outliers(pathNames['Raw'], 2000, 2, window_size=15)
         self.assertIsInstance(outliers, dict)
 
 #
