### 2025-06-05 15c87b1
Updated handle_outliers module
Renamed "detect_outliers" module to "handle_outliers"

Renamed "detect_outliers" function to "detect_spectral_outliers"
