### 2025-06-05 3f4a92e
Added x-scale control
Added a slider to control the zoom on the x-scale in "plot_dashboard"
