### 2025-06-06 0711f17
Updated plot_dashboard
min_x now updates when the plot is changed (not all plots start at 0)
