### 2025-06-06 a890337
Added NaN Exceptions
Added exceptions for dataframes with NaN values.
