+++ src/EMGFlow/preprocess_signals.py
@@ -123,6 +123,8 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
     Exception
         An exception is raised if 'col' is not a column of 'Signal'.
     Exception
+        An exception is raised if 'col' contains NaN values.
+    Exception
         An exception is raised if 'sampling_rate' is less or equal to 0.
     Exception
         An exception is raised if a Hz value in 'notch_vals' is greater than
@@ -135,9 +137,15 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
 
     """
 
+    # An exception is raised if 'col' is not a column of 'Signal'.
     if col not in list(Signal.columns.values):
         raise Exception("Column " + str(col) + " not in Signal")
     
+    # An exception is raised if 'col' contains NaN values.
+    if Signal[col].isnull().values.any():
+        raise Exception("NaN values found in column: " + str(col))
+    
+    # An exception is raised if 'sampling_rate' is less or equal to 0.
     if sampling_rate <= 0:
         raise Exception("Sampling rate must be greater or equal to 0")
 
@@ -174,9 +182,10 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
         """
         
         Signal = Signal.copy()
-        
         (Hz, Q) = notch
         
+        # An exception is raised if a Hz value in 'notch_vals' is greater than
+        # sampling_rate/2 or less than 0.
         if Hz > sampling_rate / 2 or Hz < 0:
             raise Exception("Notch filter frequency must be between 0 and " + str(sampling_rate / 2) + " (sampling_rate/2)")
         
@@ -241,6 +250,9 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
     Exception
         An exception is raised if any column in 'cols' is not found in any of
         the Signal files read.
+    Exception
+        An exception is raised if any column in 'cols' in any of the Signal
+        files read contain NaN values.
     Exception
         An exception is raised if 'sampling_rate' is less or equal to 0.
     Exception
@@ -342,12 +354,14 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high):
     Exception
         An exception is raised if 'col' is not a column of 'Signal'.
     Exception
-        An exception is raised if 'sampling_rate' is less or equal to 0.
+        An exception is raised if 'col' contains NaN values.
     Exception
-        An exception is raised if 'high' is not higher than 'low'.
+        An exception is raised if 'sampling_rate' is less or equal to 0.
     Exception
         An exception is raised if 'high' or 'low' are higher than 1/2 of
         'sampling_rate'.
+    Exception
+        An exception is raised if 'high' is not higher than 'low'.
 
     Returns
     -------
@@ -356,15 +370,24 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high):
 
     """
     
+    # An exception is raised if 'col' is not a column of 'Signal'.
     if col not in list(Signal.columns.values):
         raise Exception("Column " + str(col) + " not in Signal.")
     
+    # An exception is raised if 'col' contains NaN values.
+    if Signal[col].isnull().values.any():
+        raise Exception("NaN values found in column: " + str(col))
+    
+    # An exception is raised if 'sampling_rate' is less or equal to 0.
     if sampling_rate <= 0:
         raise Exception("Sampling rate must be greater or equal to 0.")
     
+    # An exception is raised if 'high' or 'low' are higher than 1/2 of
+    # 'sampling_rate'.
     if high > sampling_rate/2 or low > sampling_rate/2:
         raise Exception("'high' and 'low' cannot be greater than 1/2 the sampling rate.")
     
+    # An exception is raised if 'high' is not higher than 'low'.
     if high <= low:
[…]
