### 2025-06-06 7c57f83
Update preprocess_signals.py
Added a NaN exception to "emg_to_psd"
