### 2025-06-06 34bb1f9
Update preprocess-signals.md
Added NaN exception documentation.
