### 2025-06-06 f34e625
Update preprocess_signals.py
Fixed nan check
