### 2025-06-09 b06e676
Added screen_artifacts function
Wrote screen_artifacts function

Updated documentation for smooth_filter_signals to be consistent
