### 2025-06-09 1fcc9ce
Update handle_outliers.py
Added documentation for apply_screen_artifacts
