### 2025-06-09 f343c93
Updated screen_artifacts_signals
