### 2025-06-09 7ee649e
Update preprocess_signals.py
Fixed documentation spelling error
