### 2025-06-09 20b18e5
Update preprocess_signals.py
apply_notch_filters now ignores NaN values
