### 2025-06-10 f884a54
Update preprocess_signals.py
Updated core preprocessing functions to work with NaN values, but raise warnings.

NaN values are ignored by removing them before performing the operation, then reinserting them afterwards. This ignores their presence from agregate functions.
