+++ src/EMGFlow/preprocess_signals.py
@@ -126,10 +126,10 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
 
     Raises
     ------
+    Warning
+        A warning is raised if 'col' contains NaN values.
     Exception
         An exception is raised if 'col' is not a column of 'Signal'.
-    Exception
-        An exception is raised if 'col' contains NaN values.
     Exception
         An exception is raised if 'sampling_rate' is less or equal to 0.
     Exception
@@ -171,6 +171,8 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
 
         Raises
         ------
+        Warning
+            A warning is raised if 'col' contains NaN values.
         Exception
             An exception is raised if the Hz value in notch is greater than
             sampling_rate/2 or less than 0.
@@ -209,7 +211,7 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
         b, a = scipy.signal.iirnotch(norm_Hz, Q)
         filtered_signal = scipy.signal.lfilter(b, a, filtered_signal)
         
-        # Reinsert NaN values
+        # Reinsert filtered values
         SignalCol = np.full_like(Signal[col], np.nan)
         SignalCol[~nan_ind] = filtered_signal
         
@@ -220,6 +222,7 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
     # Applies apply_notch_filter for every notch_val tuple
     for i in range(len(notch_vals)):
         notch_Signal[col] = apply_notch_filter(notch_Signal, col, sampling_rate, notch_vals[i])
+        
     return notch_Signal
 
 #
@@ -234,7 +237,7 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
     Parameters
     ----------
     in_path : str
-        Filepath to a directory to read Signal files.
+        Filepath to a directory to read signal files.
     out_path : str
         Filepath to an output directory.
     sampling_rate : float
@@ -245,7 +248,7 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
         Q-score (an intensity score where a higher number means a less
         extreme filter).
     cols : list-str, optional
-        List of columns of the Signal to apply the filter to. The default is
+        List of columns of the signal to apply the filter to. The default is
         None, in which case the filter is applied to every column except for
         'Time'.
     expression : str, optional
@@ -258,17 +261,17 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
     file_ext : str, optional
         File extension for files to read. Only reads files with this extension.
         The default is 'csv'.
-
+    
     Raises
     ------
     Warning
         Raises a warning if no files in 'in_path' match with 'expression'.
+    Warning
+        A warning is raised if any column in 'cols' in any of the signal files
+        read contain NaN values.
     Exception
         An exception is raised if any column in 'cols' is not found in any of
-        the Signal files read.
-    Exception
-        An exception is raised if any column in 'cols' in any of the Signal
-        files read contain NaN values.
+        the signal files read.
     Exception
         An exception is raised if 'sampling_rate' is less or equal to 0.
     Exception
@@ -283,7 +286,6 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
         An exception is raised if 'expression' is not None or a valid regular
         expression.
     
-
     Returns
     -------
     None.
@@ -367,10 +369,10 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high):
[…]
