### 2025-06-11 e296516
Updated artifact screening function names
Renamed artifact screening function names
- apply_screen_artifacts --> apply_fill_missing
- screen_artifacts_signals --> apply_fill_missing
