### 2025-06-11 74e4573
Renamed calc_spec_flux
Renamed calc_spec_flux to calc_sflx to be more consistent with other names

Renamed calc_sf to calc_sflt to be distinct from sflx
