+++ src/EMGFlow/extract_features.py
@@ -512,97 +512,6 @@ def calc_ap(Signal, col):
 # =============================================================================
 #
 
-def calc_spec_flux(Signal1, diff, col, sampling_rate, diff_sr=None):
-    """
-    Calculate the spectral flux of a Signal.
-
-    Parameters
-    ----------
-    Signal1 : pd.DataFrame
-        A Pandas dataframe containing a 'Time' column, and additional columns
-        for signal data.
-    diff : float, pd.DataFrame
-        The divisor of the calculation. If a percentage is provided, it will
-        calculate the spectral flux of the percentage of the Signal with one
-        minus the percentage of the Signal.
-    col : str
-        Column of 'Signal1' the feature is calculated from. If a second signal
-        is provided for 'diff', a column of the same name should be available
-        for use.
-    sampling_rate : float
-        Sampling rate of the Signal.
-    diff_sr : float, optional
-        Sampling rate for the second Signal if provided. The default is None,
-        in which case if a second Signal is provided, the sampling rate is
-        assumed to be the same as the first.
-
-    Raises
-    ------
-    Exception
-        An exception is raised if 'col' is not a column of 'Signal1'.
-    Exception
-        An exception is raised if 'sampling_rate' is less or equal to 0.
-    Exception
-        An exception is raised if 'diff' is a float, but isn't between 0 and 1.
-    Exception
-        An exception is raised if 'diff' is a dataframe and does not contain
-        'col'.
-    Exception
-        An exception is raised if 'diff_sr' is less or equal to 0.
-    Exception
-        An exception is raised if 'diff' is an invalid data type.
-
-    Returns
-    -------
-    flux : float
-        Spectral flux of the Signal.
-
-    """
-    
-    if col not in list(Signal1.columns.values):
-        raise Exception("Column " + str(col) + " not in Signal1")
-        
-    if sampling_rate <= 0:
-        raise Exception("Sampling rate cannot be 0 or negative")
-    
-    # Separate Signal1 by div and find spectral flux
-    if isinstance(diff, float):
-        if diff >= 1 or diff <= 0:
-            raise Exception("diff must be a float between 0 and 1")
-        
-        # Find column divider index
-        diff_ind = int(len(Signal1[col]) * diff)
-        # Take the PSD of each signal
-        psd1 = emg_to_psd(Signal1[col][:diff_ind], sampling_rate=sampling_rate)
-        psd2 = emg_to_psd(Signal1[col][diff_ind:], sampling_rate=sampling_rate)
-        # Calculate the spectral flux
-        flux = np.sum((psd1['Power'] - psd2['Power']) ** 2)
-        
-    # Find spectral flux of Signal1 by div
-    elif isinstance(diff, pd.DataFrame):
-        if col not in list(diff.columns.values):
-            raise Exception("Column " + str(col) + " not in diff")
-        
-        # If no second sampling rate, assume same sampling rate as first Signal
-        if diff_sr == None: diff_sr = sampling_rate
-        
-        if diff_sr <= 0:
-            raise Exception("Sampling rate cannot be 0 or negative")
-        # Take the PSD of each signal
-        psd1 = emg_to_psd(Signal1[col], sampling_rate=sampling_rate)
-        psd2 = emg_to_psd(diff[col], sampling_rate=diff_sr)
-        # Calculate the spectral flux
-        flux = np.sum((psd1['Power'] - psd2['Power']) ** 2)
-    
-    else:
-        raise Exception("Invalid data type given for diff: " + str(type(diff)))
-    
-    return flux
-
-#
-# =============================================================================
-#
-
[…]
