### 2025-06-11 b2e70e1
Update test_extract_features.py
Updated calc_sflt and calc_sflx in tests
