+++ tests/test_extract_features.py
@@ -126,14 +126,6 @@ class TestSimple(unittest.TestCase):
         AP = EMGFlow.calc_ap(Signal, 'EMG_zyg')
         self.assertIsInstance(AP, float)
     
-    def test_calc_spec_flux(self):
-        pathNames = EMGFlow.make_paths()
-        EMGFlow.make_sample_data(pathNames)
-        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
-        Signal = EMGFlow.read_file_type(filePath, 'csv')
-        specFlux = EMGFlow.calc_spec_flux(Signal, 0.5, 'EMG_zyg', 2000)
-        self.assertIsInstance(specFlux, float)
-    
     def test_calc_mdf(self):
         pathNames = EMGFlow.make_paths()
         EMGFlow.make_sample_data(pathNames)
@@ -199,14 +191,22 @@ class TestSimple(unittest.TestCase):
         SF = EMGFlow.calc_sf(PSD)
         self.assertIsInstance(SF, float)
     
-    def test_calc_ss(self):
+    def test_calc_sflt(self):
         pathNames = EMGFlow.make_paths()
         EMGFlow.make_sample_data(pathNames)
         filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
         Signal = EMGFlow.read_file_type(filePath, 'csv')
         PSD = EMGFlow.emg_to_psd(Signal['EMG_zyg'], 2000)
-        SS = EMGFlow.calc_ss(PSD)
+        SS = EMGFlow.calc_sflt(PSD)
         self.assertIsInstance(SS, float)
+
+    def test_calc_sflx(self):
+        pathNames = EMGFlow.make_paths()
+        EMGFlow.make_sample_data(pathNames)
+        filePath = os.path.join(pathNames['Raw'], '01', 'sample_data_01.csv')
+        Signal = EMGFlow.read_file_type(filePath, 'csv')
+        specFlux = EMGFlow.calc_sflx(Signal, 0.5, 'EMG_zyg', 2000)
+        self.assertIsInstance(specFlux, float)
     
     def test_calc_sd(self):
         pathNames = EMGFlow.make_paths()
