### 2025-06-11 6de81a1
Update test_extract_features.py
Fixed name bug
