### 2025-06-11 d4b9899
Updated with new sample data
Updated the example picture of plot_dashboard with the new sample data

Updated the example of a signal file in getting-started with the sample data
