+++ docs/guide/getting-started.md
@@ -13,22 +13,22 @@ pip install EMGFlow
 _EMGFlow_ extracts a comprehensive set of 32 statistical features from sEMG signals, achieved with only a few lines of code:
 
 ```python
-import EMGFlow
+import EMGFlow as ef
 
 # Get path dictionary
-path_names = EMGFlow.make_paths()
+path_names = ef.make_paths()
 
 # Load sample data
-EMGFlow.make_sample_data(path_names)
+ef.make_sample_data(path_names)
 
 # Preprocess signals
-EMGFlow.clean_signals(path_names, sampling_rate=2000)
+ef.clean_signals(path_names, sampling_rate=2000)
 
 # Plot data on the "EMG_zyg" column
-EMGFlow.plot_dashboard(path_names, 'EMG_zyg', 'mV')
+ef.plot_dashboard(path_names, 'EMG_zyg', 'mV')
 
 # Extract features and save results in "Features.csv" in feature_path
-df = EMGFlow.extract_features(path_names, sampling_rate=2000)
+df = ef.extract_features(path_names, sampling_rate=2000)
 ```
 
 ## Input data format
@@ -39,14 +39,17 @@ _EMGFlow_ accepts data in plaintext .CSV file format. Files should have the foll
 - Col 1 - Labelled Time, and contains the timestamps of sampled data
 - Col 2:n - Assumed to be sEMG or related signal data.
 
-| Time     | Data-1 | ... | Data-n |
-| -------- | ------ | --- | ------ |
-| 0        | 0      | ... |  -.04  |
-| 0.1      | 0.32   | ... |   .9   |
-| ...      | ...    | ... |   ...  |
-| 20.0     | -1.2   | ... |   1.7  |
+As an example, here is the first few rows of some sample data included in the package:
 
-Note, the Time column can be omitted when the sample rate is known.
+| Time    | EMG_zyg   | EMG_cor   |
+| ------- | --------- | --------- |
+| 50.0005 | -0.145569 | -0.097046 |
+| 50.0010 | -0.129089 | -0.100708 |
+| 50.0015 | -0.118713 | -0.101929 |
+| 50.0020 | -0.118103 | -0.104675 |
+| 50.0025 | -0.104370 | -0.097656 |
+
+The "Time" column can generally be omitted when the sampling rate is known, though some functions require a "Time" column.
 
 ## EMGFlow Pipeline Processing Steps
 
