### 2025-06-11 a6db15f
Update getting-started.md
Added "..." to the end of the example file
