### 2025-06-12 3678bbb
Update plot_signals.py
Fixed a bug that caused the feature path to be removed from the original path_names dictionary input
