### 2025-06-14 #7696 @faroit
@srlivingstone @WiIIson do you have updates on the revision?
