### 2025-06-20 f5f0038
Added NaN handling
Modified all functions to handle NaN values
- Takes a max_gap_ms parameter
- If NaN segments are less than this time, they are ignored
- If NaN segments are greater than this time, the data is divided and a filter is applied to each segment
- If value sequences are less than this time, they are set to NaN
