+++ src/EMGFlow/preprocess_signals.py
@@ -1,4 +1,5 @@
 import scipy
+import itertools
 import pandas as pd
 import numpy as np
 import os
@@ -104,7 +105,7 @@ def emg_to_psd(sig_vals, sampling_rate=1000, normalize=True):
 # =============================================================================
 #
 
-def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
+def apply_notch_filters(Signal, col, sampling_rate, notch_vals, min_gap_ms=30.0):
     """
     Apply a list of notch filters for given frequencies and Q-factors to a
     column of the provided data.
@@ -123,6 +124,11 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
         applied. Hz is the frequency to apply the filter to, and Q is the
         Q-score (an intensity score where a higher number means a less extreme
         filter).
+    min_gap_ms : float
+        Minimum length of time (in ms) for missing data. Segments of missing
+        data less than this threshold are ignored in the filter calculation,
+        and segments greater than this are used to divide the data, with the
+        filter applied to each part.
 
     Raises
     ------
@@ -146,6 +152,10 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
     # An exception is raised if 'col' is not a column of 'Signal'.
     if col not in list(Signal.columns.values):
         raise Exception("Column " + str(col) + " not in Signal")
+        
+    # A warning is raised if 'col' contains NaN values.
+    if Signal[col].isnull().values.any():
+        warnings.warn("Warning: NaN values detected in dataframe. These values will be ignored.")
     
     # An exception is raised if 'sampling_rate' is less or equal to 0.
     if sampling_rate <= 0:
@@ -179,7 +189,7 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
 
         Returns
         -------
-        SignalCol : pd.Series
+        notch_signal_col : pd.Series
             A Pandas series of the provided column with the notch filter
             applied.
 
@@ -187,6 +197,13 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
         
         Signal = Signal.copy()
         (Hz, Q) = notch
+        
+        if Signal[col].isnull().values.any():
+            raise Exception("NaN values were detected in the dataframe.")
+        
+        # An exception is raised if 'col' is not a column of 'Signal'.
+        if col not in list(Signal.columns.values):
+            raise Exception("Column " + str(col) + " not in Signal")
         
         # An exception is raised if a Hz value in 'notch_vals' is greater than
         # sampling_rate/2 or less than 0.
@@ -197,31 +214,58 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals):
         nyq_freq = sampling_rate / 2
         norm_Hz = Hz / nyq_freq
         
-        # Check for NaN - warn and remove if found
-        if Signal[col].isnull().values.any():
-            warnings.warn("Warning: NaN values detected in dataframe. These values will be ignored.")
-        
-        # Find indices of NaN values
-        nan_ind = np.isnan(Signal[col])
-        
         # Remove NaN values
-        filtered_signal = Signal[col][~nan_ind]
+        notch_signal_col = Signal[col]
         
         # Use scipy notch filter using normalized frequency
         b, a = scipy.signal.iirnotch(norm_Hz, Q)
-        filtered_signal = scipy.signal.lfilter(b, a, filtered_signal)
+        notch_signal_col = scipy.signal.lfilter(b, a, notch_signal_col)
         
-        # Reinsert filtered values
[…]
