### 2025-06-20 00957f5
Update test_preprocess_signals.py
Updated preprocessing tests with new formats
