### 2025-06-20 f3e66ec
Updated test functions
Updated with new formats
