### 2025-06-24 dbeac66
Updated emg_to_psd
emg_to_psd now works with NaN values.

Updated tests with new calling format.
