### 2025-06-24 19db12e
Update handle_outliers.py
Fixed bug with emg_to_psd
