### 2025-06-24 d0dcb28
Update extract_features.py
Updated bug with emg_to_psd not working
