+++ src/EMGFlow/extract_features.py
@@ -854,8 +854,8 @@ def calc_sflx(Signal1, diff, col, sampling_rate, diff_sr=None):
         # Find column divider index
         diff_ind = int(len(Signal1[col]) * diff)
         # Take the PSD of each signal
-        psd1 = emg_to_psd(Signal1[:diff_ind], col, sampling_rate=sampling_rate)
-        psd2 = emg_to_psd(Signal1[diff_ind:], col, sampling_rate=sampling_rate)
+        psd1 = emg_to_psd(Signal1.iloc[:diff_ind].reset_index(), col, sampling_rate=sampling_rate)
+        psd2 = emg_to_psd(Signal1.iloc[diff_ind:].reset_index(), col, sampling_rate=sampling_rate)
         # Calculate the spectral flux
         flux = np.sum((psd1['Power'] - psd2['Power']) ** 2)
         
