### 2025-06-25 fb8bd94
Update preprocess_signals.py
Fixed bug causing temp_dat to be interpreted as a tuple in newer versions of python
