### 2025-06-26 7de3cdd
Update extract_features.py
Made all features compatible with NaN values
