+++ src/EMGFlow/extract_features.py
@@ -24,7 +24,7 @@ A collection of functions for extracting features.
 
 def calc_iemg(Signal, col, sampling_rate):
     """
-    Calculate the Integreated EMG (IEMG) of a Signal.
+    Calculate the Integreated EMG (IEMG) of a Signal. Ignores NaNs.
 
     Parameters
     ----------
@@ -56,7 +56,10 @@ def calc_iemg(Signal, col, sampling_rate):
     if sampling_rate <= 0:
         raise Exception("Sampling rate cannot be less or equal to 0.")
     
-    IEMG = np.nansum(np.abs(Signal[col]) * sampling_rate)
+    # Get valid values
+    col_vals = Signal.dropna()[col].values
+    
+    IEMG = np.nansum(np.abs(col_vals) * sampling_rate)
     return IEMG
 
 #
@@ -65,7 +68,7 @@ def calc_iemg(Signal, col, sampling_rate):
 
 def calc_mav(Signal, col):
     """
-    Calculate the Mean Absolute Value (MAV) of a Signal.
+    Calculate the Mean Absolute Value (MAV) of a Signal. Ignores NaNs.
 
     Parameters
     ----------
@@ -90,8 +93,11 @@ def calc_mav(Signal, col):
     if col not in list(Signal.columns.values):
         raise Exception("Column '" + str(col) + "' not found in the Signal")
     
-    N = len(Signal[col])
-    MAV = np.sum(np.abs(Signal[col])) / N
+    # Get valid values
+    col_vals = Signal.dropna()[col].values
+    
+    N = len(col_vals)
+    MAV = np.sum(np.abs(col_vals)) / N
     return MAV
 
 #
@@ -100,7 +106,8 @@ def calc_mav(Signal, col):
 
 def calc_mmav1(Signal, col):
     """
-    Calculate the Modified Mean Absolute Value 1 (MMAV1) of a Signal.
+    Calculate the Modified Mean Absolute Value 1 (MMAV1) of a Signal. Ignores
+    NaNs.
 
     Parameters
     ----------
@@ -125,8 +132,11 @@ def calc_mmav1(Signal, col):
     if col not in list(Signal.columns.values):
         raise Exception("Column '" + str(col) + "' not found in the Signal.")
     
-    N = len(Signal[col])
-    vals = list(np.abs(Signal[col]))
+    # Get valid values
+    col_vals = Signal.dropna()[col].values
+    
+    N = len(col_vals)
+    vals = list(np.abs(col_vals))
     total = 0
     for n in range(N):
         if (0.25*N <= n) and (n <= 0.75*N):
@@ -142,7 +152,8 @@ def calc_mmav1(Signal, col):
 
 def calc_mmav2(Signal, col):
     """
-    Calculate the Modified Mean Absolute Value 2 (MMAV2) of a Signal.
+    Calculate the Modified Mean Absolute Value 2 (MMAV2) of a Signal. Ignores
+    NaNs.
 
     Parameters
     ----------
@@ -167,8 +178,11 @@ def calc_mmav2(Signal, col):
     if col not in list(Signal.columns.values):
         raise Exception("Column '" + str(col) + "' not found in the Signal.")
     
-    N = len(Signal[col])
-    vals = list(np.abs(Signal[col]))
+    # Get valid values
+    col_vals = Signal.dropna()[col].values
+    
+    N = len(col_vals)
+    vals = list(np.abs(col_vals))
     total = 0
     for n in range(N):
         if (0.25*N <= n) and (n <= 0.75*N):
@@ -186,7 +200,7 @@ def calc_mmav2(Signal, col):
 
[…]
