+++ src/EMGFlow/handle_outliers.py
@@ -239,14 +239,9 @@ def detect_spectral_outliers(in_path, sampling_rate, threshold, cols=None, low=N
 # =============================================================================
 #
 
-def apply_fill_missing(Signal, col, method='robust', interpolation='pchip', maxgap=100):
+def screen_artefacts(Signal, col, method='robust'):
     """
-    Removes outlier artifacts from a column of the provided data. Works in two
-    stages:
-        1. Identifies outliers using 'method', sets them to NaN.
-        2. Fills in NaN values using 'interpolation'.
-    
-    Either stage can be skipped by setting the parameter to None.
+    Creates a NaN mask for a column of a signal dataframe.
 
     Parameters
     ----------
@@ -254,42 +249,20 @@ def apply_fill_missing(Signal, col, method='robust', interpolation='pchip', maxg
         A Pandas dataframe containing a 'Time' column, and additional columns
         for signal data.
     col : str
-        Column of 'Signal' the filter is applied to.
+        Column of 'Signal' the NaN mask is created from.
     method : str, optional
         The outlier detection method being used. Valid options are 'robust',
         'normal', or None. The default is 'robust'.
-    interpolation : str, optional
-        The interpolation method being used to fill outliers. Valid options are
-        'pchip', 'spline' or None. The default is 'pchip'.
-    maxgap : int, optional
-        The maximum number of NaN values that can be present for the
-        interpolation method to trigger. If more than this number of NaN values
-        are present (including those flagged as outliers), a warning will raise
-        and the interpolation step will be skipped. The default is 100.
 
     Raises
     ------
-    Warning
-        A warning is raised if more than 'maxgap' NaN values are detected in
-        Signal at the interpolation stage.
     Exception
         An exception is raised if 'col' is not a column of 'Signal'.
-    Exception
-        An exception is raised if 'Time' is not a column of 'Signal'.
-    Exception
-        An exception is raised if 'method' is not a valid outlier detection
-        method.
-    Exception
-        An exception is raised if 'interpolation' is not a valid interpolation
-        method.
-    Exception
-        An exception is raised if not enough valid values are present to
-        perform interpolation.
 
     Returns
     -------
-    Signal : pd.DataFrame
-        A copy of the 'Signal' dataframe with the artifact screening applied.
+    masked_Signal : pd.DataFrame
+        A copy of the 'Signal' dataframe with an added column for the NaN mask.
 
     """
     
@@ -297,123 +270,88 @@ def apply_fill_missing(Signal, col, method='robust', interpolation='pchip', maxg
     if col not in list(Signal.columns.values):
         raise Exception("Column " + str(col) + " not in Signal")
     
-    # An exception is raised if 'Signal' does not have a 'Time' column
-    if 'Time' not in list(Signal.columns.values):
-        raise Exception('Signal is missing a "Time" column.')
+    masked_Signal = Signal.copy()
     
-    Signal = Signal.copy()
-    Signal = Signal.set_index('Time')
-    
-    # 1. Identify outlier parameters
+    # Identify outlier parameters
     if method=='robust':
         # Set high and low with robust method
-        high = np.nanmedian(Signal[col]) + 5*(1.482*scipy.stats.median_abs_deviation(Signal[col], nan_policy='omit'))
-        low = np.nanmedian(Signal[col]) - 5*(1.482*scipy.stats.median_abs_deviation(Signal[col], nan_policy='omit'))
+        high = np.nanmedian(masked_Signal[col]) + 5*(1.482*scipy.stats.median_abs_deviation(masked_Signal[col], nan_policy='omit'))
[…]
