### 2025-06-30 05f8b87
Added "Filled" path folder
Added a "Filled" path

Updated plot_dashboard to work with "Filled" path

Updated emg_to_psd to take a NaN mask
