### 2025-06-30 a93ad81
Updated workflow with "Filled" folder
Added multi-file functions for screening artefacts and filling missing data

Updated plot_signals to be compatible with data in the "Filled" folder

Updated extract_features to calculate missing data
