### 2025-07-02 7c2bc62
Update extract_features.py
Fixed bug with handling the Filled path in the extract_features function
