+++ src/EMGFlow/__init__.py
@@ -2,5 +2,5 @@ __version__ = "1.0.17"
 from .access_files import *
 from .preprocess_signals import *
 from .extract_features import *
-from .handle_outliers import *
+from .handle_missing_data import *
 from .plot_signals import *
