+++ src/EMGFlow/plot_signals.py
@@ -23,7 +23,7 @@ A collection of functions for plotting subject data.
 # =============================================================================
 #
 
-def plot_dashboard(path_names, col, units, expression=None, file_ext='csv', auto_run=True):
+def plot_dashboard(path_names, col, units, expression=None, file_ext='csv', use_mask=True, auto_run=True):
     """
     Generate a shiny dashboard of different processing stages for a given
     column.
@@ -192,9 +192,10 @@ def plot_dashboard(path_names, col, units, expression=None, file_ext='csv', auto
                     if col not in list(sigDF.columns.values):
                         raise Exception("Column " + str(col) + " not in Signal " + str(filename))
                     
-                    mask_col = 'mask_' + str(col)
-                    if mask_col in list(sigDF.columns.values):
-                        sigDF.loc[~sigDF[mask_col], col] = np.nan
+                    if use_mask:
+                        mask_col = 'mask_' + str(col)
+                        if mask_col in list(sigDF.columns.values):
+                            sigDF.loc[~sigDF[mask_col], col] = np.nan
                     
                     ax.plot(sigDF['Time'], sigDF[col], alpha=0.5, linewidth=1)
                 # Set legend for multiple plots
@@ -208,9 +209,10 @@ def plot_dashboard(path_names, col, units, expression=None, file_ext='csv', auto
                 if col not in list(sigDF.columns.values):
                     raise Exception("Column " + str(col) + " not in Signal " + str(filename))
                 
-                mask_col = 'mask_' + str(col)
-                if mask_col in list(sigDF.columns.values):
-                    sigDF.loc[~sigDF[mask_col], col] = np.nan
+                if use_mask:
+                    mask_col = 'mask_' + str(col)
+                    if mask_col in list(sigDF.columns.values):
+                        sigDF.loc[~sigDF[mask_col], col] = np.nan
                 
                 # Get colour data
                 i = (names.index(column) + 1) % len(colours)
