+++ src/EMGFlow/handle_outliers.py
@@ -1,534 +1,0 @@
-import pandas as pd
-import numpy as np
-import scipy
-import scipy.interpolate
-import scipy.signal
-import scipy.stats
-import os
-import re
-import warnings
-from tqdm import tqdm
-
-from .preprocess_signals import emg_to_psd
-from .access_files import *
-
-#
-# =============================================================================
-#
-
-"""
-A collection of functions for finding outliers while testing.
-"""
-
-#
-# =============================================================================
-#
-
-def detect_spectral_outliers(in_path, sampling_rate, threshold, cols=None, low=None, high=None, metric=np.median, expression=None, window_size=200, file_ext='csv'):
-    """
-    Looks at all Signals contained in a filepath, returns a dictionary of file
-    names and locations that have outliers.
-    
-    Works by interpolating an inverse function from the peaks of the signal's
-    spectrum. The function then calculates the 'metric' aggregate of the
-    differences between the predicted spectrum intensity of the inverse
-    function, and the actual spectrum intensity of the peaks. Finally, if the
-    largest difference between the predicted and actual values is greater than
-    the metric average multiplied by the threshold value, the file is flagged
-    for having an outlier and is added to the dictionary.
-
-    Parameters
-    ----------
-    in_path : str
-        Filepath to a directory to read Signal files.
-    sampling_rate : float
-        Sampling rate of the Signal.
-    threshold : float
-        The number of times greater than the metric a value has to be to be
-        considered an outlier.
-    cols : list-str, optional
-        List of columns of the Signal to search for outliers in. The default is
-        None, in which case outliers are searched for in every column except
-        for 'time'.
-    low : float, optional
-        Lower frequency limit of where to search for outliers. Should be the
-        same as lower limit for bandpass filtering, or some value that
-        eliminates the irrelevant lower frequency ranges. The default is None,
-        in which case no lower threshold is used.
-    high : float, optional
-        Upper frequency limit of where to search for outliers. Should be the
-        same as upper limit for bandpass filtering, or some value that
-        eliminates the irrelevant upper frequency ranges. The default is None,
-        in which case no upper threshold is used.
-    metric : function, optional
-        Some summary function that defines the metric used for finding
-        outliers. The default is 'np.median', but others such as 'np.mean' can
-        be used instead.
-    expression : str, optional
-        A regular expression. If provided, will only search for outliers in
-        files whose names match the regular expression. The default is None.
-    window_size : int, optional
-        The window size to use when filtering for local maxima. The default is
-        200.
-    file_ext : str, optional
-        File extension for files to read. Only reads files with this extension.
-        The default is 'csv'.
-
-    Raises
-    ------
-    Warning
-        A warning is raised if 'window_size' is greater than half the size of
-        a file read by the function.
-    Exception
-        An exception is raised if 'window_size' is not an integer greater than
-        0.
-    Exception
-        An exception is raised if 'sampling_rate' is less or equal to 0.
-    Exception
-        An exception is raised if 'threshold' is less or equal to 0.
-    Exception
-        An exception is raised if 'low' is greater than 'high'.
-    Exception
-        An exception is raised if 'low' or 'high' are negative.
-    Exception
-        An exception is raised if 'metric' is not a valid summary function.
-    Exception
-        An exception is raised if a column in 'cols' is not in a data file.
-    Exception
-        An exception is raised if a file cannot not be read in 'in_path'.
-    Exception
[…]

[…]
