### 2025-07-02 fa33439
Updated paths
Updated file paths for make_paths
