### 2025-07-02 be815a9
Fixed tests
Updated tests for new formats
