### 2025-07-03 af19258
Updated data, Fixed Notch Filter
Fixed the notch filter function - wasnt working properly with NaN data

Updated sample data with NaN values
