### 2025-07-03 3a8a167
Updated examples
Updated examples page with new examples
- Added steps for artefact screening and filling missing data
- Added an example for using your own data paths
