### 2025-07-03 158163d
Updated documentation, simplified code
Updated some documentation

Removed "convert_map_files"
