+++ src/EMGFlow/handle_missing_data.py
@@ -1,5 +1,4 @@
 import pandas as pd
-import numpy as np
 import scipy
 import scipy.interpolate
 import scipy.signal
