+++ src/EMGFlow/preprocess_signals.py
@@ -1,5 +1,4 @@
 import scipy
-import itertools
 import pandas as pd
 import numpy as np
 import os
