### 2025-07-08 0a3b421
Updated code and documentation
Updated documentation for some functions

Updated PSD function with additional line- confirmed that all functions from preprocess_signals work identically to older versions
