+++ src/EMGFlow/extract_features.py
@@ -24,7 +24,7 @@ A collection of functions for extracting features.
 
 def calc_iemg(Signal, col, sampling_rate):
     """
-    Calculate the Integreated EMG (IEMG) of a Signal. Ignores NaNs.
+    Calculate the Integreated EMG (IEMG) of a column of 'Signal'. Ignores NaNs.
 
     Parameters
     ----------
@@ -46,7 +46,7 @@ def calc_iemg(Signal, col, sampling_rate):
     Returns
     -------
     IEMG : float
-        IEMG of the Signal.
+        The calculated IEMG.
 
     """
     
@@ -68,7 +68,8 @@ def calc_iemg(Signal, col, sampling_rate):
 
 def calc_mav(Signal, col):
     """
-    Calculate the Mean Absolute Value (MAV) of a Signal. Ignores NaNs.
+    Calculate the Mean Absolute Value (MAV) of a column of 'Signal'. Ignores
+    NaNs.
 
     Parameters
     ----------
@@ -86,7 +87,7 @@ def calc_mav(Signal, col):
     Returns
     -------
     MAV : float
-        MAV of the Signal.
+        The calculated MAV.
 
     """
     
@@ -106,8 +107,8 @@ def calc_mav(Signal, col):
 
 def calc_mmav1(Signal, col):
     """
-    Calculate the Modified Mean Absolute Value 1 (MMAV1) of a Signal. Ignores
-    NaNs.
+    Calculate the Modified Mean Absolute Value 1 (MMAV1) of a column of
+    'Signal'. Ignores NaNs.
 
     Parameters
     ----------
@@ -125,7 +126,7 @@ def calc_mmav1(Signal, col):
     Returns
     -------
     MMAV1 : float
-        MMAV1 of the Signal.
+        The calculated MMAV1.
 
     """
     
@@ -152,8 +153,8 @@ def calc_mmav1(Signal, col):
 
 def calc_mmav2(Signal, col):
     """
-    Calculate the Modified Mean Absolute Value 2 (MMAV2) of a Signal. Ignores
-    NaNs.
+    Calculate the Modified Mean Absolute Value 2 (MMAV2) of a column of
+    'Signal'. Ignores NaNs.
 
     Parameters
     ----------
@@ -171,7 +172,7 @@ def calc_mmav2(Signal, col):
     Returns
     -------
     MMAV2 : float
-        MMAV2 of the Signal.
+        The calculated MMAV2'.
 
     """
     
@@ -200,7 +201,8 @@ def calc_mmav2(Signal, col):
 
 def calc_ssi(Signal, col, sampling_rate):
     """
-    Calculate the Simple Square Integreal (SSI) of a Signal. Ignores NaNs.
+    Calculate the Simple Square Integral (SSI) of a column of 'Signal'.
+    Ignores NaNs.
 
     Parameters
     ----------
@@ -222,7 +224,7 @@ def calc_ssi(Signal, col, sampling_rate):
     Returns
     -------
     SSI : float
-        SSI of the Signal.
+        The calculated SSI.
 
     """
     
@@ -244,7 +246,7 @@ def calc_ssi(Signal, col, sampling_rate):
 
 def calc_var(Signal, col):
     """
-    Calculate the Variance (VAR) of a Signal. Ignores NaNs.
+    Calculate the Variance (VAR) of a column of 'Signal'. Ignores NaNs.
 
     Parameters
     ----------
@@ -262,7 +264,7 @@ def calc_var(Signal, col):
     Returns
     -------
     VAR : float
-        VAR of the Signal.
+        The calculated VAR.
 
     """
     
@@ -282,7 +284,7 @@ def calc_var(Signal, col):
 
[…]
