### 2025-07-08 09411e3
Updated processing steps page
Changed "getting-started" page to "processing-steps"

Added table for sources for the sample data
