### 2025-07-08 6ab7fb2
Update what-is-emgflow.md
Cleaned up dead links
