### 2025-07-08 cda583d
Update config.ts
fixed mistake
