### 2025-07-09 5986115
Updated sample data, plotting functions and bugfixes
Updated sample data time data.

Updated plotting function to have step labels, and re-ordered layering.

Fixed a bug in the main series functions where NaN values were interfering with calculations.
