### 2025-07-14 5671fea
Updated docs
Updated docs and confirmed correctness in all modules except preprocess_signals
