### 2025-07-14 60d2b22
Update plot_signals.py
Updated plotting color scheme
