### 2025-07-14 b1421ac
Updated docs
Updated docs in plot_signals

Updated docs in preprocess_signals - everything up to and including notch_filter_signals is correct
