+++ src/EMGFlow/preprocess_signals.py
@@ -22,9 +22,9 @@ A collection of functions for filtering Signals.
 
 def emg_to_psd(Signal, col, sampling_rate=1000, normalize=True, min_gap_ms=30.0, nan_mask=None):
     """
-    Creates a PSD graph of a Signal. Uses the Welch method, meaning it can be
-    used as a Long Term Average Spectrum (LTAS).
-
+    Creates a PSD dataframe of a signal. Uses the Welch method, meaning it can
+    be used as a Long Term Average Spectrum (LTAS).
+    
     Parameters
     ----------
     Signal : pd.DataFrame
@@ -33,24 +33,31 @@ def emg_to_psd(Signal, col, sampling_rate=1000, normalize=True, min_gap_ms=30.0,
     col : str
         Column of 'Signal' the filter is applied to.
     sampling_rate : float
-        Sampling rate of 'sig_vals'. This is the number of entries recorded per
-        second, or the inverse of the difference between entries.
+        Sampling rate of 'Signal'.
     normalize : bool, optional
         If True, will normalize the result. If False, will not. The default is
         True.
     min_gap_ms : float, optional
-        
-    mask : pd.Series, optional
+        The minimum length (in ms) for data to be considered valid. If a length
+        of data is less than this time, it is set to NaN. If a length of
+        invalid data is less than this time, it is ignored in calculations. The
+        default is 30.0.
+    nan_mask : pd.Series, optional
+        Optional series that controls the calculation of the function. Can be
+        a True/False mask that is the same size as the selected column, and
+        will set all associated False values in the column to NaN in the
+        calculation. The default is None, in which case no NaN masking will be
+        done.
 
     Raises
     ------
     Exception
-        An exception is raised if 'sig_vals' is a pd.DataFrame, not a column of
-        a dataframe.
+        An exception is raised if 'col' is not a column of 'Signal'.
     Exception
-        An exception is raised if 'sig_vals' contains NaN values.
+        An exception is raised if 'sampling_rate' is less than or equal to 0.
     Exception
-        An exception is raised if 'sampling_rate' is less or equal to 0.
+        An exception is raised if 'nan_mask' is an incorrect data type, or not
+        the same length as the column
 
     Returns
     -------
@@ -63,10 +70,10 @@ def emg_to_psd(Signal, col, sampling_rate=1000, normalize=True, min_gap_ms=30.0,
     
     # An exception is raised if 'col' is not a column of 'Signal'.
     if col not in list(Signal.columns.values):
-        raise Exception("Column " + str(col) + " not in Signal")
+        raise Exception("Column " + str(col) + " not in 'Signal'")
     
     if sampling_rate <= 0:
-        raise Exception("Sampling rate must be greater or equal to 0")
+        raise Exception("'sampling_rate' must be greater than 0")
     
     PSD_Signal = Signal.copy()
     
@@ -180,12 +187,12 @@ def apply_notch_filters(Signal, col, sampling_rate, notch_vals, min_gap_ms=30.0)
         applied. Hz is the frequency to apply the filter to, and Q is the
         Q-score (an intensity score where a higher number means a less extreme
         filter).
-    min_gap_ms : float
-        Minimum length of time (in ms) for missing data. Segments of missing
-        data less than this threshold are ignored in the filter calculation,
-        and segments greater than this are used to divide the data, with the
-        filter applied to each part.
-
+    min_gap_ms : float, optional
+        The minimum length (in ms) for data to be considered valid. If a length
+        of data is less than this time, it is set to NaN. If a length of
+        invalid data is less than this time, it is ignored in calculations. The
+        default is 30.0.
+        
     Raises
     ------
     Warning
[…]
