### 2025-07-15 8c5de70
Updated plot_signals and docs
Updated plot_signals to use Viridis colour scheme

Updated docs with changes
