+++ src/EMGFlow/preprocess_signals.py
@@ -322,6 +322,9 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
     ------
     Warning
         A warning is raised if no files in 'in_path' match with 'expression'.
+    Exception
+        An exception is raised if 'expression' is not None or a valid regular
+        expression.
         
     Warning
         A warning is raised if a column in 'cols' contains NaN values.
@@ -335,10 +338,6 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
         'sampling_rate'/2 or less than 0.
         
     Exception
-        An exception is raised if 'expression' is not None or a valid regular
-        expression.
-        
-    Exception
         An exception is raised if a file could not be read.
     Exception
         An exception is raised if an unsupported file format was provided for
@@ -350,13 +349,19 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
 
     """
     
+    if expression is not None:
+        try:
+            re.compile(expression)
+        except:
+            raise Exception("Invalid regex expression provided")
+    
     # Convert out_path to absolute
     if not os.path.isabs(out_path):
         out_path = os.path.abspath(out_path)
     
     # Get dictionary of file locations
     if exp_copy:
-        file_dirs = map_files(in_path, file_ext=file_ext, expression=expression)
+        file_dirs = map_files(in_path, file_ext=file_ext)
     else:
         file_dirs = map_files(in_path, file_ext=file_ext, expression=expression)
         if len(file_dirs) == 0:
@@ -403,7 +408,8 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
 
 def apply_bandpass_filter(Signal, col, sampling_rate, low, high, min_gap_ms=30.0):
     """
-    Apply a bandpass filter to a Signal for a given lower and upper limit.
+    Apply a bandpass filter to for a given lower and upper limit to a column of
+    the provided data.
 
     Parameters
     ----------
@@ -419,10 +425,10 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high, min_gap_ms=30.0
     high : float
         Upper frequency limit of the bandpass filter.
     min_gap_ms : float, optional
-        Minimum length of time (in ms) for missing data. Segments of missing
-        data less than this threshold are ignored in the filter calculation,
-        and segments greater than this are used to divide the data, with the
-        filter applied to each part. The default is 30.
+        The minimum length (in ms) for data to be considered valid. If a length
+        of data is less than this time, it is set to NaN. If a length of
+        invalid data is less than this time, it is ignored in calculations. The
+        default is 30.0.
 
     Raises
     ------
@@ -455,7 +461,7 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high, min_gap_ms=30.0
     
     # An exception is raised if 'sampling_rate' is less or equal to 0.
     if sampling_rate <= 0:
-        raise Exception("Sampling rate must be greater or equal to 0.")
+        raise Exception("Sampling rate must be greater than 0.")
     
     # An exception is raised if 'high' or 'low' are higher than 1/2 of
     # 'sampling_rate'.
@@ -530,9 +536,9 @@ def bandpass_filter_signals(in_path, out_path, sampling_rate, low=20, high=450, 
         Filepath to an output directory.
     sampling_rate : float
         Sampling rate of the signal files.
-    low : float
+    low : float, optional
         Lower frequency limit of the bandpass filter. The default is 20.
-    high : float
+    high : float, optional
         Upper frequency limit of the bandpass filter. The default is 450.
     cols : list, optional
         List of columns of the Signal to apply the filter to. The default is
[…]
