### 2025-07-15 203305c
Revised docs
Revised docs to fix errors and apply consistency
