+++ src/EMGFlow/preprocess_signals.py
@@ -31,7 +31,7 @@ def emg_to_psd(Signal, col, sampling_rate=1000, normalize=True, min_gap_ms=30.0,
         A Pandas dataframe containing a 'Time' column, and additional columns
         for signal data.
     col : str
-        Column of 'Signal' the filter is applied to.
+        Column of 'Signal' the PSD is created from.
     sampling_rate : float
         Sampling rate of 'Signal'.
     normalize : bool, optional
@@ -304,7 +304,7 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
         Q-score (an intensity score where a higher number means a less
         extreme filter).
     cols : list-str, optional
-        List of columns of the signal to apply the filter to. The default is
+        List of columns of the signals to apply the filter to. The default is
         None, in which case the filter is applied to every column except for
         'Time'.
     expression : str, optional
@@ -524,14 +524,14 @@ def apply_bandpass_filter(Signal, col, sampling_rate, low, high, min_gap_ms=30.0
 
 def bandpass_filter_signals(in_path, out_path, sampling_rate, low=20, high=450, cols=None, expression=None, exp_copy=False, file_ext='csv'):
     """
-    Apply bandpass filters to all Signals in a folder. Writes filtered Signals
-    to an output folder, and generates a file structure
-    matching the input folder.
+    Apply bandpass filters to all signal files in a folder. Writes filtered
+    signal files to an output folder, and generates a file structure matching
+    the input folder.
     
     Parameters
     ----------
     in_path : dict-str
-        Filepath to a directory to read Signal files.
+        Filepath to a directory to read signal files.
     out_path : str
         Filepath to an output directory.
     sampling_rate : float
@@ -540,8 +540,8 @@ def bandpass_filter_signals(in_path, out_path, sampling_rate, low=20, high=450, 
         Lower frequency limit of the bandpass filter. The default is 20.
     high : float, optional
         Upper frequency limit of the bandpass filter. The default is 450.
-    cols : list, optional
-        List of columns of the Signal to apply the filter to. The default is
+    cols : list-str, optional
+        List of columns of the signals to apply the filter to. The default is
         None, in which case the filter is applied to every column except for
         'Time'.
     expression : str, optional
@@ -1128,15 +1128,15 @@ def apply_loess_smooth(Signal, col, sampling_rate, window_size, min_gap_ms=30.0)
 
 def smooth_filter_signals(in_path, out_path, sampling_rate, window_size, cols=None, expression=None, exp_copy=False, file_ext='csv', method='rms', min_gap_ms=30.0, sigma=1):  
     """
-    Apply smoothing filters to all Signals in a folder. Writes filtered signal
-    files to an output folder, and generates a file structure matching the
-    input folder. The method used to smooth the signals can be specified, but
-    is RMS as default.
+    Apply smoothing filters to all signal files in a folder. Writes filtered
+    signal files to an output folder, and generates a file structure matching
+    the input folder. The method used to smooth the signals can be specified,
+    but is RMS as default.
 
     Parameters
     ----------
     in_path : dict-str
-        Filepath to a directory to read Signal files.
+        Filepath to a directory to read signal files.
     out_path : str
         Filepath to an output directory.
     sampling_rate : float
@@ -1144,9 +1144,9 @@ def smooth_filter_signals(in_path, out_path, sampling_rate, window_size, cols=No
     window_size : int, float
         Size of the window of the filter.
     cols : list-str, optional
-        List of columns of the Signal to apply the filter to. The default is
+        List of columns of the signals to apply the filter to. The default is
         None, in which case the filter is applied to every column except for
-        'time'.
+        'Time'.
     expression : str, optional
         A regular expression. If provided, will only filter files whose names
         match the regular expression. The default is None.
[…]
