### 2025-07-15 ceec940
Update preprocess_signals.py
Updated core functions to be compatible if the data has NaN masks
