### 2025-07-15 c8b0912
Fixed RegEx parameters
Fixed error with regex parameters not working
