### 2025-07-16 8b03d60
Fixed exp_copy
Fixed file writing bug related to exp_copy
