+++ src/EMGFlow/preprocess_signals.py
@@ -394,8 +394,10 @@ def notch_filter_signals(in_path, out_path, sampling_rate, notch, cols=None, exp
         elif (file[-len(file_ext):] == file_ext) and exp_copy:
             # Copy the file even if it doesn't match if exp_copy is true
             data = read_file_type(file_dirs[file], file_ext)
+            
             out_file = out_path + file_dirs[file][len(in_path):]
-            out_folder = out_file[:len(out_file) - len(file)]
+            out_folder = out_file[:len(out_file) - len(os.path.basename(out_file)) - 1]
+            
             os.makedirs(out_folder, exist_ok=True)
             data.to_csv(out_file, index=False)
     
@@ -632,8 +634,10 @@ def bandpass_filter_signals(in_path, out_path, sampling_rate, low=20, high=450, 
         elif (file[-len(file_ext):] == file_ext) and exp_copy:
             # Copy the file even if it doesn't match if exp_copy is true
             data = read_file_type(file_dirs[file], file_ext)
+            
             out_file = out_path + file_dirs[file][len(in_path):]
-            out_folder = out_file[:len(out_file) - len(file)]
+            out_folder = out_file[:len(out_file) - len(os.path.basename(out_file)) - 1]
+            
             os.makedirs(out_folder, exist_ok=True)
             data.to_csv(out_file, index=False)
             
@@ -1255,8 +1259,10 @@ def smooth_filter_signals(in_path, out_path, sampling_rate, window_size, cols=No
         elif (file[-len(file_ext):] == file_ext) and exp_copy:
             # Copy the file even if it doesn't match if exp_copy is true
             data = read_file_type(file_dirs[file], file_ext)
+            
             out_file = out_path + file_dirs[file][len(in_path):]
-            out_folder = out_file[:len(out_file) - len(file)]
+            out_folder = out_file[:len(out_file) - len(os.path.basename(out_file)) - 1]
+            
             os.makedirs(out_folder, exist_ok=True)
             data.to_csv(out_file, index=False)
     return
@@ -1701,8 +1707,10 @@ def screen_artefact_signals(in_path, out_path, sampling_rate, method='robust', c
         elif (file[-len(file_ext):] == file_ext) and exp_copy:
             # Copy the file even if it doesn't match if exp_copy is true
             data = read_file_type(file_dirs[file], file_ext)
+            
             out_file = out_path + file_dirs[file][len(in_path):]
-            out_folder = out_file[:len(out_file) - len(file)]
+            out_folder = out_file[:len(out_file) - len(os.path.basename(out_file)) - 1]
+            
             os.makedirs(out_folder, exist_ok=True)
             data.to_csv(out_file, index=False)
     
