### 2025-07-16 be0a34b
Updated file format docs
Removed extra file format info in the overview page

Updated file format info in the processing-steps page
- Added explanation for NaNs
- Updated Time data in example
- Rewrote some things
