+++ docs/guide/processing-steps.md
@@ -39,18 +39,41 @@ _EMGFlow_ accepts data in plaintext .CSV file format. Files should have the foll
 - Col 1 - Labelled Time, and contains the timestamps of sampled data
 - Col 2:n - Assumed to be sEMG or related signal data.
 
-As an example, here is the first few rows of some sample data included in the package:
+As an example, here is the first few rows of some sample data included in the package with added NaN values:
 
-| Time    | EMG_zyg   | EMG_cor   |
-| ------- | --------- | --------- |
-| 50.0005 | -0.145569 | -0.097046 |
-| 50.0010 | -0.129089 | -0.100708 |
-| 50.0015 | -0.118713 | -0.101929 |
-| 50.0020 | -0.118103 | -0.104675 |
-| 50.0025 | -0.104370 | -0.097656 |
-| ...     | ...       | ...       |
+| Time   | EMG_zyg   | EMG_cor   |
+| ------ | --------- | --------- |
+| 0.0005 | -0.145569 | -0.097046 |
+| 0.0010 | -0.129089 | -0.100708 |
+| 0.0015 | -0.118713 | -0.101929 |
+| 0.0020 | NaN       | NaN       |
+| 0.0025 | -0.104370 | -0.097656 |
+| ...    | ...       | ...       |
 
-The "Time" column can generally be omitted when the sampling rate is known, though some functions require a "Time" column.
+For proper calculations, ensure the 'Time' column has an equal difference between each sequential row.
+
+EMGFlow uses `pd.read_csv()` from the Pandas package to read in files. Continuing with the documentation for Pandas, the following values are interpreted as `NaN`:
+- " "
+- "#N/A"
+- "#N/A N/A"
+- "#NA"
+- "-1.#IND"
+- "-1.#QNAN"
+- "-NaN"
+- "-nan"
+- "1.#IND"
+- "1.#QNAN"
+- "\<NA\>"
+- "N/A"
+- "NA"
+- "NULL"
+- "NaN"
+- "None"
+- "n/a"
+- "nan"
+- "null"
+
+In the future, we plan to make EMGFlow compatible with other file formats.
 
 ## EMGFlow Pipeline Processing Steps
 
