+++ docs/reference/access-files.md
@@ -12,7 +12,6 @@ mindmap
             make_sample_data
             read_file_type
             map_files
-            convert_map_files
             map_files_fuse
         HO(Handle Outliers)
         PrS(Preprocess Signals)
@@ -30,7 +29,9 @@ mindmap
 
 Generates a file structure for signal files, and returns a dictionary of the locations for these files.
 
-Creates 'Raw', 'Notch', 'Bandpass', 'Smooth' and 'Feature' subfolders at a given location. If no path is given, will create a 'Data' folder in the current working directory, with these subfolders inside.
+Creates '1_raw', '2_notch', '3_bandpass', '4_smooth', '5_filled' and '6_feature' subfolders at a given location. If no path is given, will create a 'Data' folder in the current working directory, with these subfolders inside.
+
+The paths to these directories can be accessed with the corresponding keys: 'Raw', 'Notch', 'Bandpass', 'Smooth', 'Filled' and 'Feature'.
 
 ```python
 make_paths(root=None)
@@ -61,9 +62,11 @@ path_names = EMGFlow.make_paths()
 
 **Description**
 
-Generates sample data in the 'Raw' folder of a provided dictionary of file locations.
+Generates sample data in the '1_raw' folder of a provided dictionary of file locations.
 
 Creates '01' and '02' folders, which each contain two sample data files ('01/sample_data_01.csv', '01/sample_data_02.csv', '02/sample_data_03.csv', '02/sample_data_04.csv')
+
+The sample data will not be written if it already exists in the folder.
 
 ```python
 make_sample_data(path_names)
@@ -133,7 +136,7 @@ An exception is raised if an unsupported file format was provided for `file_ext`
 
 ```python
 # Read a csv file
-path = '/Data/Raw/01/sample_data_01.csv'
+path = '/Data/1_raw/01/sample_data_01.csv'
 ext = 'csv'
 df = EMGFlow.read_file_type(path, ext)
 ```
@@ -173,7 +176,7 @@ An exception is raised if `expression` is not None or a valid regular expression
 **Returns**
 
 `file_dirs`: dict-str
-- Returns a dictionary of file name and location keys/values.
+- Returns a dictionary of file name keys and file path location values.
 
 **Example**
 
@@ -190,59 +193,11 @@ file_loc_2 = EMGFlow.map_files('data', expression='^DATA_')
 
 
 
-## `convert_map_files`
-
-**Description**
-
-A more advanced version of `map_files` that can coerce other data types into the `map_files` format.
-
-If provided a dictionary (assumed to be a file location map), it is returned, filtered by `expression` if provided.
-
-```python
-convert_map_files(file_obj, file_ext='csv', experssion=None)
-```
-
-**Parameters**
-
-`file_obj`: str, dict-str
-- The file location object. This can be a string to a file location, or an already formed dictionary of file locations.
-
-`file_ext`: str, optional ('csv')
-- File extension for files to read. Only reads files with this extension. The default is 'csv'.
-
-`expression`: str, optional (None)
-- A regular expression. If provided, will only count files whose names match the regular expression. The default is None.
-
-**Raises**
-
-An exception is raised if `file_ext` is an unsupported file format.
-
-An exception is raised if `expression` is not None or a valid regular expression.
-
-**Returns**
-
-`file_dirs`: dict-str
-- Returns a dictionary of file name and location keys/values.
-
-**Example**
-
-```python
-# Read in file locations normally
-file_loc_1 = EMGFlow.convert_map_files('/data')
-
-# Filter an existing dataframe with a regular expression
-file_loc_2 = EMGFlow.convert_map_files(file_loc_1, expression='^01')
-```
-
-
-
-
-
 ## `map_files_fuse`
 
 **Description**
 
[…]
