### 2025-07-17 b07c284
Updated docs, fixed bugs
Fixed bug in extract_features where mask columns weren't ignored.

Updated documentation for extract_features, and functions with the expression parameter
