### 2025-10-07 #7696 @faroit
@rkurchin thanks for checking in, I was sick for the last 2 weeks and forgot to put my absence notification here
