### 2025-11-03 #7696 @srlivingstone
Hi @faroit , @samiralavi and @wbaccinelli , thank you for your patience. @WiIIson and I have completed our revisions, and responded to and closed all previous Issues. In our new 1.0 release, we made extensive revisions to the code, documentation, and supporting paper. These changes touched on five main areas:

1. Usability and accessibility. The API now reads cleanly and predictably: full snake_case throughout the package, consistent parameter ordering, and fully spelled argument names improved readability. A `clean_signals()` wrapper, `make_paths()` helper, and bundled sample data let new users run an end-to-end pipeline without prior signal-processing expertise.
2. Robustness and Data Quality. Preprocessing is more resilient to real data: core functions are now NaN-tolerant; a new dedicated `fill_missing_signals()` stage and a “percent missing” metric make incompleteness explicit and measurable; a new `screen_artefact_signals()` stage handles common sources of contamination; and warnings about NaNs are replaced with deterministic handling.
3. Technical capabilities. The toolbox adds capability beyond basic filtering. Full-wave rectification, improved PSD estimation, and new artefact screening and interpolation yield more stable amplitude and spectral summaries.
4. Visualization and interactive diagnostics. The dashboard is more effective and self-managing: It is multi-threaded, closes cleanly when the tab closes, and adds zoom plus spectral views.
5. Reproducibility and communication. Unit tests run identically in CI and locally. Documentation was moved to a clearer VitePress site and the README was updated to match the code. GitHub Discussions was enabled to support community discussions, reflected in our dedicated Contributions section. The JOSS manuscript was extensively revised to reflect the above improvements (API consistency, pipeline order, NaN handling, artefact screening and interpolation, dashboard enhancements, and documentation upgrades), keeping paper, code, and docs aligned.

We look forward to hearing your comments.
