### 2026-01-07 #7696 @srlivingstone
Hi @arfon and @faroit

Thanks for sharing. To clarify the use of AI to the editor and reviewers - EMGFlow code was written manually by @WiIIson and myself. We used GenAI to assist in clarifying some aspects of our new/revised code features in our 2025-11 resubmission, but implemented these manually. We used GenAI to copy-edit a final draft of the JOSS paper, and our manuscript includes a declaration stating its use.

@faroit - I understand that our submission is grandfathered in and will be reviewed under the pre-2026 checklist. However, I want to be clear that our submission conforms to the new 2026 JOSS guidelines. We will soon post an updated version of our manuscript that includes the revised "AI usage policy" [statement](https://joss.readthedocs.io/en/latest/submitting.html#ai-usage-policy).

Best
Steven.
