### 2026-01-23 #7696 @faroit
@srlivingstone thanks for the update and proactively adding the AI usage policy. This looks all good to me. We can now move over to the final part of the review.
