### 2026-01-23 #7696 @faroit
@srlivingstone pdf looks good. I would say it would be nice if the code examples aren't interrupted by page breaks but its not a big deal
