### 2026-01-23 #7696 @faroit
@srlivingstone please follow the post-review guidelines above https://github.com/openjournals/joss-reviews/issues/7696#issuecomment-3792600002

Also, could you please fix a couple more DOIs (see above)? 🙏
