### 2026-02-23 #7696 @srlivingstone
Hi @faroit - thank you for your patience. We have completed the author checklist, as below:

**Author checklist**
Affiliations: Confirmed
Release: 1.1.2
Archive: https://doi.org/10.5281/zenodo.18748950
License: Confirmed
