### 2026-03-24 #7696 @faroit
@srlivingstone sorry for the delay, i did a proof read in https://github.com/WiIIson/EMGFlow-Python-Package/pull/10, please have a look.
