### 2026-03-25 #7696 @srlivingstone
> [@srlivingstone](https://github.com/srlivingstone) sorry for the delay, i did a proof read in [WiIIson/EMGFlow-Python-Package#10](https://github.com/WiIIson/EMGFlow-Python-Package/pull/10), please have a look.

Hi @faroit thanks for those edits. We've approved and merged. Please let me know if you need anything further from us.
