### 2026-03-30 #7696 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
