### 2026-03-30 #7696 @Kevin-Mattheus-Moerman
@WiIIson As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks:

👉 No, there are no releases on the repository. Please create a release and create/use a version tag that we'll assign here and which will be listed over in the archive.

👉 No, see other comment above. The version tag should match a tagged release on the repository. Please let me know what the tag will be and I'll use it here as well. I see you have `v1` over at the ZENODO archive. If you want to use that then use `v1` also for the version tag on your GitHub repository (in this case, include the `v` as well). If you decide on a different version tag then you need to either edit the current ZENODO archive listing to have the correct tag (you can easity edit this manualy), or you should create a archived version and inform me about the new DOI.

- [x] Copyedit pass/checked for typos
