### 2026-03-30 #7696 @srlivingstone
Hi @Kevin-Mattheus-Moerman thanks for that. @WiIIson has released v1.1.2 on Github. I've updated the Zenodo project archive version tag to v1.1.2 - thanks for pointing out the discrepancy (upload itself is unchanged). Below are the updated links. Please let me know if you need anything further.
https://zenodo.org/records/18748950
https://doi.org/10.5281/zenodo.18748950
