### 2026-03-30 #7696 @Kevin-Mattheus-Moerman
@srlivingstone thanks for making those changes. One minor thing is that the archive version tag should include the `v` in `v1.1.2` as well. I'll proceed now, but if you could make that final edit that would be good.
